Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
getCPLConfigOption(ConfigOption)
setCPLConfigOption(ConfigOption, value)
get_cached_orig_PROJ_LIB()
get_cached_orig_GDAL_DATA()
get_cached_set_PROJ_LIB()
get_cached_set_GDAL_DATA()

Arguments

driver
An object inheriting from class 'GDALDriver'

str
A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

ConfigOption
CPL configure option documented in https://trac.osgeo.org/gdal/wiki/ConfigOptions and elsewhere in GDAL source code

value
a string value to set a CPL option; NULL is used to unset the CPL option

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

get_cached_orig_PROJ_LIB, get_cached_orig_GDAL_DATA

The values of environment variables PROJ_LIB and GDAL_DATA as read when this package was loaded

get_cached_set_PROJ_LIB, get_cached_set_GDAL_DATA

If not "", the values set when loading this package to point to metadata files included in CRAN binary packages

Note

Loading the rgdal package changes the GDAL\_DATA environmental variable to the GDAL support files bundled with the package.

Author

Timothy H. Keitt, modified by Roger Bivand

See also

Examples

gdalDrivers()

#>                name                                                long_name
#> 1           AAIGrid                                      Arc/Info ASCII Grid
#> 2              ACE2                                                     ACE2
#> 3              ADRG                            ARC Digitized Raster Graphics
#> 4               AIG                                     Arc/Info Binary Grid
#> 5            AirSAR                                AirSAR Polarimetric Image
#> 6               ARG                                Azavea Raster Grid format
#> 7               BAG                               Bathymetry Attributed Grid
#> 8            BIGGIF                       Graphics Interchange Format (.gif)
#> 9               BLX                                     Magellan topo (.blx)
#> 10              BMP                     MS Windows Device Independent Bitmap
#> 11              BSB                              Maptech BSB Nautical Charts
#> 12               BT                      VTP .bt (Binary Terrain) 1.3 Format
#> 13              BYN                         Natural Resources Canada's Geoid
#> 14              CAD                                           AutoCAD Driver
#> 15             CALS                                            CALS (Type 1)
#> 16             CEOS                                               CEOS Image
#> 17            COASP                          DRDC COASP SAR Processor Raster
#> 18              COG                        Cloud optimized GeoTIFF generator
#> 19            COSAR               COSAR Annotated Binary Matrix (TerraSAR-X)
#> 20              CPG                                          Convair PolGASP
#> 21          CTable2                                 CTable2 Datum Grid Shift
#> 22              CTG                           USGS LULC Composite Theme Grid
#> 23             DAAS          Airbus DS Intelligence Data As A Service driver
#> 24          DERIVED               Derived datasets using VRT pixel functions
#> 25            DIMAP                                               SPOT DIMAP
#> 26            DIPEx                                                    DIPEx
#> 27             DOQ1                                     USGS DOQ (Old Style)
#> 28             DOQ2                                     USGS DOQ (New Style)
#> 29             DTED                                    DTED Elevation Raster
#> 30          ECRGTOC                                          ECRG TOC format
#> 31            EEDAI                              Earth Engine Data API Image
#> 32             EHdr                                       ESRI .hdr Labelled
#> 33              EIR                                        Erdas Imagine Raw
#> 34             ELAS                                                     ELAS
#> 35             ENVI                                       ENVI .hdr Labelled
#> 36              ERS                                   ERMapper .ers Labelled
#> 37             ESAT                                     Envisat Image Format
#> 38            ESRIC                                       Esri Compact Cache
#> 39             FAST                                        EOSAT FAST Format
#> 40              FIT                                                FIT Image
#> 41          FujiBAS                                   Fuji BAS Scanner Image
#> 42           GenBin                           Generic Binary (.hdr Labelled)
#> 43              GFF Ground-based SAR Applications Testbed File Format (.gff)
#> 44              GIF                       Graphics Interchange Format (.gif)
#> 45              GMT                                   GMT NetCDF Grid Format
#> 46             GPKG                                               GeoPackage
#> 47   GRASSASCIIGrid                                         GRASS ASCII Grid
#> 48             GRIB                             GRIdded Binary (.grb, .grb2)
#> 49            GS7BG                     Golden Software 7 Binary Grid (.grd)
#> 50             GSAG                        Golden Software ASCII Grid (.grd)
#> 51             GSBG                       Golden Software Binary Grid (.grd)
#> 52              GSC                                              GSC Geogrid
#> 53            GTiff                                                  GeoTIFF
#> 54              GTX                                 NOAA Vertical Datum .GTX
#> 55              GXF                             GeoSoft Grid Exchange Format
#> 56             HDF4                       Hierarchical Data Format Release 4
#> 57        HDF4Image                                             HDF4 Dataset
#> 58             HDF5                       Hierarchical Data Format Release 5
#> 59        HDF5Image                                             HDF5 Dataset
#> 60              HF2                               HF2/HFZ heightfield raster
#> 61              HFA                              Erdas Imagine Images (.img)
#> 62             HTTP                                    HTTP Fetching Wrapper
#> 63              IDA                                  Image Data and Analysis
#> 64            ILWIS                                         ILWIS Raster Map
#> 65             INGR                                        Intergraph Raster
#> 66             IRIS                             IRIS data (.PPI, .CAPPi etc)
#> 67             ISCE                                              ISCE raster
#> 68              ISG                      International Service for the Geoid
#> 69            ISIS2                  USGS Astrogeology ISIS cube (Version 2)
#> 70            ISIS3                  USGS Astrogeology ISIS cube (Version 3)
#> 71       JAXAPALSAR               JAXA PALSAR Product Reader (Level 1.1/1.5)
#> 72             JDEM                                      Japanese DEM (.mem)
#> 73      JP2OpenJPEG               JPEG-2000 driver based on OpenJPEG library
#> 74             JPEG                                                JPEG JFIF
#> 75  KMLSUPEROVERLAY                                        Kml Super Overlay
#> 76              KRO                                                KOLOR Raw
#> 77              L1B                     NOAA Polar Orbiter Level 1b Data Set
#> 78              LAN                                          Erdas .LAN/.GIS
#> 79              LCP                        FARSITE v.4 Landscape File (.lcp)
#> 80         Leveller                                     Leveller heightfield
#> 81           LOSLAS                        NADCON .los/.las Datum Grid Shift
#> 82              MAP                                         OziExplorer .MAP
#> 83          MBTiles                                                  MBTiles
#> 84              MEM                                         In Memory Raster
#> 85              MFF                                        Vexcel MFF Raster
#> 86             MFF2                                 Vexcel MFF2 (HKV) Raster
#> 87              MRF                                       Meta Raster Format
#> 88             MSGN                           EUMETSAT Archive native (.nat)
#> 89              NDF                                        NLAPS Data Format
#> 90           netCDF                               Network Common Data Format
#> 91         NGSGEOID                              NOAA NGS Geoid Height Grids
#> 92              NGW                                              NextGIS Web
#> 93             NITF                     National Imagery Transmission Format
#> 94             NTv2                                    NTv2 Datum Grid Shift
#> 95          NWT_GRC               Northwood Classified Grid Format .grc/.tab
#> 96          NWT_GRD                  Northwood Numeric Grid Format .grd/.tab
#> 97           OGCAPI                                                   OGCAPI
#> 98              OZI                                   OziExplorer Image File
#> 99             PAux                                        PCI .aux Labelled
#> 100          PCIDSK                                     PCIDSK Database File
#> 101        PCRaster                                     PCRaster Raster File
#> 102             PDF                                           Geospatial PDF
#> 103             PDS                               NASA Planetary Data System
#> 104            PDS4                             NASA Planetary Data System 4
#> 105        PLMOSAIC                                  Planet Labs Mosaics API
#> 106        PLSCENES                                   Planet Labs Scenes API
#> 107             PNG                                Portable Network Graphics
#> 108             PNM                          Portable Pixmap Format (netpbm)
#> 109   PostGISRaster                                    PostGIS Raster driver
#> 110             PRF                                      Racurs PHOTOMOD PRF
#> 111               R                                      R Object Data Store
#> 112      Rasterlite                                               Rasterlite
#> 113             RDA                   DigitalGlobe Raster Data Access driver
#> 114             RIK                                  Swedish Grid RIK (.rik)
#> 115             RMF                                     Raster Matrix Format
#> 116         ROI_PAC                                           ROI_PAC raster
#> 117          RPFTOC                         Raster Product Format TOC format
#> 118         RRASTER                                                 R Raster
#> 119             RS2                                   RadarSat 2 XML Product
#> 120             RST                                        Idrisi Raster A.1
#> 121            SAFE                              Sentinel-1 SAR SAFE Product
#> 122            SAGA                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)
#> 123        SAR_CEOS                                           CEOS SAR Image
#> 124            SDTS                                              SDTS Raster
#> 125       SENTINEL2                                               Sentinel 2
#> 126             SGI                                SGI Image File Format 1.0
#> 127          SIGDEM                       Scaled Integer Gridded DEM .sigdem
#> 128          SNODAS                            Snow Data Assimilation System
#> 129             SRP                      Standard Raster Product (ASRP/USRP)
#> 130         SRTMHGT                                      SRTMHGT File Format
#> 131          STACTA               Spatio-Temporal Asset Catalog Tiled Assets
#> 132        Terragen                                     Terragen heightfield
#> 133             TGA                              TGA/TARGA Image File Format
#> 134             TIL                                          EarthWatch .TIL
#> 135             TSX                                       TerraSAR-X Product
#> 136         USGSDEM                       USGS Optional ASCII DEM (and CDED)
#> 137           VICAR                                          MIPL VICAR file
#> 138             VRT                                           Virtual Raster
#> 139             WCS                                 OGC Web Coverage Service
#> 140             WMS                                      OGC Web Map Service
#> 141            WMTS                                 OGC Web Map Tile Service
#> 142             XPM                                        X11 PixMap Format
#> 143             XYZ                                        ASCII Gridded XYZ
#> 144            ZMap                                           ZMap Plus Grid
#>     create  copy isRaster
#> 1    FALSE  TRUE     TRUE
#> 2    FALSE FALSE     TRUE
#> 3     TRUE FALSE     TRUE
#> 4    FALSE FALSE     TRUE
#> 5    FALSE FALSE     TRUE
#> 6    FALSE  TRUE     TRUE
#> 7     TRUE  TRUE     TRUE
#> 8    FALSE FALSE     TRUE
#> 9    FALSE  TRUE     TRUE
#> 10    TRUE FALSE     TRUE
#> 11   FALSE FALSE     TRUE
#> 12    TRUE FALSE     TRUE
#> 13    TRUE FALSE     TRUE
#> 14   FALSE FALSE     TRUE
#> 15   FALSE  TRUE     TRUE
#> 16   FALSE FALSE     TRUE
#> 17   FALSE FALSE     TRUE
#> 18   FALSE  TRUE     TRUE
#> 19   FALSE FALSE     TRUE
#> 20   FALSE FALSE     TRUE
#> 21    TRUE FALSE     TRUE
#> 22   FALSE FALSE     TRUE
#> 23   FALSE FALSE     TRUE
#> 24   FALSE FALSE     TRUE
#> 25   FALSE FALSE     TRUE
#> 26   FALSE FALSE     TRUE
#> 27   FALSE FALSE     TRUE
#> 28   FALSE FALSE     TRUE
#> 29   FALSE  TRUE     TRUE
#> 30   FALSE FALSE     TRUE
#> 31   FALSE FALSE     TRUE
#> 32    TRUE  TRUE     TRUE
#> 33   FALSE FALSE     TRUE
#> 34    TRUE FALSE     TRUE
#> 35    TRUE FALSE     TRUE
#> 36    TRUE FALSE     TRUE
#> 37   FALSE FALSE     TRUE
#> 38   FALSE FALSE     TRUE
#> 39   FALSE FALSE     TRUE
#> 40   FALSE  TRUE     TRUE
#> 41   FALSE FALSE     TRUE
#> 42   FALSE FALSE     TRUE
#> 43   FALSE FALSE     TRUE
#> 44   FALSE  TRUE     TRUE
#> 45   FALSE  TRUE     TRUE
#> 46    TRUE  TRUE     TRUE
#> 47   FALSE FALSE     TRUE
#> 48   FALSE  TRUE     TRUE
#> 49    TRUE  TRUE     TRUE
#> 50   FALSE  TRUE     TRUE
#> 51    TRUE  TRUE     TRUE
#> 52   FALSE FALSE     TRUE
#> 53    TRUE  TRUE     TRUE
#> 54    TRUE FALSE     TRUE
#> 55   FALSE FALSE     TRUE
#> 56   FALSE FALSE     TRUE
#> 57    TRUE FALSE     TRUE
#> 58   FALSE FALSE     TRUE
#> 59   FALSE FALSE     TRUE
#> 60   FALSE  TRUE     TRUE
#> 61    TRUE  TRUE     TRUE
#> 62   FALSE FALSE     TRUE
#> 63    TRUE FALSE     TRUE
#> 64    TRUE  TRUE     TRUE
#> 65    TRUE  TRUE     TRUE
#> 66   FALSE FALSE     TRUE
#> 67    TRUE FALSE     TRUE
#> 68   FALSE FALSE     TRUE
#> 69    TRUE FALSE     TRUE
#> 70    TRUE  TRUE     TRUE
#> 71   FALSE FALSE     TRUE
#> 72   FALSE FALSE     TRUE
#> 73   FALSE  TRUE     TRUE
#> 74   FALSE  TRUE     TRUE
#> 75   FALSE  TRUE     TRUE
#> 76    TRUE FALSE     TRUE
#> 77   FALSE FALSE     TRUE
#> 78    TRUE FALSE     TRUE
#> 79   FALSE  TRUE     TRUE
#> 80    TRUE FALSE     TRUE
#> 81   FALSE FALSE     TRUE
#> 82   FALSE FALSE     TRUE
#> 83    TRUE  TRUE     TRUE
#> 84    TRUE FALSE     TRUE
#> 85    TRUE  TRUE     TRUE
#> 86    TRUE  TRUE     TRUE
#> 87    TRUE  TRUE     TRUE
#> 88   FALSE FALSE     TRUE
#> 89   FALSE FALSE     TRUE
#> 90    TRUE  TRUE     TRUE
#> 91   FALSE FALSE     TRUE
#> 92    TRUE  TRUE     TRUE
#> 93    TRUE  TRUE     TRUE
#> 94    TRUE FALSE     TRUE
#> 95   FALSE FALSE     TRUE
#> 96    TRUE  TRUE     TRUE
#> 97   FALSE FALSE     TRUE
#> 98   FALSE FALSE     TRUE
#> 99    TRUE FALSE     TRUE
#> 100   TRUE FALSE     TRUE
#> 101   TRUE  TRUE     TRUE
#> 102   TRUE  TRUE     TRUE
#> 103  FALSE FALSE     TRUE
#> 104   TRUE  TRUE     TRUE
#> 105  FALSE FALSE     TRUE
#> 106  FALSE FALSE     TRUE
#> 107  FALSE  TRUE     TRUE
#> 108   TRUE FALSE     TRUE
#> 109  FALSE  TRUE     TRUE
#> 110  FALSE FALSE     TRUE
#> 111  FALSE  TRUE     TRUE
#> 112  FALSE  TRUE     TRUE
#> 113  FALSE FALSE     TRUE
#> 114  FALSE FALSE     TRUE
#> 115   TRUE FALSE     TRUE
#> 116   TRUE FALSE     TRUE
#> 117  FALSE FALSE     TRUE
#> 118   TRUE  TRUE     TRUE
#> 119  FALSE FALSE     TRUE
#> 120   TRUE  TRUE     TRUE
#> 121  FALSE FALSE     TRUE
#> 122   TRUE  TRUE     TRUE
#> 123  FALSE FALSE     TRUE
#> 124  FALSE FALSE     TRUE
#> 125  FALSE FALSE     TRUE
#> 126   TRUE FALSE     TRUE
#> 127  FALSE  TRUE     TRUE
#> 128  FALSE FALSE     TRUE
#> 129  FALSE FALSE     TRUE
#> 130  FALSE  TRUE     TRUE
#> 131  FALSE FALSE     TRUE
#> 132   TRUE FALSE     TRUE
#> 133  FALSE FALSE     TRUE
#> 134  FALSE FALSE     TRUE
#> 135  FALSE FALSE     TRUE
#> 136  FALSE  TRUE     TRUE
#> 137   TRUE  TRUE     TRUE
#> 138   TRUE  TRUE     TRUE
#> 139  FALSE FALSE     TRUE
#> 140  FALSE  TRUE     TRUE
#> 141  FALSE  TRUE     TRUE
#> 142  FALSE  TRUE     TRUE
#> 143  FALSE  TRUE     TRUE
#> 144  FALSE  TRUE     TRUE

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)

#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x68c51e0>
#> 

getDriverLongName(getDriver(x))

#> [1] "JPEG JFIF"

GDAL.close(x)