Class "GDALMajorObject"

"GDALMajorObject" is a virtual base class for all GDAL objects.

getDescription(object)

Objects from the Class

Objects can be created by calls of the form new("GDALMajorObject", ...), but are only created for classes that extend this class.

Slots

handle:

Object of class "externalptr", used internally; not for public consumption

Methods

No methods defined with class "GDALMajorObject" in the signature.

Arguments

object

an object inheriting from "GDALMajorObject"

Details

getDescription:

returns a descrption string associated with the object. No setter method is defined because GDAL dataset objects use the description to hold the filename attached to the dataset. It would not be good to change that mid-stream.

References

https://gdal.org/

Author

Timothy H. Keitt, modified by Roger Bivand

See also

Examples

driver <- new('GDALDriver', as.character(getGDALDriverNames()[1,1]))
driver
#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x556e470>
#> 
rm(driver)
logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
x
#> An object of class "GDALReadOnlyDataset"
#> Slot "handle":
#> <pointer: 0x4cf9dd0>
#> 
getDescription(x)
#> [1] "/tmp/RtmpnkEu0w/temp_libpath101e75200293a/rgdal/pictures/logo.jpg"
dim(x)
#> [1] 175 200
GDAL.close(x)