GDALReadOnlyDataset-class {rgdal}	R Documentation

Class "GDALReadOnlyDataset"

Description

GDALReadOnlyDataset is the base class for a GDAL Dataset classes. Only read operations are supported. Both GDALDataset and GDALTransientDataset inherit these read operations while providing additional write operations (see GDALDataset-class). GDALReadOnlyDataset-class inherits from GDALMajorObject-class.

Usage

GDAL.close(dataset)
GDAL.open(filename, read.only = TRUE, silent=FALSE,
                 allowedDrivers = NULL, options=NULL)
getDriver(dataset)

getColorTable(dataset, band = 1)
getGeoTransFunc(dataset)

Arguments

dataset
An object inheriting from class 'GDALReadOnlyDataset'

filename
name of file to contain raster data object; will be normalized with normalizePath if it is a file

band
The band number (1-based) to read from

read.only
A logical flag indicating whether to open the file as a GDALReadOnlyDataset or as a writable GDALDataset

silent
logical; if TRUE, comment and non-fatal CPL driver errors suppressed

allowedDrivers
a character vector of suggested driver short names may be provided starting from GDAL 2.0

options
open options may be passed to raster drivers starting from GDAL 2.0; very few drivers support these options

Details

GDAL.open and GDAL.close are shorter versions of new("GDALReadOnlyDataset", ...) and closeDataset(). Because GDAL.close through closeDataset() uses the finalization mechanism to destroy the handles to the dataset and its driver, messages such as:

"Closing GDAL dataset handle 0x8ff7900... destroyed ... done."

may appear when GDAL.close is run, or at some later stage. getDriver returns an object inheriting from class 'GDALDriver'. getColorTable returns the dataset colour table (currently does not support RGB imaging). getGeoTransFunc returns a warping function.

Objects from the Class

Objects can be created by calls of the form new("GDALReadOnlyDataset", filename, handle). ~~ describe objects here ~~

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject" ~~

Extends

Class "GDALMajorObject", directly.

Methods

closeDataset

signature(dataset = "GDALReadOnlyDataset"): ...

dim

signature(x = "GDALReadOnlyDataset"): ...

initialize

signature(.Object = "GDALReadOnlyDataset"): ...

Author(s)

Timothy H. Keitt, modified by Roger Bivand

References

http://www.gdal.org/

See Also

See also GDALDriver-class, GDALDataset-class, GDALTransientDataset-class.

Examples

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
dim(x)
plot(density(getRasterTable(x)$band1))
#displayDataset(x)
#displayDataset(x, col=function(x){rev(cm.colors(x))})
#im <- displayDataset(x, col=function(x){rev(cm.colors(x))}, reset.par=FALSE)
#contour(1:attr(im, "size")[2], 1:attr(im, "size")[1],
# t(attr(im, "index"))[,attr(im, "size")[1]:1], nlevels = 1,
# levels = 100, col = 'black', add = TRUE)
GDAL.close(x)
logo <- system.file("pictures/Rlogo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
dim(x)
#displayDataset(x)
GDAL.close(x)

[Package rgdal version 1.4-8 Index]