GDALTransientDataset-class {rgdal}	R Documentation

Class "GDALTransientDataset"

Description

GDALTransientDataset is identical to GDALDataset-class except that transient datasets are not associated with any user-visible file. Transient datasets delete their associated file data when closed. See saveDataset and saveDatasetAs.

Objects from the Class

Objects can be created by calls of the form new("GDALTransientDataset", driver, rows, cols, bands, type, options, fname, handle).

driver

A "GDALDriver" object that determines the storage format

rows

Number of rows in the newly created dataset

cols

Number of columns in the newly created dataset

bands

Number of bands to create

type

A GDAL type name as listed in .GDALDataTypes

options

Driver specific options

fname

default NULL, used internally to pass through a file name with a required extension (RST driver has this problem)

handle

Used internally; not for public consumption

Slots

handle:

Object of class "externalptr", from class "GDALDataset", used internally; not for public consumption

Extends

Class "GDALDataset", directly. Class "GDALReadOnlyDataset", by class "GDALDataset". Class "GDALMajorObject", by class "GDALDataset".

Methods

closeDataset

signature(dataset = "GDALTransientDataset"): ...

initialize

signature(.Object = "GDALTransientDataset"): ...

Author(s)

Timothy H. Keitt, modified by Roger Bivand

See Also

See also GDALDriver-class, GDALReadOnlyDataset-class

Examples

list.files(tempdir())
x <- new('GDALTransientDataset', driver=new('GDALDriver', "GTiff"), rows=100,
 cols=100, bands=3, type='Byte')
dim(x)
list.files(tempdir())
GDAL.close(x)
list.files(tempdir())

[Package rgdal version 1.4-8 Index]