is_proj_CDN_enabled {rgdal}	R Documentation

PROJ search paths and content download network handling

Description

From PROJ 7 (and partly 7.1), it is becoming possible to use transformation grids downloaded on demand to improve coordinate operation accuracy from a content download network (CDN). These functions report on and control the use of the CDN.

Usage

is_proj_CDN_enabled()
enable_proj_CDN()
disable_proj_CDN()
proj_CDN_user_writable_dir()
get_proj_search_paths()
set_proj_search_paths(paths)

Arguments

paths
a character vector of existing directories

Details

The PROJ user-writable CDN directory is set as soon as the internal search path is queried, and for most uses, the default value will allow all programs using PROJ such as R packages, QGIS, GRASS, etc., to access any downloaded grids. Grids are checked for staleness at regular intervals. This directory may be set to a non-default value with the PROJ_USER_WRITABLE_DIRECTORY environment variable before rgdal (and any other package using PROJ) is loaded and attached, from PROJ >= 7.1.0.

Value

Logical values and/or character vector search paths, often NULL for earlier versions of PROJ.

Author(s)

Roger Bivand

References

https://cdn.proj.org/.

Examples

is_proj_CDN_enabled()
proj_CDN_user_writable_dir()
get_proj_search_paths()

[Package rgdal version 1.5-25 Index]