llgridlines {rgdal}	R Documentation

Plot long-lat grid over projected data

Description

Plot long-lat grid over projected data

Usage

llgridlines(obj, easts, norths, ndiscr = 20, lty = 2, offset=0.5, side="WS",
llcrs = "+proj=longlat +datum=WGS84", plotLines = TRUE, plotLabels =
TRUE, ...)

Arguments

obj
object, deriving from Spatial having projection specified

easts
numeric; see gridlines

norths
numeric; see gridlines

ndiscr
numeric; see gridlines

offset
numeric; see gridat

side
character, default “WS”; see gridat; available from sp 0.9-84

lty
line type to be used for grid lines

llcrs
proj4string of longitude - latitude

plotLines
logical; plot lines?

plotLabels
logical; plot labels?

...
graphics arguments passed to plot function for lines and text function for labels

Value

none; side effect is that grid lines and lables are plotted

See Also

is.projected, CRS-class

Examples

data(meuse)
coordinates(meuse) = ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
plot(meuse) 
llgridlines(meuse, lty=3)
plot(meuse) 
llgridlines(meuse, lty=3, side = "EN", offset = 0.2)

[Package rgdal version 1.4-8 Index]