List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()
GDAL_OSR_PROJ()
GDALis3ormore()
PROJis6ormore()
new_proj_and_gdal()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu. From PROJ 6, "datum" is not available. From PROJ 7.1.0, "units" returns the conversion factor as numeric, not character.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

https://proj.org/

Author

Roger Bivand Roger.Bivand@nhh.no

Note

Loading the rgdal package may change the PROJ\_LIB environmental variable to the PROJ.4 support files if bundled with binary packages.

Examples

getPROJ4VersionInfo()

#> [1] "Rel. 8.1.1, September 1st, 2021, [PJ_VERSION: 811]"
#> attr(,"short")
#> [1] 811

projInfo()

#>             name
#> 1     adams_hemi
#> 2      adams_ws1
#> 3      adams_ws2
#> 4            aea
#> 5           aeqd
#> 6         affine
#> 7           airy
#> 8         aitoff
#> 9           alsk
#> 10         apian
#> 11        august
#> 12      axisswap
#> 13         bacon
#> 14    bertin1953
#> 15          bipc
#> 16         boggs
#> 17         bonne
#> 18       calcofi
#> 19          cart
#> 20          cass
#> 21            cc
#> 22          ccon
#> 23           cea
#> 24         chamb
#> 25         collg
#> 26     col_urban
#> 27        comill
#> 28         crast
#> 29      defmodel
#> 30   deformation
#> 31         denoy
#> 32          eck1
#> 33          eck2
#> 34          eck3
#> 35          eck4
#> 36          eck5
#> 37          eck6
#> 38       eqearth
#> 39           eqc
#> 40          eqdc
#> 41         euler
#> 42        etmerc
#> 43         fahey
#> 44          fouc
#> 45        fouc_s
#> 46          gall
#> 47          geoc
#> 48    geogoffset
#> 49          geos
#> 50         gins8
#> 51       gn_sinu
#> 52          gnom
#> 53         goode
#> 54          gs48
#> 55          gs50
#> 56         guyou
#> 57        hammer
#> 58        hatano
#> 59       healpix
#> 60      rhealpix
#> 61       helmert
#> 62    hgridshift
#> 63        horner
#> 64           igh
#> 65         igh_o
#> 66         imw_p
#> 67          isea
#> 68          kav5
#> 69          kav7
#> 70        krovak
#> 71         labrd
#> 72          laea
#> 73        lagrng
#> 74          larr
#> 75          lask
#> 76        lonlat
#> 77        latlon
#> 78           lcc
#> 79          lcca
#> 80          leac
#> 81        lee_os
#> 82         loxim
#> 83          lsat
#> 84         mbt_s
#> 85       mbt_fps
#> 86        mbtfpp
#> 87        mbtfpq
#> 88        mbtfps
#> 89          merc
#> 90        mil_os
#> 91          mill
#> 92       misrsom
#> 93          moll
#> 94   molobadekas
#> 95    molodensky
#> 96         murd1
#> 97         murd2
#> 98         murd3
#> 99      natearth
#> 100    natearth2
#> 101         nell
#> 102       nell_h
#> 103        nicol
#> 104        nsper
#> 105         nzmg
#> 106         noop
#> 107      ob_tran
#> 108         ocea
#> 109          oea
#> 110        omerc
#> 111        ortel
#> 112        ortho
#> 113       pconic
#> 114    patterson
#> 115     peirce_q
#> 116     pipeline
#> 117         poly
#> 118          pop
#> 119         push
#> 120        putp1
#> 121        putp2
#> 122        putp3
#> 123       putp3p
#> 124       putp4p
#> 125        putp5
#> 126       putp5p
#> 127        putp6
#> 128       putp6p
#> 129      qua_aut
#> 130          qsc
#> 131        robin
#> 132        rouss
#> 133        rpoly
#> 134          sch
#> 135          set
#> 136         sinu
#> 137       somerc
#> 138        stere
#> 139       sterea
#> 140      gstmerc
#> 141          tcc
#> 142         tcea
#> 143        times
#> 144     tinshift
#> 145       tissot
#> 146        tmerc
#> 147      tobmerc
#> 148  topocentric
#> 149        tpeqd
#> 150        tpers
#> 151  unitconvert
#> 152          ups
#> 153         urm5
#> 154       urmfps
#> 155          utm
#> 156        vandg
#> 157       vandg2
#> 158       vandg3
#> 159       vandg4
#> 160        vitk1
#> 161   vgridshift
#> 162         wag1
#> 163         wag2
#> 164         wag3
#> 165         wag4
#> 166         wag5
#> 167         wag6
#> 168         wag7
#> 169      webmerc
#> 170        weren
#> 171        wink1
#> 172        wink2
#> 173       wintri
#> 174 xyzgridshift
#>                                                         description
#> 1                                      Adams Hemisphere in a Square
#> 2                                         Adams World in a Square I
#> 3                                        Adams World in a Square II
#> 4                                                 Albers Equal Area
#> 5                                             Azimuthal Equidistant
#> 6                                             Affine transformation
#> 7                                                              Airy
#> 8                                                            Aitoff
#> 9                                  Modified Stereographic of Alaska
#> 10                                                 Apian Globular I
#> 11                                              August Epicycloidal
#> 12                                                    Axis ordering
#> 13                                                   Bacon Globular
#> 14                                                      Bertin 1953
#> 15                              Bipolar conic of western hemisphere
#> 16                                                  Boggs Eumorphic
#> 17                                          Bonne (Werner lat_1=90)
#> 18                        Cal Coop Ocean Fish Invest Lines/Stations
#> 19                                   Geodetic/cartesian conversions
#> 20                                                          Cassini
#> 21                                              Central Cylindrical
#> 22                                                    Central Conic
#> 23                                           Equal Area Cylindrical
#> 24                                             Chamberlin Trimetric
#> 25                                                        Collignon
#> 26                                                   Colombia Urban
#> 27                                                   Compact Miller
#> 28                                   Craster Parabolic (Putnins P4)
#> 29                                                Deformation model
#> 30                                             Kinematic grid shift
#> 31                                          Denoyer Semi-Elliptical
#> 32                                                         Eckert I
#> 33                                                        Eckert II
#> 34                                                       Eckert III
#> 35                                                        Eckert IV
#> 36                                                         Eckert V
#> 37                                                        Eckert VI
#> 38                                                      Equal Earth
#> 39                           Equidistant Cylindrical (Plate Carree)
#> 40                                                Equidistant Conic
#> 41                                                            Euler
#> 42                                     Extended Transverse Mercator
#> 43                                                            Fahey
#> 44                                                          Foucaut
#> 45                                               Foucaut Sinusoidal
#> 46                                        Gall (Gall Stereographic)
#> 47                                              Geocentric Latitude
#> 48                                                Geographic Offset
#> 49                                     Geostationary Satellite View
#> 50                                        Ginsburg VIII (TsNIIGAiK)
#> 51                                        General Sinusoidal Series
#> 52                                                         Gnomonic
#> 53                                                 Goode Homolosine
#> 54                                Modified Stereographic of 48 U.S.
#> 55                                Modified Stereographic of 50 U.S.
#> 56                                                            Guyou
#> 57                                     Hammer & Eckert-Greifendorff
#> 58                                   Hatano Asymmetrical Equal Area
#> 59                                                          HEALPix
#> 60                                                         rHEALPix
#> 61                   3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 62                                            Horizontal grid shift
#> 63                                     Horner polynomial evaluation
#> 64                                     Interrupted Goode Homolosine
#> 65                        Interrupted Goode Homolosine Oceanic View
#> 66                         International Map of the World Polyconic
#> 67                                    Icosahedral Snyder Equal Area
#> 68                                                      Kavraisky V
#> 69                                                    Kavraisky VII
#> 70                                                           Krovak
#> 71                                                          Laborde
#> 72                                     Lambert Azimuthal Equal Area
#> 73                                                         Lagrange
#> 74                                                         Larrivee
#> 75                                                        Laskowski
#> 76                                              Lat/long (Geodetic)
#> 77                                        Lat/long (Geodetic alias)
#> 78                                          Lambert Conformal Conic
#> 79                              Lambert Conformal Conic Alternative
#> 80                                         Lambert Equal Area Conic
#> 81                                        Lee Oblated Stereographic
#> 82                                                       Loximuthal
#> 83                                        Space oblique for LANDSAT
#> 84                           McBryde-Thomas Flat-Polar Sine (No. 1)
#> 85                            McBryde-Thomas Flat-Pole Sine (No. 2)
#> 86                              McBride-Thomas Flat-Polar Parabolic
#> 87                                McBryde-Thomas Flat-Polar Quartic
#> 88                             McBryde-Thomas Flat-Polar Sinusoidal
#> 89                                                         Mercator
#> 90                                     Miller Oblated Stereographic
#> 91                                               Miller Cylindrical
#> 92                                           Space oblique for MISR
#> 93                                                        Mollweide
#> 94                                Molodensky-Badekas transformation
#> 95                                             Molodensky transform
#> 96                                                        Murdoch I
#> 97                                                       Murdoch II
#> 98                                                      Murdoch III
#> 99                                                    Natural Earth
#> 100                                                 Natural Earth 2
#> 101                                                            Nell
#> 102                                                     Nell-Hammer
#> 103                                               Nicolosi Globular
#> 104                                          Near-sided perspective
#> 105                                            New Zealand Map Grid
#> 106                                                    No operation
#> 107                                  General Oblique Transformation
#> 108                                  Oblique Cylindrical Equal Area
#> 109                                              Oblated Equal Area
#> 110                                                Oblique Mercator
#> 111                                                   Ortelius Oval
#> 112                                                    Orthographic
#> 113                                               Perspective Conic
#> 114                                           Patterson Cylindrical
#> 115                                              Peirce Quincuncial
#> 116                                 Transformation pipeline manager
#> 117                                            Polyconic (American)
#> 118                   Retrieve coordinate value from pipeline stack
#> 119                         Save coordinate value on pipeline stack
#> 120                                                      Putnins P1
#> 121                                                      Putnins P2
#> 122                                                      Putnins P3
#> 123                                                     Putnins P3'
#> 124                                                     Putnins P4'
#> 125                                                      Putnins P5
#> 126                                                     Putnins P5'
#> 127                                                      Putnins P6
#> 128                                                     Putnins P6'
#> 129                                                Quartic Authalic
#> 130                                Quadrilateralized Spherical Cube
#> 131                                                        Robinson
#> 132                                         Roussilhe Stereographic
#> 133                                           Rectangular Polyconic
#> 134                                    Spherical Cross-track Height
#> 135                                            Set coordinate value
#> 136                                   Sinusoidal (Sanson-Flamsteed)
#> 137                                            Swiss. Obl. Mercator
#> 138                                                   Stereographic
#> 139                               Oblique Stereographic Alternative
#> 140 Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 141                                  Transverse Central Cylindrical
#> 142                               Transverse Cylindrical Equal Area
#> 143                                                           Times
#> 144                              Triangulation based transformation
#> 145                                                          Tissot
#> 146                                             Transverse Mercator
#> 147                                                 Tobler-Mercator
#> 148                               Geocentric/Topocentric conversion
#> 149                                           Two Point Equidistant
#> 150                                              Tilted perspective
#> 151                                                 Unit conversion
#> 152                                   Universal Polar Stereographic
#> 153                                                        Urmaev V
#> 154                                    Urmaev Flat-Polar Sinusoidal
#> 155                             Universal Transverse Mercator (UTM)
#> 156                                             van der Grinten (I)
#> 157                                              van der Grinten II
#> 158                                             van der Grinten III
#> 159                                              van der Grinten IV
#> 160                                                     Vitkovsky I
#> 161                                             Vertical grid shift
#> 162                                         Wagner I (Kavraisky VI)
#> 163                                                       Wagner II
#> 164                                                      Wagner III
#> 165                                                       Wagner IV
#> 166                                                        Wagner V
#> 167                                                       Wagner VI
#> 168                                                      Wagner VII
#> 169                                  Web Mercator / Pseudo Mercator
#> 170                                                   Werenskiold I
#> 171                                                        Winkel I
#> 172                                                       Winkel II
#> 173                                                   Winkel Tripel
#> 174                                           Geocentric grid shift

projInfo("ellps")

#>         name          major                  ell
#> 1      MERIT    a=6378137.0           rf=298.257
#> 2      SGS85    a=6378136.0           rf=298.257
#> 3      GRS80    a=6378137.0     rf=298.257222101
#> 4      IAU76    a=6378140.0           rf=298.257
#> 5       airy  a=6377563.396       rf=299.3249646
#> 6     APL4.9    a=6378137.0            rf=298.25
#> 7      NWL9D    a=6378145.0            rf=298.25
#> 8   mod_airy  a=6377340.189        b=6356034.446
#> 9     andrae   a=6377104.43             rf=300.0
#> 10    danish a=6377019.2563             rf=300.0
#> 11   aust_SA    a=6378160.0            rf=298.25
#> 12     GRS67    a=6378160.0    rf=298.2471674270
#> 13   GSK2011    a=6378136.5       rf=298.2564151
#> 14    bessel  a=6377397.155       rf=299.1528128
#> 15  bess_nam  a=6377483.865       rf=299.1528128
#> 16    clrk66    a=6378206.4          b=6356583.8
#> 17    clrk80  a=6378249.145          rf=293.4663
#> 18 clrk80ign    a=6378249.2 rf=293.4660212936269
#> 19       CPM    a=6375738.7            rf=334.29
#> 20    delmbr     a=6376428.             rf=311.5
#> 21   engelis   a=6378136.05          rf=298.2566
#> 22   evrst30  a=6377276.345          rf=300.8017
#> 23   evrst48  a=6377304.063          rf=300.8017
#> 24   evrst56  a=6377301.243          rf=300.8017
#> 25   evrst69  a=6377295.664          rf=300.8017
#> 26   evrstSS  a=6377298.556          rf=300.8017
#> 27   fschr60     a=6378166.             rf=298.3
#> 28  fschr60m     a=6378155.             rf=298.3
#> 29   fschr68     a=6378150.             rf=298.3
#> 30   helmert     a=6378200.             rf=298.3
#> 31     hough    a=6378270.0              rf=297.
#> 32      intl    a=6378388.0              rf=297.
#> 33     krass    a=6378245.0             rf=298.3
#> 34     kaula     a=6378163.            rf=298.24
#> 35     lerch     a=6378139.           rf=298.257
#> 36     mprts     a=6397300.              rf=191.
#> 37  new_intl    a=6378157.5          b=6356772.2
#> 38   plessis     a=6376523.           b=6355863.
#> 39      PZ90    a=6378136.0         rf=298.25784
#> 40    SEasia    a=6378155.0       b=6356773.3205
#> 41   walbeck    a=6376896.0       b=6355834.8467
#> 42     WGS60    a=6378165.0             rf=298.3
#> 43     WGS66    a=6378145.0            rf=298.25
#> 44     WGS72    a=6378135.0            rf=298.26
#> 45     WGS84    a=6378137.0     rf=298.257223563
#> 46    sphere    a=6370997.0          b=6370997.0
#>                                description
#> 1                               MERIT 1983
#> 2                Soviet Geodetic System 85
#> 3                     GRS 1980(IUGG, 1980)
#> 4                                 IAU 1976
#> 5                                Airy 1830
#> 6                      Appl. Physics. 1965
#> 7                 Naval Weapons Lab., 1965
#> 8                            Modified Airy
#> 9               Andrae 1876 (Den., Iclnd.)
#> 10          Andrae 1876 (Denmark, Iceland)
#> 11         Australian Natl & S. Amer. 1969
#> 12                       GRS 67(IUGG 1967)
#> 13                                GSK-2011
#> 14                             Bessel 1841
#> 15                   Bessel 1841 (Namibia)
#> 16                             Clarke 1866
#> 17                        Clarke 1880 mod.
#> 18                      Clarke 1880 (IGN).
#> 19         Comm. des Poids et Mesures 1799
#> 20                 Delambre 1810 (Belgium)
#> 21                            Engelis 1985
#> 22                            Everest 1830
#> 23                            Everest 1948
#> 24                            Everest 1956
#> 25                            Everest 1969
#> 26               Everest (Sabah & Sarawak)
#> 27            Fischer (Mercury Datum) 1960
#> 28                   Modified Fischer 1960
#> 29                            Fischer 1968
#> 30                            Helmert 1906
#> 31                                   Hough
#> 32 International 1924 (Hayford 1909, 1910)
#> 33                        Krassovsky, 1942
#> 34                              Kaula 1961
#> 35                              Lerch 1979
#> 36                         Maupertius 1738
#> 37                  New International 1967
#> 38                   Plessis 1817 (France)
#> 39                                   PZ-90
#> 40                          Southeast Asia
#> 41                                 Walbeck
#> 42                                  WGS 60
#> 43                                  WGS 66
#> 44                                  WGS 72
#> 45                                  WGS 84
#> 46               Normal Sphere (r=6370997)

projInfo("units")

#>        id     to_meter               name
#> 1      mm 1.000000e-03         millimetre
#> 2      cm 1.000000e-02         centimetre
#> 3       m 1.000000e+00              metre
#> 4      ft 3.048000e-01               foot
#> 5   us-ft 3.048006e-01     US survey foot
#> 6    fath 1.828800e+00             fathom
#> 7     kmi 1.852000e+03      nautical mile
#> 8   us-ch 2.011684e+01    US survey chain
#> 9   us-mi 1.609347e+03     US survey mile
#> 10     km 1.000000e+03          kilometre
#> 11 ind-ft 3.047984e-01 Indian foot (1937)
#> 12 ind-yd 9.143952e-01 Indian yard (1937)
#> 13     mi 1.609344e+03       Statute mile
#> 14     yd 9.144000e-01               yard
#> 15     ch 2.011680e+01              chain
#> 16   link 2.011680e-01               link
#> 17     dm 1.000000e-02          decimeter
#> 18     in 2.540000e-02               inch
#> 19 ind-ch 2.011670e+01       Indian chain
#> 20  us-in 2.540005e-02     US survey inch
#> 21  us-yd 9.144018e-01     US survey yard