GDALcall {rgdal}	R Documentation

Wrapper functions to allow more direct calling of rgdal C code

Description

These functions allow more direct access to some of the rgdal C API. These are advanced methods intended for package developers only.

Usage

GDALcall(object, option, ...)
rawTransform(projfrom, projto, n, x, y, z=NULL)

Arguments

object
GDALTransientDataset (option = 'SetGeoTransform', 'SetProject') or GDALRasterBand (the other options)

option
character. One of 'SetGeoTransform', 'SetProject', 'SetNoDataValue', 'SetStatistics', 'SetRasterColorTable' or 'SetCategoryNames')

...
additional arguments. The values to be set

projfrom
character. PROJ.4 coordinate reference system (CRS) description

projto
character. PROJ.4 CRS description

n
number of coordinates

x
x coordinates

y
y coordinates

z
z coordinates

Value

GDALcall does not return anything. rawTransform returns a matrix of transformed coordinates.

Author(s)

Robert Hijmans

[Package rgdal version 1.4-8 Index]