GDALDataset-class {rgdal}	R Documentation

Class "GDALDataset"

Description

GDALDataset extends GDALReadOnlyDataset-class with data update commands.

Usage

putRasterData(dataset, rasterData, band = 1, offset = c(0, 0))
saveDataset(dataset, filename, options=NULL, returnNewObj=FALSE)
copyDataset(dataset, driver, strict = FALSE, options = NULL, fname=NULL)
deleteDataset(dataset)
saveDatasetAs(dataset, filename, driver = NULL, options=NULL)

Arguments

dataset
An object inheriting from class 'GDALDataset'

rasterData
A data array with length(dim(rasterData)) = 2

band
The band number (1-based) to read from

offset
Number of rows and columns from the origin (usually the upper left corner) to begin reading from

filename
name of file to contain raster data object; will be normalized with normalizePath

returnNewObj
until and including 0.5-27, saveDataset returned an invisible copy of the new file handle, which was then only finalized when the garbage collector ran. The old behaviour can be retained by setting to FALSE, the default behaviour is to close the hande and not return it.

driver
GDAL driver name to use for saving raster data object

strict
TRUE if the copy must be strictly equivalent, or more normally FALSE indicating that the copy may adapt as needed for the output format

options
Driver specific options (currently passed to GDAL)

fname
default NULL, used internally to pass through a file name with a required extension (RST driver has this problem)

Details

putRasterData:

writes data contained in rasterData to the dataset, begining at offset rows and columns from the origin (usually the upper left corner). Data type conversion is automatic.

saveDataset:

saves a raster data object in a file using the driver of the object

saveDatasetAs:

saves a raster data object in a file using the specified driver

copyDataset:

make a copy of raster data object in a file using the specified driver

deleteDataset:

delete the file from which the raster data object was read (should only delete files opened as GDALDataset objects

Objects from the Class

Objects can be created by calls of the form new("GDALDataset", filename, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALReadOnlyDataset", used internally; not for public consumption

Extends

Class "GDALReadOnlyDataset", directly. Class "GDALMajorObject", by class "GDALReadOnlyDataset".

Methods

initialize

signature(.Object = "GDALDataset"): ...

Author(s)

Timothy H. Keitt, modified by Roger Bivand

See Also

GDALDriver-class, GDALReadOnlyDataset-class, GDALTransientDataset-class

[Package rgdal version 1.4-8 Index]