Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
get_cached_orig_PROJ_LIB()
get_cached_orig_GDAL_DATA()
get_cached_set_PROJ_LIB()
get_cached_set_GDAL_DATA()

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Arguments

driver

An object inheriting from class 'GDALDriver'

str

A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

get_cached_orig_PROJ_LIB, get_cached_orig_GDAL_DATA

The values of environment variables PROJ_LIB and GDAL_DATA as read when this package was loaded

get_cached_set_PROJ_LIB, get_cached_set_GDAL_DATA

If not "", the values set when loading this package to point to metadata files included in CRAN binary packages

Note

Loading the rgdal package changes the GDAL_DATA environmental variable to the GDAL support files bundled with the package.

Author

Timothy H. Keitt, modified by Roger Bivand

See also

Examples

gdalDrivers()
#>                name                                                long_name
#> 1           AAIGrid                                      Arc/Info ASCII Grid
#> 2              ACE2                                                     ACE2
#> 3              ADRG                            ARC Digitized Raster Graphics
#> 4               AIG                                     Arc/Info Binary Grid
#> 5               ARG                                Azavea Raster Grid format
#> 6            AirSAR                                AirSAR Polarimetric Image
#> 7               BAG                               Bathymetry Attributed Grid
#> 8            BIGGIF                       Graphics Interchange Format (.gif)
#> 9               BLX                                     Magellan topo (.blx)
#> 10              BMP                     MS Windows Device Independent Bitmap
#> 11              BSB                              Maptech BSB Nautical Charts
#> 12               BT                      VTP .bt (Binary Terrain) 1.3 Format
#> 13              BYN                         Natural Resources Canada's Geoid
#> 14              CAD                                           AutoCAD Driver
#> 15             CALS                                            CALS (Type 1)
#> 16             CEOS                                               CEOS Image
#> 17            COASP                          DRDC COASP SAR Processor Raster
#> 18              COG                        Cloud optimized GeoTIFF generator
#> 19            COSAR               COSAR Annotated Binary Matrix (TerraSAR-X)
#> 20              CPG                                          Convair PolGASP
#> 21              CTG                           USGS LULC Composite Theme Grid
#> 22          CTable2                                 CTable2 Datum Grid Shift
#> 23             DAAS          Airbus DS Intelligence Data As A Service driver
#> 24          DERIVED               Derived datasets using VRT pixel functions
#> 25            DIMAP                                               SPOT DIMAP
#> 26            DIPEx                                                    DIPEx
#> 27             DOQ1                                     USGS DOQ (Old Style)
#> 28             DOQ2                                     USGS DOQ (New Style)
#> 29             DTED                                    DTED Elevation Raster
#> 30          ECRGTOC                                          ECRG TOC format
#> 31            EEDAI                              Earth Engine Data API Image
#> 32             EHdr                                       ESRI .hdr Labelled
#> 33              EIR                                        Erdas Imagine Raw
#> 34             ELAS                                                     ELAS
#> 35             ENVI                                       ENVI .hdr Labelled
#> 36              ERS                                   ERMapper .ers Labelled
#> 37             ESAT                                     Envisat Image Format
#> 38            ESRIC                                       Esri Compact Cache
#> 39             FAST                                        EOSAT FAST Format
#> 40              FIT                                                FIT Image
#> 41              GFF Ground-based SAR Applications Testbed File Format (.gff)
#> 42              GIF                       Graphics Interchange Format (.gif)
#> 43             GPKG                                               GeoPackage
#> 44   GRASSASCIIGrid                                         GRASS ASCII Grid
#> 45             GRIB                             GRIdded Binary (.grb, .grb2)
#> 46            GS7BG                     Golden Software 7 Binary Grid (.grd)
#> 47             GSAG                        Golden Software ASCII Grid (.grd)
#> 48             GSBG                       Golden Software Binary Grid (.grd)
#> 49              GSC                                              GSC Geogrid
#> 50              GTX                                 NOAA Vertical Datum .GTX
#> 51            GTiff                                                  GeoTIFF
#> 52              GXF                             GeoSoft Grid Exchange Format
#> 53           GenBin                           Generic Binary (.hdr Labelled)
#> 54             HDF4                       Hierarchical Data Format Release 4
#> 55        HDF4Image                                             HDF4 Dataset
#> 56             HDF5                       Hierarchical Data Format Release 5
#> 57        HDF5Image                                             HDF5 Dataset
#> 58              HF2                               HF2/HFZ heightfield raster
#> 59              HFA                              Erdas Imagine Images (.img)
#> 60             HTTP                                    HTTP Fetching Wrapper
#> 61            ILWIS                                         ILWIS Raster Map
#> 62             IRIS                             IRIS data (.PPI, .CAPPi etc)
#> 63             ISCE                                              ISCE raster
#> 64              ISG                      International Service for the Geoid
#> 65            ISIS2                  USGS Astrogeology ISIS cube (Version 2)
#> 66            ISIS3                  USGS Astrogeology ISIS cube (Version 3)
#> 67       JAXAPALSAR               JAXA PALSAR Product Reader (Level 1.1/1.5)
#> 68             JDEM                                      Japanese DEM (.mem)
#> 69      JP2OpenJPEG               JPEG-2000 driver based on OpenJPEG library
#> 70             JPEG                                                JPEG JFIF
#> 71              KEA                                  KEA Image Format (.kea)
#> 72  KMLSUPEROVERLAY                                        Kml Super Overlay
#> 73              KRO                                                KOLOR Raw
#> 74              L1B                     NOAA Polar Orbiter Level 1b Data Set
#> 75              LAN                                          Erdas .LAN/.GIS
#> 76              LCP                        FARSITE v.4 Landscape File (.lcp)
#> 77           LOSLAS                        NADCON .los/.las Datum Grid Shift
#> 78         Leveller                                     Leveller heightfield
#> 79              MAP                                         OziExplorer .MAP
#> 80          MBTiles                                                  MBTiles
#> 81              MEM                                         In Memory Raster
#> 82              MFF                                        Vexcel MFF Raster
#> 83             MFF2                                 Vexcel MFF2 (HKV) Raster
#> 84              MRF                                       Meta Raster Format
#> 85             MSGN                           EUMETSAT Archive native (.nat)
#> 86              NDF                                        NLAPS Data Format
#> 87         NGSGEOID                              NOAA NGS Geoid Height Grids
#> 88              NGW                                              NextGIS Web
#> 89             NITF                     National Imagery Transmission Format
#> 90             NTv2                                    NTv2 Datum Grid Shift
#> 91          NWT_GRC               Northwood Classified Grid Format .grc/.tab
#> 92          NWT_GRD                  Northwood Numeric Grid Format .grd/.tab
#> 93           OGCAPI                                                   OGCAPI
#> 94              OZI                                   OziExplorer Image File
#> 95             PAux                                        PCI .aux Labelled
#> 96           PCIDSK                                     PCIDSK Database File
#> 97         PCRaster                                     PCRaster Raster File
#> 98              PDF                                           Geospatial PDF
#> 99              PDS                               NASA Planetary Data System
#> 100            PDS4                             NASA Planetary Data System 4
#> 101        PLMOSAIC                                  Planet Labs Mosaics API
#> 102        PLSCENES                                   Planet Labs Scenes API
#> 103             PNG                                Portable Network Graphics
#> 104             PNM                          Portable Pixmap Format (netpbm)
#> 105             PRF                                      Racurs PHOTOMOD PRF
#> 106   PostGISRaster                                    PostGIS Raster driver
#> 107               R                                      R Object Data Store
#> 108             RIK                                  Swedish Grid RIK (.rik)
#> 109             RMF                                     Raster Matrix Format
#> 110         ROI_PAC                                           ROI_PAC raster
#> 111          RPFTOC                         Raster Product Format TOC format
#> 112         RRASTER                                                 R Raster
#> 113             RS2                                   RadarSat 2 XML Product
#> 114             RST                                        Idrisi Raster A.1
#> 115      Rasterlite                                               Rasterlite
#> 116            SAFE                              Sentinel-1 SAR SAFE Product
#> 117            SAGA                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)
#> 118        SAR_CEOS                                           CEOS SAR Image
#> 119            SDTS                                              SDTS Raster
#> 120       SENTINEL2                                               Sentinel 2
#> 121             SGI                                SGI Image File Format 1.0
#> 122          SIGDEM                       Scaled Integer Gridded DEM .sigdem
#> 123          SNODAS                            Snow Data Assimilation System
#> 124             SRP                      Standard Raster Product (ASRP/USRP)
#> 125         SRTMHGT                                      SRTMHGT File Format
#> 126          STACIT                      Spatio-Temporal Asset Catalog Items
#> 127          STACTA               Spatio-Temporal Asset Catalog Tiled Assets
#> 128             TGA                              TGA/TARGA Image File Format
#> 129             TIL                                          EarthWatch .TIL
#> 130             TSX                                       TerraSAR-X Product
#> 131        Terragen                                     Terragen heightfield
#> 132         USGSDEM                       USGS Optional ASCII DEM (and CDED)
#> 133           VICAR                                          MIPL VICAR file
#> 134             VRT                                           Virtual Raster
#> 135             WCS                                 OGC Web Coverage Service
#> 136            WEBP                                                     WEBP
#> 137             WMS                                      OGC Web Map Service
#> 138            WMTS                                 OGC Web Map Tile Service
#> 139             XPM                                        X11 PixMap Format
#> 140             XYZ                                        ASCII Gridded XYZ
#> 141            ZMap                                           ZMap Plus Grid
#> 142            Zarr                                                     Zarr
#> 143          netCDF                               Network Common Data Format
#>     create  copy isRaster
#> 1    FALSE  TRUE     TRUE
#> 2    FALSE FALSE     TRUE
#> 3     TRUE FALSE     TRUE
#> 4    FALSE FALSE     TRUE
#> 5    FALSE  TRUE     TRUE
#> 6    FALSE FALSE     TRUE
#> 7     TRUE  TRUE     TRUE
#> 8    FALSE FALSE     TRUE
#> 9    FALSE  TRUE     TRUE
#> 10    TRUE FALSE     TRUE
#> 11   FALSE FALSE     TRUE
#> 12    TRUE FALSE     TRUE
#> 13    TRUE FALSE     TRUE
#> 14   FALSE FALSE     TRUE
#> 15   FALSE  TRUE     TRUE
#> 16   FALSE FALSE     TRUE
#> 17   FALSE FALSE     TRUE
#> 18   FALSE  TRUE     TRUE
#> 19   FALSE FALSE     TRUE
#> 20   FALSE FALSE     TRUE
#> 21   FALSE FALSE     TRUE
#> 22    TRUE FALSE     TRUE
#> 23   FALSE FALSE     TRUE
#> 24   FALSE FALSE     TRUE
#> 25   FALSE FALSE     TRUE
#> 26   FALSE FALSE     TRUE
#> 27   FALSE FALSE     TRUE
#> 28   FALSE FALSE     TRUE
#> 29   FALSE  TRUE     TRUE
#> 30   FALSE FALSE     TRUE
#> 31   FALSE FALSE     TRUE
#> 32    TRUE  TRUE     TRUE
#> 33   FALSE FALSE     TRUE
#> 34    TRUE FALSE     TRUE
#> 35    TRUE FALSE     TRUE
#> 36    TRUE FALSE     TRUE
#> 37   FALSE FALSE     TRUE
#> 38   FALSE FALSE     TRUE
#> 39   FALSE FALSE     TRUE
#> 40   FALSE  TRUE     TRUE
#> 41   FALSE FALSE     TRUE
#> 42   FALSE  TRUE     TRUE
#> 43    TRUE  TRUE     TRUE
#> 44   FALSE FALSE     TRUE
#> 45   FALSE  TRUE     TRUE
#> 46    TRUE  TRUE     TRUE
#> 47   FALSE  TRUE     TRUE
#> 48    TRUE  TRUE     TRUE
#> 49   FALSE FALSE     TRUE
#> 50    TRUE FALSE     TRUE
#> 51    TRUE  TRUE     TRUE
#> 52   FALSE FALSE     TRUE
#> 53   FALSE FALSE     TRUE
#> 54   FALSE FALSE     TRUE
#> 55    TRUE FALSE     TRUE
#> 56   FALSE FALSE     TRUE
#> 57   FALSE FALSE     TRUE
#> 58   FALSE  TRUE     TRUE
#> 59    TRUE  TRUE     TRUE
#> 60   FALSE FALSE     TRUE
#> 61    TRUE  TRUE     TRUE
#> 62   FALSE FALSE     TRUE
#> 63    TRUE FALSE     TRUE
#> 64   FALSE FALSE     TRUE
#> 65    TRUE FALSE     TRUE
#> 66    TRUE  TRUE     TRUE
#> 67   FALSE FALSE     TRUE
#> 68   FALSE FALSE     TRUE
#> 69   FALSE  TRUE     TRUE
#> 70   FALSE  TRUE     TRUE
#> 71    TRUE  TRUE     TRUE
#> 72   FALSE  TRUE     TRUE
#> 73    TRUE FALSE     TRUE
#> 74   FALSE FALSE     TRUE
#> 75    TRUE FALSE     TRUE
#> 76   FALSE  TRUE     TRUE
#> 77   FALSE FALSE     TRUE
#> 78    TRUE FALSE     TRUE
#> 79   FALSE FALSE     TRUE
#> 80    TRUE  TRUE     TRUE
#> 81    TRUE FALSE     TRUE
#> 82    TRUE  TRUE     TRUE
#> 83    TRUE  TRUE     TRUE
#> 84    TRUE  TRUE     TRUE
#> 85   FALSE FALSE     TRUE
#> 86   FALSE FALSE     TRUE
#> 87   FALSE FALSE     TRUE
#> 88    TRUE  TRUE     TRUE
#> 89    TRUE  TRUE     TRUE
#> 90    TRUE FALSE     TRUE
#> 91   FALSE FALSE     TRUE
#> 92    TRUE  TRUE     TRUE
#> 93   FALSE FALSE     TRUE
#> 94   FALSE FALSE     TRUE
#> 95    TRUE FALSE     TRUE
#> 96    TRUE FALSE     TRUE
#> 97    TRUE  TRUE     TRUE
#> 98    TRUE  TRUE     TRUE
#> 99   FALSE FALSE     TRUE
#> 100   TRUE  TRUE     TRUE
#> 101  FALSE FALSE     TRUE
#> 102  FALSE FALSE     TRUE
#> 103  FALSE  TRUE     TRUE
#> 104   TRUE FALSE     TRUE
#> 105  FALSE FALSE     TRUE
#> 106  FALSE  TRUE     TRUE
#> 107  FALSE  TRUE     TRUE
#> 108  FALSE FALSE     TRUE
#> 109   TRUE FALSE     TRUE
#> 110   TRUE FALSE     TRUE
#> 111  FALSE FALSE     TRUE
#> 112   TRUE  TRUE     TRUE
#> 113  FALSE FALSE     TRUE
#> 114   TRUE  TRUE     TRUE
#> 115  FALSE  TRUE     TRUE
#> 116  FALSE FALSE     TRUE
#> 117   TRUE  TRUE     TRUE
#> 118  FALSE FALSE     TRUE
#> 119  FALSE FALSE     TRUE
#> 120  FALSE FALSE     TRUE
#> 121   TRUE FALSE     TRUE
#> 122  FALSE  TRUE     TRUE
#> 123  FALSE FALSE     TRUE
#> 124  FALSE FALSE     TRUE
#> 125  FALSE  TRUE     TRUE
#> 126  FALSE FALSE     TRUE
#> 127  FALSE FALSE     TRUE
#> 128  FALSE FALSE     TRUE
#> 129  FALSE FALSE     TRUE
#> 130  FALSE FALSE     TRUE
#> 131   TRUE FALSE     TRUE
#> 132  FALSE  TRUE     TRUE
#> 133   TRUE  TRUE     TRUE
#> 134   TRUE  TRUE     TRUE
#> 135  FALSE FALSE     TRUE
#> 136  FALSE  TRUE     TRUE
#> 137  FALSE  TRUE     TRUE
#> 138  FALSE  TRUE     TRUE
#> 139  FALSE  TRUE     TRUE
#> 140  FALSE  TRUE     TRUE
#> 141  FALSE  TRUE     TRUE
#> 142   TRUE FALSE     TRUE
#> 143   TRUE  TRUE     TRUE
logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)
#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x5d1e330>
#> 
getDriverLongName(getDriver(x))
#> [1] "JPEG JFIF"
GDAL.close(x)