Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
getCPLConfigOption(ConfigOption)
setCPLConfigOption(ConfigOption, value)

Arguments

driver
An object inheriting from class 'GDALDriver'

str
A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

ConfigOption
CPL configure option documented in http://trac.osgeo.org/gdal/wiki/ConfigOptions and elsewhere in GDAL source code

value
a string value to set a CPL option; NULL is used to unset the CPL option

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

Note

Loading the rgdal package changes the GDAL\_DATA environmental variable to the GDAL support files bundled with the package.

See also

Examples

gdalDrivers()

#>                    name
#> 1               AAIGrid
#> 2                  ACE2
#> 3                  ADRG
#> 4                   AIG
#> 5                AirSAR
#> 6                   ARG
#> 7                   BAG
#> 8                BIGGIF
#> 9                   BLX
#> 10                  BMP
#> 11                  BSB
#> 12                   BT
#> 13                  BYN
#> 14                  CAD
#> 15                 CALS
#> 16                 CEOS
#> 17                COASP
#> 18                COSAR
#> 19                  CPG
#> 20              CTable2
#> 21                  CTG
#> 22              DERIVED
#> 23                DIMAP
#> 24                DIPEx
#> 25                 DOQ1
#> 26                 DOQ2
#> 27                 DTED
#> 28              E00GRID
#> 29              ECRGTOC
#> 30                EEDAI
#> 31                 EHdr
#> 32                  EIR
#> 33                 ELAS
#> 34                 ENVI
#> 35                  ERS
#> 36                 ESAT
#> 37                 FAST
#> 38                  FIT
#> 39              FujiBAS
#> 40               GenBin
#> 41                  GFF
#> 42                  GIF
#> 43                  GMT
#> 44                 GPKG
#> 45       GRASSASCIIGrid
#> 46                 GRIB
#> 47                GS7BG
#> 48                 GSAG
#> 49                 GSBG
#> 50                  GSC
#> 51                GTiff
#> 52                  GTX
#> 53                  GXF
#> 54                 HDF5
#> 55            HDF5Image
#> 56                  HF2
#> 57                  HFA
#> 58                 HTTP
#> 59                  IDA
#> 60  IGNFHeightASCIIGrid
#> 61                ILWIS
#> 62                 INGR
#> 63                 IRIS
#> 64                 ISCE
#> 65                ISIS2
#> 66                ISIS3
#> 67           JAXAPALSAR
#> 68                 JDEM
#> 69          JP2OpenJPEG
#> 70                 JPEG
#> 71      KMLSUPEROVERLAY
#> 72                  KRO
#> 73                  L1B
#> 74                  LAN
#> 75                  LCP
#> 76             Leveller
#> 77               LOSLAS
#> 78                  MAP
#> 79              MBTiles
#> 80                  MEM
#> 81                  MFF
#> 82                 MFF2
#> 83                  MRF
#> 84                 MSGN
#> 85                  NDF
#> 86               netCDF
#> 87             NGSGEOID
#> 88                  NGW
#> 89                 NITF
#> 90                 NTv1
#> 91                 NTv2
#> 92              NWT_GRC
#> 93              NWT_GRD
#> 94                  OZI
#> 95                 PAux
#> 96               PCIDSK
#> 97             PCRaster
#> 98                  PDF
#> 99                  PDS
#> 100                PDS4
#> 101            PLMOSAIC
#> 102            PLSCENES
#> 103                 PNG
#> 104                 PNM
#> 105       PostGISRaster
#> 106                 PRF
#> 107                   R
#> 108          Rasterlite
#> 109                 RDA
#> 110                 RIK
#> 111                 RMF
#> 112             ROI_PAC
#> 113              RPFTOC
#> 114             RRASTER
#> 115                 RS2
#> 116                 RST
#> 117                SAFE
#> 118                SAGA
#> 119            SAR_CEOS
#> 120                SDTS
#> 121           SENTINEL2
#> 122                 SGI
#> 123              SIGDEM
#> 124              SNODAS
#> 125                 SRP
#> 126             SRTMHGT
#> 127            Terragen
#> 128                 TIL
#> 129                 TSX
#> 130             USGSDEM
#> 131               VICAR
#> 132                 VRT
#> 133                 WCS
#> 134                 WMS
#> 135                WMTS
#> 136                 XPM
#> 137                 XYZ
#> 138                ZMap
#>                                                    long_name create  copy
#> 1                                        Arc/Info ASCII Grid  FALSE  TRUE
#> 2                                                       ACE2  FALSE FALSE
#> 3                              ARC Digitized Raster Graphics   TRUE FALSE
#> 4                                       Arc/Info Binary Grid  FALSE FALSE
#> 5                                  AirSAR Polarimetric Image  FALSE FALSE
#> 6                                  Azavea Raster Grid format  FALSE  TRUE
#> 7                                 Bathymetry Attributed Grid  FALSE  TRUE
#> 8                         Graphics Interchange Format (.gif)  FALSE FALSE
#> 9                                       Magellan topo (.blx)  FALSE  TRUE
#> 10                      MS Windows Device Independent Bitmap   TRUE FALSE
#> 11                               Maptech BSB Nautical Charts  FALSE FALSE
#> 12                       VTP .bt (Binary Terrain) 1.3 Format   TRUE FALSE
#> 13                          Natural Resources Canada's Geoid   TRUE FALSE
#> 14                                            AutoCAD Driver  FALSE FALSE
#> 15                                             CALS (Type 1)  FALSE  TRUE
#> 16                                                CEOS Image  FALSE FALSE
#> 17                           DRDC COASP SAR Processor Raster  FALSE FALSE
#> 18                COSAR Annotated Binary Matrix (TerraSAR-X)  FALSE FALSE
#> 19                                           Convair PolGASP  FALSE FALSE
#> 20                                  CTable2 Datum Grid Shift   TRUE FALSE
#> 21                            USGS LULC Composite Theme Grid  FALSE FALSE
#> 22                Derived datasets using VRT pixel functions  FALSE FALSE
#> 23                                                SPOT DIMAP  FALSE FALSE
#> 24                                                     DIPEx  FALSE FALSE
#> 25                                      USGS DOQ (Old Style)  FALSE FALSE
#> 26                                      USGS DOQ (New Style)  FALSE FALSE
#> 27                                     DTED Elevation Raster  FALSE  TRUE
#> 28                                  Arc/Info Export E00 GRID  FALSE FALSE
#> 29                                           ECRG TOC format  FALSE FALSE
#> 30                               Earth Engine Data API Image  FALSE FALSE
#> 31                                        ESRI .hdr Labelled   TRUE  TRUE
#> 32                                         Erdas Imagine Raw  FALSE FALSE
#> 33                                                      ELAS   TRUE FALSE
#> 34                                        ENVI .hdr Labelled   TRUE FALSE
#> 35                                    ERMapper .ers Labelled   TRUE FALSE
#> 36                                      Envisat Image Format  FALSE FALSE
#> 37                                         EOSAT FAST Format  FALSE FALSE
#> 38                                                 FIT Image  FALSE  TRUE
#> 39                                    Fuji BAS Scanner Image  FALSE FALSE
#> 40                            Generic Binary (.hdr Labelled)  FALSE FALSE
#> 41  Ground-based SAR Applications Testbed File Format (.gff)  FALSE FALSE
#> 42                        Graphics Interchange Format (.gif)  FALSE  TRUE
#> 43                                    GMT NetCDF Grid Format  FALSE  TRUE
#> 44                                                GeoPackage   TRUE  TRUE
#> 45                                          GRASS ASCII Grid  FALSE FALSE
#> 46                              GRIdded Binary (.grb, .grb2)  FALSE  TRUE
#> 47                      Golden Software 7 Binary Grid (.grd)   TRUE  TRUE
#> 48                         Golden Software ASCII Grid (.grd)  FALSE  TRUE
#> 49                        Golden Software Binary Grid (.grd)   TRUE  TRUE
#> 50                                               GSC Geogrid  FALSE FALSE
#> 51                                                   GeoTIFF   TRUE  TRUE
#> 52                                  NOAA Vertical Datum .GTX   TRUE FALSE
#> 53                              GeoSoft Grid Exchange Format  FALSE FALSE
#> 54                        Hierarchical Data Format Release 5  FALSE FALSE
#> 55                                              HDF5 Dataset  FALSE FALSE
#> 56                                HF2/HFZ heightfield raster  FALSE  TRUE
#> 57                               Erdas Imagine Images (.img)   TRUE  TRUE
#> 58                                     HTTP Fetching Wrapper  FALSE FALSE
#> 59                                   Image Data and Analysis   TRUE FALSE
#> 60                   IGN France height correction ASCII Grid  FALSE FALSE
#> 61                                          ILWIS Raster Map   TRUE  TRUE
#> 62                                         Intergraph Raster   TRUE  TRUE
#> 63                              IRIS data (.PPI, .CAPPi etc)  FALSE FALSE
#> 64                                               ISCE raster   TRUE FALSE
#> 65                   USGS Astrogeology ISIS cube (Version 2)   TRUE FALSE
#> 66                   USGS Astrogeology ISIS cube (Version 3)   TRUE  TRUE
#> 67                JAXA PALSAR Product Reader (Level 1.1/1.5)  FALSE FALSE
#> 68                                       Japanese DEM (.mem)  FALSE FALSE
#> 69                JPEG-2000 driver based on OpenJPEG library  FALSE  TRUE
#> 70                                                 JPEG JFIF  FALSE  TRUE
#> 71                                         Kml Super Overlay  FALSE  TRUE
#> 72                                                 KOLOR Raw   TRUE FALSE
#> 73                      NOAA Polar Orbiter Level 1b Data Set  FALSE FALSE
#> 74                                           Erdas .LAN/.GIS   TRUE FALSE
#> 75                         FARSITE v.4 Landscape File (.lcp)  FALSE  TRUE
#> 76                                      Leveller heightfield   TRUE FALSE
#> 77                         NADCON .los/.las Datum Grid Shift  FALSE FALSE
#> 78                                          OziExplorer .MAP  FALSE FALSE
#> 79                                                   MBTiles   TRUE  TRUE
#> 80                                          In Memory Raster   TRUE FALSE
#> 81                                         Vexcel MFF Raster   TRUE  TRUE
#> 82                                  Vexcel MFF2 (HKV) Raster   TRUE  TRUE
#> 83                                        Meta Raster Format   TRUE  TRUE
#> 84                            EUMETSAT Archive native (.nat)  FALSE FALSE
#> 85                                         NLAPS Data Format  FALSE FALSE
#> 86                                Network Common Data Format   TRUE  TRUE
#> 87                               NOAA NGS Geoid Height Grids  FALSE FALSE
#> 88                                               NextGIS Web   TRUE FALSE
#> 89                      National Imagery Transmission Format   TRUE  TRUE
#> 90                                     NTv1 Datum Grid Shift  FALSE FALSE
#> 91                                     NTv2 Datum Grid Shift   TRUE FALSE
#> 92                Northwood Classified Grid Format .grc/.tab  FALSE FALSE
#> 93                   Northwood Numeric Grid Format .grd/.tab   TRUE  TRUE
#> 94                                    OziExplorer Image File  FALSE FALSE
#> 95                                         PCI .aux Labelled   TRUE FALSE
#> 96                                      PCIDSK Database File   TRUE FALSE
#> 97                                      PCRaster Raster File   TRUE  TRUE
#> 98                                            Geospatial PDF   TRUE  TRUE
#> 99                                NASA Planetary Data System  FALSE FALSE
#> 100                             NASA Planetary Data System 4   TRUE  TRUE
#> 101                                  Planet Labs Mosaics API  FALSE FALSE
#> 102                                   Planet Labs Scenes API  FALSE FALSE
#> 103                                Portable Network Graphics  FALSE  TRUE
#> 104                          Portable Pixmap Format (netpbm)   TRUE FALSE
#> 105                                    PostGIS Raster driver  FALSE  TRUE
#> 106                                      Racurs PHOTOMOD PRF  FALSE FALSE
#> 107                                      R Object Data Store  FALSE  TRUE
#> 108                                               Rasterlite  FALSE  TRUE
#> 109                   DigitalGlobe Raster Data Access driver  FALSE FALSE
#> 110                                  Swedish Grid RIK (.rik)  FALSE FALSE
#> 111                                     Raster Matrix Format   TRUE FALSE
#> 112                                           ROI_PAC raster   TRUE FALSE
#> 113                         Raster Product Format TOC format  FALSE FALSE
#> 114                                                 R Raster   TRUE  TRUE
#> 115                                   RadarSat 2 XML Product  FALSE FALSE
#> 116                                        Idrisi Raster A.1   TRUE  TRUE
#> 117                              Sentinel-1 SAR SAFE Product  FALSE FALSE
#> 118                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)   TRUE  TRUE
#> 119                                           CEOS SAR Image  FALSE FALSE
#> 120                                              SDTS Raster  FALSE FALSE
#> 121                                               Sentinel 2  FALSE FALSE
#> 122                                SGI Image File Format 1.0   TRUE FALSE
#> 123                       Scaled Integer Gridded DEM .sigdem  FALSE  TRUE
#> 124                            Snow Data Assimilation System  FALSE FALSE
#> 125                      Standard Raster Product (ASRP/USRP)  FALSE FALSE
#> 126                                      SRTMHGT File Format  FALSE  TRUE
#> 127                                     Terragen heightfield   TRUE FALSE
#> 128                                          EarthWatch .TIL  FALSE FALSE
#> 129                                       TerraSAR-X Product  FALSE FALSE
#> 130                       USGS Optional ASCII DEM (and CDED)  FALSE  TRUE
#> 131                                          MIPL VICAR file  FALSE FALSE
#> 132                                           Virtual Raster   TRUE  TRUE
#> 133                                 OGC Web Coverage Service  FALSE FALSE
#> 134                                      OGC Web Map Service  FALSE  TRUE
#> 135                                 OGC Web Map Tile Service  FALSE  TRUE
#> 136                                        X11 PixMap Format  FALSE  TRUE
#> 137                                        ASCII Gridded XYZ  FALSE  TRUE
#> 138                                           ZMap Plus Grid  FALSE  TRUE
#>     isRaster
#> 1       TRUE
#> 2       TRUE
#> 3       TRUE
#> 4       TRUE
#> 5       TRUE
#> 6       TRUE
#> 7       TRUE
#> 8       TRUE
#> 9       TRUE
#> 10      TRUE
#> 11      TRUE
#> 12      TRUE
#> 13      TRUE
#> 14      TRUE
#> 15      TRUE
#> 16      TRUE
#> 17      TRUE
#> 18      TRUE
#> 19      TRUE
#> 20      TRUE
#> 21      TRUE
#> 22      TRUE
#> 23      TRUE
#> 24      TRUE
#> 25      TRUE
#> 26      TRUE
#> 27      TRUE
#> 28      TRUE
#> 29      TRUE
#> 30      TRUE
#> 31      TRUE
#> 32      TRUE
#> 33      TRUE
#> 34      TRUE
#> 35      TRUE
#> 36      TRUE
#> 37      TRUE
#> 38      TRUE
#> 39      TRUE
#> 40      TRUE
#> 41      TRUE
#> 42      TRUE
#> 43      TRUE
#> 44      TRUE
#> 45      TRUE
#> 46      TRUE
#> 47      TRUE
#> 48      TRUE
#> 49      TRUE
#> 50      TRUE
#> 51      TRUE
#> 52      TRUE
#> 53      TRUE
#> 54      TRUE
#> 55      TRUE
#> 56      TRUE
#> 57      TRUE
#> 58      TRUE
#> 59      TRUE
#> 60      TRUE
#> 61      TRUE
#> 62      TRUE
#> 63      TRUE
#> 64      TRUE
#> 65      TRUE
#> 66      TRUE
#> 67      TRUE
#> 68      TRUE
#> 69      TRUE
#> 70      TRUE
#> 71      TRUE
#> 72      TRUE
#> 73      TRUE
#> 74      TRUE
#> 75      TRUE
#> 76      TRUE
#> 77      TRUE
#> 78      TRUE
#> 79      TRUE
#> 80      TRUE
#> 81      TRUE
#> 82      TRUE
#> 83      TRUE
#> 84      TRUE
#> 85      TRUE
#> 86      TRUE
#> 87      TRUE
#> 88      TRUE
#> 89      TRUE
#> 90      TRUE
#> 91      TRUE
#> 92      TRUE
#> 93      TRUE
#> 94      TRUE
#> 95      TRUE
#> 96      TRUE
#> 97      TRUE
#> 98      TRUE
#> 99      TRUE
#> 100     TRUE
#> 101     TRUE
#> 102     TRUE
#> 103     TRUE
#> 104     TRUE
#> 105     TRUE
#> 106     TRUE
#> 107     TRUE
#> 108     TRUE
#> 109     TRUE
#> 110     TRUE
#> 111     TRUE
#> 112     TRUE
#> 113     TRUE
#> 114     TRUE
#> 115     TRUE
#> 116     TRUE
#> 117     TRUE
#> 118     TRUE
#> 119     TRUE
#> 120     TRUE
#> 121     TRUE
#> 122     TRUE
#> 123     TRUE
#> 124     TRUE
#> 125     TRUE
#> 126     TRUE
#> 127     TRUE
#> 128     TRUE
#> 129     TRUE
#> 130     TRUE
#> 131     TRUE
#> 132     TRUE
#> 133     TRUE
#> 134     TRUE
#> 135     TRUE
#> 136     TRUE
#> 137     TRUE
#> 138     TRUE

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)

#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x62a41e0>
#> 

getDriverLongName(getDriver(x))

#> [1] "JPEG JFIF"

GDAL.close(x)