Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
getCPLConfigOption(ConfigOption)
setCPLConfigOption(ConfigOption, value)
get_cached_orig_PROJ_LIB()
get_cached_orig_GDAL_DATA()
get_cached_set_PROJ_LIB()
get_cached_set_GDAL_DATA()

Arguments

driver
An object inheriting from class 'GDALDriver'

str
A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

ConfigOption
CPL configure option documented in http://trac.osgeo.org/gdal/wiki/ConfigOptions and elsewhere in GDAL source code

value
a string value to set a CPL option; NULL is used to unset the CPL option

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

get_cached_orig_PROJ_LIB, get_cached_orig_GDAL_DATA

The values of environment variables PROJ_LIB and GDAL_DATA as read when this package was loaded

get_cached_set_PROJ_LIB, get_cached_set_GDAL_DATA

If not "", the values set when loading this package to point to metadata files included in CRAN binary packages

Note

Loading the rgdal package changes the GDAL\_DATA environmental variable to the GDAL support files bundled with the package.

See also

Examples

gdalDrivers()

#>                    name
#> 1               AAIGrid
#> 2                  ACE2
#> 3                  ADRG
#> 4                   AIG
#> 5                AirSAR
#> 6                   ARG
#> 7                   BAG
#> 8                BIGGIF
#> 9                   BLX
#> 10                  BMP
#> 11                  BSB
#> 12                   BT
#> 13                  BYN
#> 14                  CAD
#> 15                 CALS
#> 16                 CEOS
#> 17                COASP
#> 18                  COG
#> 19                COSAR
#> 20                  CPG
#> 21              CTable2
#> 22                  CTG
#> 23                 DAAS
#> 24              DERIVED
#> 25                DIMAP
#> 26                DIPEx
#> 27                 DOQ1
#> 28                 DOQ2
#> 29                 DTED
#> 30              E00GRID
#> 31              ECRGTOC
#> 32                EEDAI
#> 33                 EHdr
#> 34                  EIR
#> 35                 ELAS
#> 36                 ENVI
#> 37                  ERS
#> 38                 ESAT
#> 39                  EXR
#> 40                 FAST
#> 41                  FIT
#> 42              FujiBAS
#> 43               GenBin
#> 44                  GFF
#> 45                  GIF
#> 46                  GMT
#> 47                 GPKG
#> 48       GRASSASCIIGrid
#> 49                 GRIB
#> 50                GS7BG
#> 51                 GSAG
#> 52                 GSBG
#> 53                  GSC
#> 54                GTiff
#> 55                  GTX
#> 56                  GXF
#> 57                 HDF5
#> 58            HDF5Image
#> 59                  HF2
#> 60                  HFA
#> 61                 HTTP
#> 62                  IDA
#> 63  IGNFHeightASCIIGrid
#> 64                ILWIS
#> 65                 INGR
#> 66                 IRIS
#> 67                 ISCE
#> 68                  ISG
#> 69                ISIS2
#> 70                ISIS3
#> 71           JAXAPALSAR
#> 72                 JDEM
#> 73          JP2OpenJPEG
#> 74                 JPEG
#> 75      KMLSUPEROVERLAY
#> 76                  KRO
#> 77                  L1B
#> 78                  LAN
#> 79                  LCP
#> 80             Leveller
#> 81               LOSLAS
#> 82                  MAP
#> 83              MBTiles
#> 84                  MEM
#> 85                  MFF
#> 86                 MFF2
#> 87                  MRF
#> 88                 MSGN
#> 89                  NDF
#> 90               netCDF
#> 91             NGSGEOID
#> 92                  NGW
#> 93                 NITF
#> 94                 NTv1
#> 95                 NTv2
#> 96              NWT_GRC
#> 97              NWT_GRD
#> 98                  OZI
#> 99                 PAux
#> 100              PCIDSK
#> 101            PCRaster
#> 102                 PDF
#> 103                 PDS
#> 104                PDS4
#> 105            PLMOSAIC
#> 106            PLSCENES
#> 107                 PNG
#> 108                 PNM
#> 109       PostGISRaster
#> 110                 PRF
#> 111                   R
#> 112          Rasterlite
#> 113                 RDA
#> 114                 RIK
#> 115                 RMF
#> 116             ROI_PAC
#> 117              RPFTOC
#> 118             RRASTER
#> 119                 RS2
#> 120                 RST
#> 121                SAFE
#> 122                SAGA
#> 123            SAR_CEOS
#> 124                SDTS
#> 125           SENTINEL2
#> 126                 SGI
#> 127              SIGDEM
#> 128              SNODAS
#> 129                 SRP
#> 130             SRTMHGT
#> 131            Terragen
#> 132                 TIL
#> 133                 TSX
#> 134             USGSDEM
#> 135               VICAR
#> 136                 VRT
#> 137                 WCS
#> 138                WEBP
#> 139                 WMS
#> 140                WMTS
#> 141                 XPM
#> 142                 XYZ
#> 143                ZMap
#>                                                    long_name create  copy
#> 1                                        Arc/Info ASCII Grid  FALSE  TRUE
#> 2                                                       ACE2  FALSE FALSE
#> 3                              ARC Digitized Raster Graphics   TRUE FALSE
#> 4                                       Arc/Info Binary Grid  FALSE FALSE
#> 5                                  AirSAR Polarimetric Image  FALSE FALSE
#> 6                                  Azavea Raster Grid format  FALSE  TRUE
#> 7                                 Bathymetry Attributed Grid  FALSE  TRUE
#> 8                         Graphics Interchange Format (.gif)  FALSE FALSE
#> 9                                       Magellan topo (.blx)  FALSE  TRUE
#> 10                      MS Windows Device Independent Bitmap   TRUE FALSE
#> 11                               Maptech BSB Nautical Charts  FALSE FALSE
#> 12                       VTP .bt (Binary Terrain) 1.3 Format   TRUE FALSE
#> 13                          Natural Resources Canada's Geoid   TRUE FALSE
#> 14                                            AutoCAD Driver  FALSE FALSE
#> 15                                             CALS (Type 1)  FALSE  TRUE
#> 16                                                CEOS Image  FALSE FALSE
#> 17                           DRDC COASP SAR Processor Raster  FALSE FALSE
#> 18                         Cloud optimized GeoTIFF generator  FALSE  TRUE
#> 19                COSAR Annotated Binary Matrix (TerraSAR-X)  FALSE FALSE
#> 20                                           Convair PolGASP  FALSE FALSE
#> 21                                  CTable2 Datum Grid Shift   TRUE FALSE
#> 22                            USGS LULC Composite Theme Grid  FALSE FALSE
#> 23           Airbus DS Intelligence Data As A Service driver  FALSE FALSE
#> 24                Derived datasets using VRT pixel functions  FALSE FALSE
#> 25                                                SPOT DIMAP  FALSE FALSE
#> 26                                                     DIPEx  FALSE FALSE
#> 27                                      USGS DOQ (Old Style)  FALSE FALSE
#> 28                                      USGS DOQ (New Style)  FALSE FALSE
#> 29                                     DTED Elevation Raster  FALSE  TRUE
#> 30                                  Arc/Info Export E00 GRID  FALSE FALSE
#> 31                                           ECRG TOC format  FALSE FALSE
#> 32                               Earth Engine Data API Image  FALSE FALSE
#> 33                                        ESRI .hdr Labelled   TRUE  TRUE
#> 34                                         Erdas Imagine Raw  FALSE FALSE
#> 35                                                      ELAS   TRUE FALSE
#> 36                                        ENVI .hdr Labelled   TRUE FALSE
#> 37                                    ERMapper .ers Labelled   TRUE FALSE
#> 38                                      Envisat Image Format  FALSE FALSE
#> 39                  Extended Dynamic Range Image File Format   TRUE  TRUE
#> 40                                         EOSAT FAST Format  FALSE FALSE
#> 41                                                 FIT Image  FALSE  TRUE
#> 42                                    Fuji BAS Scanner Image  FALSE FALSE
#> 43                            Generic Binary (.hdr Labelled)  FALSE FALSE
#> 44  Ground-based SAR Applications Testbed File Format (.gff)  FALSE FALSE
#> 45                        Graphics Interchange Format (.gif)  FALSE  TRUE
#> 46                                    GMT NetCDF Grid Format  FALSE  TRUE
#> 47                                                GeoPackage   TRUE  TRUE
#> 48                                          GRASS ASCII Grid  FALSE FALSE
#> 49                              GRIdded Binary (.grb, .grb2)  FALSE  TRUE
#> 50                      Golden Software 7 Binary Grid (.grd)   TRUE  TRUE
#> 51                         Golden Software ASCII Grid (.grd)  FALSE  TRUE
#> 52                        Golden Software Binary Grid (.grd)   TRUE  TRUE
#> 53                                               GSC Geogrid  FALSE FALSE
#> 54                                                   GeoTIFF   TRUE  TRUE
#> 55                                  NOAA Vertical Datum .GTX   TRUE FALSE
#> 56                              GeoSoft Grid Exchange Format  FALSE FALSE
#> 57                        Hierarchical Data Format Release 5  FALSE FALSE
#> 58                                              HDF5 Dataset  FALSE FALSE
#> 59                                HF2/HFZ heightfield raster  FALSE  TRUE
#> 60                               Erdas Imagine Images (.img)   TRUE  TRUE
#> 61                                     HTTP Fetching Wrapper  FALSE FALSE
#> 62                                   Image Data and Analysis   TRUE FALSE
#> 63                   IGN France height correction ASCII Grid  FALSE FALSE
#> 64                                          ILWIS Raster Map   TRUE  TRUE
#> 65                                         Intergraph Raster   TRUE  TRUE
#> 66                              IRIS data (.PPI, .CAPPi etc)  FALSE FALSE
#> 67                                               ISCE raster   TRUE FALSE
#> 68                       International Service for the Geoid  FALSE FALSE
#> 69                   USGS Astrogeology ISIS cube (Version 2)   TRUE FALSE
#> 70                   USGS Astrogeology ISIS cube (Version 3)   TRUE  TRUE
#> 71                JAXA PALSAR Product Reader (Level 1.1/1.5)  FALSE FALSE
#> 72                                       Japanese DEM (.mem)  FALSE FALSE
#> 73                JPEG-2000 driver based on OpenJPEG library  FALSE  TRUE
#> 74                                                 JPEG JFIF  FALSE  TRUE
#> 75                                         Kml Super Overlay  FALSE  TRUE
#> 76                                                 KOLOR Raw   TRUE FALSE
#> 77                      NOAA Polar Orbiter Level 1b Data Set  FALSE FALSE
#> 78                                           Erdas .LAN/.GIS   TRUE FALSE
#> 79                         FARSITE v.4 Landscape File (.lcp)  FALSE  TRUE
#> 80                                      Leveller heightfield   TRUE FALSE
#> 81                         NADCON .los/.las Datum Grid Shift  FALSE FALSE
#> 82                                          OziExplorer .MAP  FALSE FALSE
#> 83                                                   MBTiles   TRUE  TRUE
#> 84                                          In Memory Raster   TRUE FALSE
#> 85                                         Vexcel MFF Raster   TRUE  TRUE
#> 86                                  Vexcel MFF2 (HKV) Raster   TRUE  TRUE
#> 87                                        Meta Raster Format   TRUE  TRUE
#> 88                            EUMETSAT Archive native (.nat)  FALSE FALSE
#> 89                                         NLAPS Data Format  FALSE FALSE
#> 90                                Network Common Data Format   TRUE  TRUE
#> 91                               NOAA NGS Geoid Height Grids  FALSE FALSE
#> 92                                               NextGIS Web   TRUE  TRUE
#> 93                      National Imagery Transmission Format   TRUE  TRUE
#> 94                                     NTv1 Datum Grid Shift  FALSE FALSE
#> 95                                     NTv2 Datum Grid Shift   TRUE FALSE
#> 96                Northwood Classified Grid Format .grc/.tab  FALSE FALSE
#> 97                   Northwood Numeric Grid Format .grd/.tab   TRUE  TRUE
#> 98                                    OziExplorer Image File  FALSE FALSE
#> 99                                         PCI .aux Labelled   TRUE FALSE
#> 100                                     PCIDSK Database File   TRUE FALSE
#> 101                                     PCRaster Raster File   TRUE  TRUE
#> 102                                           Geospatial PDF   TRUE  TRUE
#> 103                               NASA Planetary Data System  FALSE FALSE
#> 104                             NASA Planetary Data System 4   TRUE  TRUE
#> 105                                  Planet Labs Mosaics API  FALSE FALSE
#> 106                                   Planet Labs Scenes API  FALSE FALSE
#> 107                                Portable Network Graphics  FALSE  TRUE
#> 108                          Portable Pixmap Format (netpbm)   TRUE FALSE
#> 109                                    PostGIS Raster driver  FALSE  TRUE
#> 110                                      Racurs PHOTOMOD PRF  FALSE FALSE
#> 111                                      R Object Data Store  FALSE  TRUE
#> 112                                               Rasterlite  FALSE  TRUE
#> 113                   DigitalGlobe Raster Data Access driver  FALSE FALSE
#> 114                                  Swedish Grid RIK (.rik)  FALSE FALSE
#> 115                                     Raster Matrix Format   TRUE FALSE
#> 116                                           ROI_PAC raster   TRUE FALSE
#> 117                         Raster Product Format TOC format  FALSE FALSE
#> 118                                                 R Raster   TRUE  TRUE
#> 119                                   RadarSat 2 XML Product  FALSE FALSE
#> 120                                        Idrisi Raster A.1   TRUE  TRUE
#> 121                              Sentinel-1 SAR SAFE Product  FALSE FALSE
#> 122                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)   TRUE  TRUE
#> 123                                           CEOS SAR Image  FALSE FALSE
#> 124                                              SDTS Raster  FALSE FALSE
#> 125                                               Sentinel 2  FALSE FALSE
#> 126                                SGI Image File Format 1.0   TRUE FALSE
#> 127                       Scaled Integer Gridded DEM .sigdem  FALSE  TRUE
#> 128                            Snow Data Assimilation System  FALSE FALSE
#> 129                      Standard Raster Product (ASRP/USRP)  FALSE FALSE
#> 130                                      SRTMHGT File Format  FALSE  TRUE
#> 131                                     Terragen heightfield   TRUE FALSE
#> 132                                          EarthWatch .TIL  FALSE FALSE
#> 133                                       TerraSAR-X Product  FALSE FALSE
#> 134                       USGS Optional ASCII DEM (and CDED)  FALSE  TRUE
#> 135                                          MIPL VICAR file   TRUE  TRUE
#> 136                                           Virtual Raster   TRUE  TRUE
#> 137                                 OGC Web Coverage Service  FALSE FALSE
#> 138                                                     WEBP  FALSE  TRUE
#> 139                                      OGC Web Map Service  FALSE  TRUE
#> 140                                 OGC Web Map Tile Service  FALSE  TRUE
#> 141                                        X11 PixMap Format  FALSE  TRUE
#> 142                                        ASCII Gridded XYZ  FALSE  TRUE
#> 143                                           ZMap Plus Grid  FALSE  TRUE
#>     isRaster
#> 1       TRUE
#> 2       TRUE
#> 3       TRUE
#> 4       TRUE
#> 5       TRUE
#> 6       TRUE
#> 7       TRUE
#> 8       TRUE
#> 9       TRUE
#> 10      TRUE
#> 11      TRUE
#> 12      TRUE
#> 13      TRUE
#> 14      TRUE
#> 15      TRUE
#> 16      TRUE
#> 17      TRUE
#> 18      TRUE
#> 19      TRUE
#> 20      TRUE
#> 21      TRUE
#> 22      TRUE
#> 23      TRUE
#> 24      TRUE
#> 25      TRUE
#> 26      TRUE
#> 27      TRUE
#> 28      TRUE
#> 29      TRUE
#> 30      TRUE
#> 31      TRUE
#> 32      TRUE
#> 33      TRUE
#> 34      TRUE
#> 35      TRUE
#> 36      TRUE
#> 37      TRUE
#> 38      TRUE
#> 39      TRUE
#> 40      TRUE
#> 41      TRUE
#> 42      TRUE
#> 43      TRUE
#> 44      TRUE
#> 45      TRUE
#> 46      TRUE
#> 47      TRUE
#> 48      TRUE
#> 49      TRUE
#> 50      TRUE
#> 51      TRUE
#> 52      TRUE
#> 53      TRUE
#> 54      TRUE
#> 55      TRUE
#> 56      TRUE
#> 57      TRUE
#> 58      TRUE
#> 59      TRUE
#> 60      TRUE
#> 61      TRUE
#> 62      TRUE
#> 63      TRUE
#> 64      TRUE
#> 65      TRUE
#> 66      TRUE
#> 67      TRUE
#> 68      TRUE
#> 69      TRUE
#> 70      TRUE
#> 71      TRUE
#> 72      TRUE
#> 73      TRUE
#> 74      TRUE
#> 75      TRUE
#> 76      TRUE
#> 77      TRUE
#> 78      TRUE
#> 79      TRUE
#> 80      TRUE
#> 81      TRUE
#> 82      TRUE
#> 83      TRUE
#> 84      TRUE
#> 85      TRUE
#> 86      TRUE
#> 87      TRUE
#> 88      TRUE
#> 89      TRUE
#> 90      TRUE
#> 91      TRUE
#> 92      TRUE
#> 93      TRUE
#> 94      TRUE
#> 95      TRUE
#> 96      TRUE
#> 97      TRUE
#> 98      TRUE
#> 99      TRUE
#> 100     TRUE
#> 101     TRUE
#> 102     TRUE
#> 103     TRUE
#> 104     TRUE
#> 105     TRUE
#> 106     TRUE
#> 107     TRUE
#> 108     TRUE
#> 109     TRUE
#> 110     TRUE
#> 111     TRUE
#> 112     TRUE
#> 113     TRUE
#> 114     TRUE
#> 115     TRUE
#> 116     TRUE
#> 117     TRUE
#> 118     TRUE
#> 119     TRUE
#> 120     TRUE
#> 121     TRUE
#> 122     TRUE
#> 123     TRUE
#> 124     TRUE
#> 125     TRUE
#> 126     TRUE
#> 127     TRUE
#> 128     TRUE
#> 129     TRUE
#> 130     TRUE
#> 131     TRUE
#> 132     TRUE
#> 133     TRUE
#> 134     TRUE
#> 135     TRUE
#> 136     TRUE
#> 137     TRUE
#> 138     TRUE
#> 139     TRUE
#> 140     TRUE
#> 141     TRUE
#> 142     TRUE
#> 143     TRUE

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)

#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x8605340>
#> 

getDriverLongName(getDriver(x))

#> [1] "JPEG JFIF"

GDAL.close(x)