Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
getCPLConfigOption(ConfigOption)
setCPLConfigOption(ConfigOption, value)

Arguments

driver
An object inheriting from class 'GDALDriver'

str
A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

ConfigOption
CPL configure option documented in http://trac.osgeo.org/gdal/wiki/ConfigOptions and elsewhere in GDAL source code

value
a string value to set a CPL option; NULL is used to unset the CPL option

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

Note

Loading the rgdal package changes the GDAL\_DATA environmental variable to the GDAL support files bundled with the package.

See also

Examples

gdalDrivers()

#>                    name
#> 1               AAIGrid
#> 2                  ACE2
#> 3                  ADRG
#> 4                   AIG
#> 5                AirSAR
#> 6                   ARG
#> 7                   BAG
#> 8                BIGGIF
#> 9                   BLX
#> 10                  BMP
#> 11                  BSB
#> 12                   BT
#> 13                  BYN
#> 14                  CAD
#> 15                 CALS
#> 16                 CEOS
#> 17                COASP
#> 18                  COG
#> 19                COSAR
#> 20                  CPG
#> 21              CTable2
#> 22                  CTG
#> 23                 DAAS
#> 24              DERIVED
#> 25                DIMAP
#> 26                DIPEx
#> 27                 DOQ1
#> 28                 DOQ2
#> 29                 DTED
#> 30              E00GRID
#> 31              ECRGTOC
#> 32                EEDAI
#> 33                 EHdr
#> 34                  EIR
#> 35                 ELAS
#> 36                 ENVI
#> 37                  ERS
#> 38                 ESAT
#> 39                 FAST
#> 40                  FIT
#> 41              FujiBAS
#> 42               GenBin
#> 43                  GFF
#> 44                  GIF
#> 45                  GMT
#> 46                 GPKG
#> 47       GRASSASCIIGrid
#> 48                 GRIB
#> 49                GS7BG
#> 50                 GSAG
#> 51                 GSBG
#> 52                  GSC
#> 53                GTiff
#> 54                  GTX
#> 55                  GXF
#> 56                 HDF5
#> 57            HDF5Image
#> 58                  HF2
#> 59                  HFA
#> 60                 HTTP
#> 61                  IDA
#> 62  IGNFHeightASCIIGrid
#> 63                ILWIS
#> 64                 INGR
#> 65                 IRIS
#> 66                 ISCE
#> 67                ISIS2
#> 68                ISIS3
#> 69           JAXAPALSAR
#> 70                 JDEM
#> 71          JP2OpenJPEG
#> 72                 JPEG
#> 73      KMLSUPEROVERLAY
#> 74                  KRO
#> 75                  L1B
#> 76                  LAN
#> 77                  LCP
#> 78             Leveller
#> 79               LOSLAS
#> 80                  MAP
#> 81              MBTiles
#> 82                  MEM
#> 83                  MFF
#> 84                 MFF2
#> 85                  MRF
#> 86                 MSGN
#> 87                  NDF
#> 88               netCDF
#> 89             NGSGEOID
#> 90                  NGW
#> 91                 NITF
#> 92                 NTv1
#> 93                 NTv2
#> 94              NWT_GRC
#> 95              NWT_GRD
#> 96                  OZI
#> 97                 PAux
#> 98               PCIDSK
#> 99             PCRaster
#> 100                 PDF
#> 101                 PDS
#> 102                PDS4
#> 103            PLMOSAIC
#> 104            PLSCENES
#> 105                 PNG
#> 106                 PNM
#> 107       PostGISRaster
#> 108                 PRF
#> 109                   R
#> 110          Rasterlite
#> 111                 RDA
#> 112                 RIK
#> 113                 RMF
#> 114             ROI_PAC
#> 115              RPFTOC
#> 116             RRASTER
#> 117                 RS2
#> 118                 RST
#> 119                SAFE
#> 120                SAGA
#> 121            SAR_CEOS
#> 122                SDTS
#> 123           SENTINEL2
#> 124                 SGI
#> 125              SIGDEM
#> 126              SNODAS
#> 127                 SRP
#> 128             SRTMHGT
#> 129            Terragen
#> 130                 TIL
#> 131                 TSX
#> 132             USGSDEM
#> 133               VICAR
#> 134                 VRT
#> 135                 WCS
#> 136                WEBP
#> 137                 WMS
#> 138                WMTS
#> 139                 XPM
#> 140                 XYZ
#> 141                ZMap
#>                                                    long_name create  copy
#> 1                                        Arc/Info ASCII Grid  FALSE  TRUE
#> 2                                                       ACE2  FALSE FALSE
#> 3                              ARC Digitized Raster Graphics   TRUE FALSE
#> 4                                       Arc/Info Binary Grid  FALSE FALSE
#> 5                                  AirSAR Polarimetric Image  FALSE FALSE
#> 6                                  Azavea Raster Grid format  FALSE  TRUE
#> 7                                 Bathymetry Attributed Grid  FALSE  TRUE
#> 8                         Graphics Interchange Format (.gif)  FALSE FALSE
#> 9                                       Magellan topo (.blx)  FALSE  TRUE
#> 10                      MS Windows Device Independent Bitmap   TRUE FALSE
#> 11                               Maptech BSB Nautical Charts  FALSE FALSE
#> 12                       VTP .bt (Binary Terrain) 1.3 Format   TRUE FALSE
#> 13                          Natural Resources Canada's Geoid   TRUE FALSE
#> 14                                            AutoCAD Driver  FALSE FALSE
#> 15                                             CALS (Type 1)  FALSE  TRUE
#> 16                                                CEOS Image  FALSE FALSE
#> 17                           DRDC COASP SAR Processor Raster  FALSE FALSE
#> 18                         Cloud optimized GeoTIFF generator  FALSE  TRUE
#> 19                COSAR Annotated Binary Matrix (TerraSAR-X)  FALSE FALSE
#> 20                                           Convair PolGASP  FALSE FALSE
#> 21                                  CTable2 Datum Grid Shift   TRUE FALSE
#> 22                            USGS LULC Composite Theme Grid  FALSE FALSE
#> 23           Airbus DS Intelligence Data As A Service driver  FALSE FALSE
#> 24                Derived datasets using VRT pixel functions  FALSE FALSE
#> 25                                                SPOT DIMAP  FALSE FALSE
#> 26                                                     DIPEx  FALSE FALSE
#> 27                                      USGS DOQ (Old Style)  FALSE FALSE
#> 28                                      USGS DOQ (New Style)  FALSE FALSE
#> 29                                     DTED Elevation Raster  FALSE  TRUE
#> 30                                  Arc/Info Export E00 GRID  FALSE FALSE
#> 31                                           ECRG TOC format  FALSE FALSE
#> 32                               Earth Engine Data API Image  FALSE FALSE
#> 33                                        ESRI .hdr Labelled   TRUE  TRUE
#> 34                                         Erdas Imagine Raw  FALSE FALSE
#> 35                                                      ELAS   TRUE FALSE
#> 36                                        ENVI .hdr Labelled   TRUE FALSE
#> 37                                    ERMapper .ers Labelled   TRUE FALSE
#> 38                                      Envisat Image Format  FALSE FALSE
#> 39                                         EOSAT FAST Format  FALSE FALSE
#> 40                                                 FIT Image  FALSE  TRUE
#> 41                                    Fuji BAS Scanner Image  FALSE FALSE
#> 42                            Generic Binary (.hdr Labelled)  FALSE FALSE
#> 43  Ground-based SAR Applications Testbed File Format (.gff)  FALSE FALSE
#> 44                        Graphics Interchange Format (.gif)  FALSE  TRUE
#> 45                                    GMT NetCDF Grid Format  FALSE  TRUE
#> 46                                                GeoPackage   TRUE  TRUE
#> 47                                          GRASS ASCII Grid  FALSE FALSE
#> 48                              GRIdded Binary (.grb, .grb2)  FALSE  TRUE
#> 49                      Golden Software 7 Binary Grid (.grd)   TRUE  TRUE
#> 50                         Golden Software ASCII Grid (.grd)  FALSE  TRUE
#> 51                        Golden Software Binary Grid (.grd)   TRUE  TRUE
#> 52                                               GSC Geogrid  FALSE FALSE
#> 53                                                   GeoTIFF   TRUE  TRUE
#> 54                                  NOAA Vertical Datum .GTX   TRUE FALSE
#> 55                              GeoSoft Grid Exchange Format  FALSE FALSE
#> 56                        Hierarchical Data Format Release 5  FALSE FALSE
#> 57                                              HDF5 Dataset  FALSE FALSE
#> 58                                HF2/HFZ heightfield raster  FALSE  TRUE
#> 59                               Erdas Imagine Images (.img)   TRUE  TRUE
#> 60                                     HTTP Fetching Wrapper  FALSE FALSE
#> 61                                   Image Data and Analysis   TRUE FALSE
#> 62                   IGN France height correction ASCII Grid  FALSE FALSE
#> 63                                          ILWIS Raster Map   TRUE  TRUE
#> 64                                         Intergraph Raster   TRUE  TRUE
#> 65                              IRIS data (.PPI, .CAPPi etc)  FALSE FALSE
#> 66                                               ISCE raster   TRUE FALSE
#> 67                   USGS Astrogeology ISIS cube (Version 2)   TRUE FALSE
#> 68                   USGS Astrogeology ISIS cube (Version 3)   TRUE  TRUE
#> 69                JAXA PALSAR Product Reader (Level 1.1/1.5)  FALSE FALSE
#> 70                                       Japanese DEM (.mem)  FALSE FALSE
#> 71                JPEG-2000 driver based on OpenJPEG library  FALSE  TRUE
#> 72                                                 JPEG JFIF  FALSE  TRUE
#> 73                                         Kml Super Overlay  FALSE  TRUE
#> 74                                                 KOLOR Raw   TRUE FALSE
#> 75                      NOAA Polar Orbiter Level 1b Data Set  FALSE FALSE
#> 76                                           Erdas .LAN/.GIS   TRUE FALSE
#> 77                         FARSITE v.4 Landscape File (.lcp)  FALSE  TRUE
#> 78                                      Leveller heightfield   TRUE FALSE
#> 79                         NADCON .los/.las Datum Grid Shift  FALSE FALSE
#> 80                                          OziExplorer .MAP  FALSE FALSE
#> 81                                                   MBTiles   TRUE  TRUE
#> 82                                          In Memory Raster   TRUE FALSE
#> 83                                         Vexcel MFF Raster   TRUE  TRUE
#> 84                                  Vexcel MFF2 (HKV) Raster   TRUE  TRUE
#> 85                                        Meta Raster Format   TRUE  TRUE
#> 86                            EUMETSAT Archive native (.nat)  FALSE FALSE
#> 87                                         NLAPS Data Format  FALSE FALSE
#> 88                                Network Common Data Format   TRUE  TRUE
#> 89                               NOAA NGS Geoid Height Grids  FALSE FALSE
#> 90                                               NextGIS Web   TRUE  TRUE
#> 91                      National Imagery Transmission Format   TRUE  TRUE
#> 92                                     NTv1 Datum Grid Shift  FALSE FALSE
#> 93                                     NTv2 Datum Grid Shift   TRUE FALSE
#> 94                Northwood Classified Grid Format .grc/.tab  FALSE FALSE
#> 95                   Northwood Numeric Grid Format .grd/.tab   TRUE  TRUE
#> 96                                    OziExplorer Image File  FALSE FALSE
#> 97                                         PCI .aux Labelled   TRUE FALSE
#> 98                                      PCIDSK Database File   TRUE FALSE
#> 99                                      PCRaster Raster File   TRUE  TRUE
#> 100                                           Geospatial PDF   TRUE  TRUE
#> 101                               NASA Planetary Data System  FALSE FALSE
#> 102                             NASA Planetary Data System 4   TRUE  TRUE
#> 103                                  Planet Labs Mosaics API  FALSE FALSE
#> 104                                   Planet Labs Scenes API  FALSE FALSE
#> 105                                Portable Network Graphics  FALSE  TRUE
#> 106                          Portable Pixmap Format (netpbm)   TRUE FALSE
#> 107                                    PostGIS Raster driver  FALSE  TRUE
#> 108                                      Racurs PHOTOMOD PRF  FALSE FALSE
#> 109                                      R Object Data Store  FALSE  TRUE
#> 110                                               Rasterlite  FALSE  TRUE
#> 111                   DigitalGlobe Raster Data Access driver  FALSE FALSE
#> 112                                  Swedish Grid RIK (.rik)  FALSE FALSE
#> 113                                     Raster Matrix Format   TRUE FALSE
#> 114                                           ROI_PAC raster   TRUE FALSE
#> 115                         Raster Product Format TOC format  FALSE FALSE
#> 116                                                 R Raster   TRUE  TRUE
#> 117                                   RadarSat 2 XML Product  FALSE FALSE
#> 118                                        Idrisi Raster A.1   TRUE  TRUE
#> 119                              Sentinel-1 SAR SAFE Product  FALSE FALSE
#> 120                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)   TRUE  TRUE
#> 121                                           CEOS SAR Image  FALSE FALSE
#> 122                                              SDTS Raster  FALSE FALSE
#> 123                                               Sentinel 2  FALSE FALSE
#> 124                                SGI Image File Format 1.0   TRUE FALSE
#> 125                       Scaled Integer Gridded DEM .sigdem  FALSE  TRUE
#> 126                            Snow Data Assimilation System  FALSE FALSE
#> 127                      Standard Raster Product (ASRP/USRP)  FALSE FALSE
#> 128                                      SRTMHGT File Format  FALSE  TRUE
#> 129                                     Terragen heightfield   TRUE FALSE
#> 130                                          EarthWatch .TIL  FALSE FALSE
#> 131                                       TerraSAR-X Product  FALSE FALSE
#> 132                       USGS Optional ASCII DEM (and CDED)  FALSE  TRUE
#> 133                                          MIPL VICAR file   TRUE  TRUE
#> 134                                           Virtual Raster   TRUE  TRUE
#> 135                                 OGC Web Coverage Service  FALSE FALSE
#> 136                                                     WEBP  FALSE  TRUE
#> 137                                      OGC Web Map Service  FALSE  TRUE
#> 138                                 OGC Web Map Tile Service  FALSE  TRUE
#> 139                                        X11 PixMap Format  FALSE  TRUE
#> 140                                        ASCII Gridded XYZ  FALSE  TRUE
#> 141                                           ZMap Plus Grid  FALSE  TRUE
#>     isRaster
#> 1       TRUE
#> 2       TRUE
#> 3       TRUE
#> 4       TRUE
#> 5       TRUE
#> 6       TRUE
#> 7       TRUE
#> 8       TRUE
#> 9       TRUE
#> 10      TRUE
#> 11      TRUE
#> 12      TRUE
#> 13      TRUE
#> 14      TRUE
#> 15      TRUE
#> 16      TRUE
#> 17      TRUE
#> 18      TRUE
#> 19      TRUE
#> 20      TRUE
#> 21      TRUE
#> 22      TRUE
#> 23      TRUE
#> 24      TRUE
#> 25      TRUE
#> 26      TRUE
#> 27      TRUE
#> 28      TRUE
#> 29      TRUE
#> 30      TRUE
#> 31      TRUE
#> 32      TRUE
#> 33      TRUE
#> 34      TRUE
#> 35      TRUE
#> 36      TRUE
#> 37      TRUE
#> 38      TRUE
#> 39      TRUE
#> 40      TRUE
#> 41      TRUE
#> 42      TRUE
#> 43      TRUE
#> 44      TRUE
#> 45      TRUE
#> 46      TRUE
#> 47      TRUE
#> 48      TRUE
#> 49      TRUE
#> 50      TRUE
#> 51      TRUE
#> 52      TRUE
#> 53      TRUE
#> 54      TRUE
#> 55      TRUE
#> 56      TRUE
#> 57      TRUE
#> 58      TRUE
#> 59      TRUE
#> 60      TRUE
#> 61      TRUE
#> 62      TRUE
#> 63      TRUE
#> 64      TRUE
#> 65      TRUE
#> 66      TRUE
#> 67      TRUE
#> 68      TRUE
#> 69      TRUE
#> 70      TRUE
#> 71      TRUE
#> 72      TRUE
#> 73      TRUE
#> 74      TRUE
#> 75      TRUE
#> 76      TRUE
#> 77      TRUE
#> 78      TRUE
#> 79      TRUE
#> 80      TRUE
#> 81      TRUE
#> 82      TRUE
#> 83      TRUE
#> 84      TRUE
#> 85      TRUE
#> 86      TRUE
#> 87      TRUE
#> 88      TRUE
#> 89      TRUE
#> 90      TRUE
#> 91      TRUE
#> 92      TRUE
#> 93      TRUE
#> 94      TRUE
#> 95      TRUE
#> 96      TRUE
#> 97      TRUE
#> 98      TRUE
#> 99      TRUE
#> 100     TRUE
#> 101     TRUE
#> 102     TRUE
#> 103     TRUE
#> 104     TRUE
#> 105     TRUE
#> 106     TRUE
#> 107     TRUE
#> 108     TRUE
#> 109     TRUE
#> 110     TRUE
#> 111     TRUE
#> 112     TRUE
#> 113     TRUE
#> 114     TRUE
#> 115     TRUE
#> 116     TRUE
#> 117     TRUE
#> 118     TRUE
#> 119     TRUE
#> 120     TRUE
#> 121     TRUE
#> 122     TRUE
#> 123     TRUE
#> 124     TRUE
#> 125     TRUE
#> 126     TRUE
#> 127     TRUE
#> 128     TRUE
#> 129     TRUE
#> 130     TRUE
#> 131     TRUE
#> 132     TRUE
#> 133     TRUE
#> 134     TRUE
#> 135     TRUE
#> 136     TRUE
#> 137     TRUE
#> 138     TRUE
#> 139     TRUE
#> 140     TRUE
#> 141     TRUE

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)

#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x111fac0>
#> 

getDriverLongName(getDriver(x))

#> [1] "JPEG JFIF"

GDAL.close(x)