Class "GDALDriver": GDAL Driver Object

GDALDriver objects encapsulate GDAL file format drivers. GDALDriver inherits from GDALMajorObject-class.

getGDALDriverNames()
gdalDrivers()
getDriverName(driver)
getDriverLongName(driver)
getGDALVersionInfo(str = "--version")
getGDALCheckVersion()
getGDALwithGEOS()
rgdal_extSoftVersion()
getCPLConfigOption(ConfigOption)
setCPLConfigOption(ConfigOption, value)
get_cached_orig_PROJ_LIB()
get_cached_orig_GDAL_DATA()
get_cached_set_PROJ_LIB()
get_cached_set_GDAL_DATA()

Arguments

driver
An object inheriting from class 'GDALDriver'

str
A string, may be one of "--version", "VERSION_NUM", "RELEASE_DATE", "RELEASE_NAME"

ConfigOption
CPL configure option documented in http://trac.osgeo.org/gdal/wiki/ConfigOptions and elsewhere in GDAL source code

value
a string value to set a CPL option; NULL is used to unset the CPL option

Objects from the Class

Objects can be created by calls of the form new("GDALDriver", name, handle), where name: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL).

Slots

handle:

Object of class "externalptr", from class "GDALMajorObject", used internally; not for public consumption

Extends

Class "GDALMajorObject", directly.

Methods

initialize

signature(.Object = "GDALDriver"): drivername: a string giving the name of a GDAL driver, handle: used internally; not for public consumption (default = NULL)

Details

getGDALDriverNames, gdalDrivers:

returns all driver names currently installed in GDAL, with their declared create and copy status (some drivers can create datasets, others can only copy from a prototype with a different driver.

getDriverName:

returns the GDAL driver name associated with the driver object.

getDriverLongName:

returns a longer driver name.

getGDALVersionInfo:

returns the version of the GDAL runtime shared object.

getGDALCheckVersion:

checks the version of the GDAL headers used when building the package (GDAL_VERSION_MAJOR, GDAL_VERSION_MINOR) - if the two versions differ, problems may arise (the C++ API/ABI may have changed), and rgdal should be re-installed

getGDALwithGEOS:

because drivers may behave differently if GDAL itself was built with GEOS support, the function uses a heuristic to check whether GDAL has access to the GEOS Union function or not

get_cached_orig_PROJ_LIB, get_cached_orig_GDAL_DATA

The values of environment variables PROJ_LIB and GDAL_DATA as read when this package was loaded

get_cached_set_PROJ_LIB, get_cached_set_GDAL_DATA

If not "", the values set when loading this package to point to metadata files included in CRAN binary packages

Note

Loading the rgdal package changes the GDAL\_DATA environmental variable to the GDAL support files bundled with the package.

See also

Examples

gdalDrivers()

#>                    name
#> 1               AAIGrid
#> 2                  ACE2
#> 3                  ADRG
#> 4                   AIG
#> 5                AirSAR
#> 6                   ARG
#> 7                   BAG
#> 8                BIGGIF
#> 9                   BLX
#> 10                  BMP
#> 11                  BSB
#> 12                   BT
#> 13                  BYN
#> 14                  CAD
#> 15                 CALS
#> 16                 CEOS
#> 17                COASP
#> 18                  COG
#> 19                COSAR
#> 20                  CPG
#> 21              CTable2
#> 22                  CTG
#> 23              DERIVED
#> 24                DIMAP
#> 25                DIPEx
#> 26                 DOQ1
#> 27                 DOQ2
#> 28                 DTED
#> 29              E00GRID
#> 30              ECRGTOC
#> 31                 EHdr
#> 32                  EIR
#> 33                 ELAS
#> 34                 ENVI
#> 35                  ERS
#> 36                 ESAT
#> 37                  EXR
#> 38                 FAST
#> 39                  FIT
#> 40              FujiBAS
#> 41               GenBin
#> 42                  GFF
#> 43                  GIF
#> 44                  GMT
#> 45                 GPKG
#> 46       GRASSASCIIGrid
#> 47                 GRIB
#> 48                GS7BG
#> 49                 GSAG
#> 50                 GSBG
#> 51                  GSC
#> 52                GTiff
#> 53                  GTX
#> 54                  GXF
#> 55                 HDF5
#> 56            HDF5Image
#> 57                  HF2
#> 58                  HFA
#> 59                  IDA
#> 60  IGNFHeightASCIIGrid
#> 61                ILWIS
#> 62                 INGR
#> 63                 IRIS
#> 64                 ISCE
#> 65                  ISG
#> 66                ISIS2
#> 67                ISIS3
#> 68           JAXAPALSAR
#> 69                 JDEM
#> 70          JP2OpenJPEG
#> 71                 JPEG
#> 72      KMLSUPEROVERLAY
#> 73                  KRO
#> 74                  L1B
#> 75                  LAN
#> 76                  LCP
#> 77             Leveller
#> 78               LOSLAS
#> 79                  MAP
#> 80              MBTiles
#> 81                  MEM
#> 82                  MFF
#> 83                 MFF2
#> 84                  MRF
#> 85                 MSGN
#> 86                  NDF
#> 87               netCDF
#> 88             NGSGEOID
#> 89                 NITF
#> 90                 NTv1
#> 91                 NTv2
#> 92              NWT_GRC
#> 93              NWT_GRD
#> 94                  OZI
#> 95                 PAux
#> 96               PCIDSK
#> 97             PCRaster
#> 98                  PDF
#> 99                  PDS
#> 100                PDS4
#> 101                 PNG
#> 102                 PNM
#> 103       PostGISRaster
#> 104                 PRF
#> 105                   R
#> 106          Rasterlite
#> 107                 RIK
#> 108                 RMF
#> 109             ROI_PAC
#> 110              RPFTOC
#> 111             RRASTER
#> 112                 RS2
#> 113                 RST
#> 114                SAFE
#> 115                SAGA
#> 116            SAR_CEOS
#> 117                SDTS
#> 118           SENTINEL2
#> 119                 SGI
#> 120              SIGDEM
#> 121              SNODAS
#> 122                 SRP
#> 123             SRTMHGT
#> 124            Terragen
#> 125                 TIL
#> 126                 TSX
#> 127             USGSDEM
#> 128               VICAR
#> 129                 VRT
#> 130                WEBP
#> 131                 XPM
#> 132                 XYZ
#> 133                ZMap
#>                                                    long_name create  copy
#> 1                                        Arc/Info ASCII Grid  FALSE  TRUE
#> 2                                                       ACE2  FALSE FALSE
#> 3                              ARC Digitized Raster Graphics   TRUE FALSE
#> 4                                       Arc/Info Binary Grid  FALSE FALSE
#> 5                                  AirSAR Polarimetric Image  FALSE FALSE
#> 6                                  Azavea Raster Grid format  FALSE  TRUE
#> 7                                 Bathymetry Attributed Grid  FALSE  TRUE
#> 8                         Graphics Interchange Format (.gif)  FALSE FALSE
#> 9                                       Magellan topo (.blx)  FALSE  TRUE
#> 10                      MS Windows Device Independent Bitmap   TRUE FALSE
#> 11                               Maptech BSB Nautical Charts  FALSE FALSE
#> 12                       VTP .bt (Binary Terrain) 1.3 Format   TRUE FALSE
#> 13                          Natural Resources Canada's Geoid   TRUE FALSE
#> 14                                            AutoCAD Driver  FALSE FALSE
#> 15                                             CALS (Type 1)  FALSE  TRUE
#> 16                                                CEOS Image  FALSE FALSE
#> 17                           DRDC COASP SAR Processor Raster  FALSE FALSE
#> 18                         Cloud optimized GeoTIFF generator  FALSE  TRUE
#> 19                COSAR Annotated Binary Matrix (TerraSAR-X)  FALSE FALSE
#> 20                                           Convair PolGASP  FALSE FALSE
#> 21                                  CTable2 Datum Grid Shift   TRUE FALSE
#> 22                            USGS LULC Composite Theme Grid  FALSE FALSE
#> 23                Derived datasets using VRT pixel functions  FALSE FALSE
#> 24                                                SPOT DIMAP  FALSE FALSE
#> 25                                                     DIPEx  FALSE FALSE
#> 26                                      USGS DOQ (Old Style)  FALSE FALSE
#> 27                                      USGS DOQ (New Style)  FALSE FALSE
#> 28                                     DTED Elevation Raster  FALSE  TRUE
#> 29                                  Arc/Info Export E00 GRID  FALSE FALSE
#> 30                                           ECRG TOC format  FALSE FALSE
#> 31                                        ESRI .hdr Labelled   TRUE  TRUE
#> 32                                         Erdas Imagine Raw  FALSE FALSE
#> 33                                                      ELAS   TRUE FALSE
#> 34                                        ENVI .hdr Labelled   TRUE FALSE
#> 35                                    ERMapper .ers Labelled   TRUE FALSE
#> 36                                      Envisat Image Format  FALSE FALSE
#> 37                  Extended Dynamic Range Image File Format   TRUE  TRUE
#> 38                                         EOSAT FAST Format  FALSE FALSE
#> 39                                                 FIT Image  FALSE  TRUE
#> 40                                    Fuji BAS Scanner Image  FALSE FALSE
#> 41                            Generic Binary (.hdr Labelled)  FALSE FALSE
#> 42  Ground-based SAR Applications Testbed File Format (.gff)  FALSE FALSE
#> 43                        Graphics Interchange Format (.gif)  FALSE  TRUE
#> 44                                    GMT NetCDF Grid Format  FALSE  TRUE
#> 45                                                GeoPackage   TRUE  TRUE
#> 46                                          GRASS ASCII Grid  FALSE FALSE
#> 47                              GRIdded Binary (.grb, .grb2)  FALSE  TRUE
#> 48                      Golden Software 7 Binary Grid (.grd)   TRUE  TRUE
#> 49                         Golden Software ASCII Grid (.grd)  FALSE  TRUE
#> 50                        Golden Software Binary Grid (.grd)   TRUE  TRUE
#> 51                                               GSC Geogrid  FALSE FALSE
#> 52                                                   GeoTIFF   TRUE  TRUE
#> 53                                  NOAA Vertical Datum .GTX   TRUE FALSE
#> 54                              GeoSoft Grid Exchange Format  FALSE FALSE
#> 55                        Hierarchical Data Format Release 5  FALSE FALSE
#> 56                                              HDF5 Dataset  FALSE FALSE
#> 57                                HF2/HFZ heightfield raster  FALSE  TRUE
#> 58                               Erdas Imagine Images (.img)   TRUE  TRUE
#> 59                                   Image Data and Analysis   TRUE FALSE
#> 60                   IGN France height correction ASCII Grid  FALSE FALSE
#> 61                                          ILWIS Raster Map   TRUE  TRUE
#> 62                                         Intergraph Raster   TRUE  TRUE
#> 63                              IRIS data (.PPI, .CAPPi etc)  FALSE FALSE
#> 64                                               ISCE raster   TRUE FALSE
#> 65                       International Service for the Geoid  FALSE FALSE
#> 66                   USGS Astrogeology ISIS cube (Version 2)   TRUE FALSE
#> 67                   USGS Astrogeology ISIS cube (Version 3)   TRUE  TRUE
#> 68                JAXA PALSAR Product Reader (Level 1.1/1.5)  FALSE FALSE
#> 69                                       Japanese DEM (.mem)  FALSE FALSE
#> 70                JPEG-2000 driver based on OpenJPEG library  FALSE  TRUE
#> 71                                                 JPEG JFIF  FALSE  TRUE
#> 72                                         Kml Super Overlay  FALSE  TRUE
#> 73                                                 KOLOR Raw   TRUE FALSE
#> 74                      NOAA Polar Orbiter Level 1b Data Set  FALSE FALSE
#> 75                                           Erdas .LAN/.GIS   TRUE FALSE
#> 76                         FARSITE v.4 Landscape File (.lcp)  FALSE  TRUE
#> 77                                      Leveller heightfield   TRUE FALSE
#> 78                         NADCON .los/.las Datum Grid Shift  FALSE FALSE
#> 79                                          OziExplorer .MAP  FALSE FALSE
#> 80                                                   MBTiles   TRUE  TRUE
#> 81                                          In Memory Raster   TRUE FALSE
#> 82                                         Vexcel MFF Raster   TRUE  TRUE
#> 83                                  Vexcel MFF2 (HKV) Raster   TRUE  TRUE
#> 84                                        Meta Raster Format   TRUE  TRUE
#> 85                            EUMETSAT Archive native (.nat)  FALSE FALSE
#> 86                                         NLAPS Data Format  FALSE FALSE
#> 87                                Network Common Data Format   TRUE  TRUE
#> 88                               NOAA NGS Geoid Height Grids  FALSE FALSE
#> 89                      National Imagery Transmission Format   TRUE  TRUE
#> 90                                     NTv1 Datum Grid Shift  FALSE FALSE
#> 91                                     NTv2 Datum Grid Shift   TRUE FALSE
#> 92                Northwood Classified Grid Format .grc/.tab  FALSE FALSE
#> 93                   Northwood Numeric Grid Format .grd/.tab   TRUE  TRUE
#> 94                                    OziExplorer Image File  FALSE FALSE
#> 95                                         PCI .aux Labelled   TRUE FALSE
#> 96                                      PCIDSK Database File   TRUE FALSE
#> 97                                      PCRaster Raster File   TRUE  TRUE
#> 98                                            Geospatial PDF   TRUE  TRUE
#> 99                                NASA Planetary Data System  FALSE FALSE
#> 100                             NASA Planetary Data System 4   TRUE  TRUE
#> 101                                Portable Network Graphics  FALSE  TRUE
#> 102                          Portable Pixmap Format (netpbm)   TRUE FALSE
#> 103                                    PostGIS Raster driver  FALSE  TRUE
#> 104                                      Racurs PHOTOMOD PRF  FALSE FALSE
#> 105                                      R Object Data Store  FALSE  TRUE
#> 106                                               Rasterlite  FALSE  TRUE
#> 107                                  Swedish Grid RIK (.rik)  FALSE FALSE
#> 108                                     Raster Matrix Format   TRUE FALSE
#> 109                                           ROI_PAC raster   TRUE FALSE
#> 110                         Raster Product Format TOC format  FALSE FALSE
#> 111                                                 R Raster   TRUE  TRUE
#> 112                                   RadarSat 2 XML Product  FALSE FALSE
#> 113                                        Idrisi Raster A.1   TRUE  TRUE
#> 114                              Sentinel-1 SAR SAFE Product  FALSE FALSE
#> 115                  SAGA GIS Binary Grid (.sdat, .sg-grd-z)   TRUE  TRUE
#> 116                                           CEOS SAR Image  FALSE FALSE
#> 117                                              SDTS Raster  FALSE FALSE
#> 118                                               Sentinel 2  FALSE FALSE
#> 119                                SGI Image File Format 1.0   TRUE FALSE
#> 120                       Scaled Integer Gridded DEM .sigdem  FALSE  TRUE
#> 121                            Snow Data Assimilation System  FALSE FALSE
#> 122                      Standard Raster Product (ASRP/USRP)  FALSE FALSE
#> 123                                      SRTMHGT File Format  FALSE  TRUE
#> 124                                     Terragen heightfield   TRUE FALSE
#> 125                                          EarthWatch .TIL  FALSE FALSE
#> 126                                       TerraSAR-X Product  FALSE FALSE
#> 127                       USGS Optional ASCII DEM (and CDED)  FALSE  TRUE
#> 128                                          MIPL VICAR file   TRUE  TRUE
#> 129                                           Virtual Raster   TRUE  TRUE
#> 130                                                     WEBP  FALSE  TRUE
#> 131                                        X11 PixMap Format  FALSE  TRUE
#> 132                                        ASCII Gridded XYZ  FALSE  TRUE
#> 133                                           ZMap Plus Grid  FALSE  TRUE
#>     isRaster
#> 1       TRUE
#> 2       TRUE
#> 3       TRUE
#> 4       TRUE
#> 5       TRUE
#> 6       TRUE
#> 7       TRUE
#> 8       TRUE
#> 9       TRUE
#> 10      TRUE
#> 11      TRUE
#> 12      TRUE
#> 13      TRUE
#> 14      TRUE
#> 15      TRUE
#> 16      TRUE
#> 17      TRUE
#> 18      TRUE
#> 19      TRUE
#> 20      TRUE
#> 21      TRUE
#> 22      TRUE
#> 23      TRUE
#> 24      TRUE
#> 25      TRUE
#> 26      TRUE
#> 27      TRUE
#> 28      TRUE
#> 29      TRUE
#> 30      TRUE
#> 31      TRUE
#> 32      TRUE
#> 33      TRUE
#> 34      TRUE
#> 35      TRUE
#> 36      TRUE
#> 37      TRUE
#> 38      TRUE
#> 39      TRUE
#> 40      TRUE
#> 41      TRUE
#> 42      TRUE
#> 43      TRUE
#> 44      TRUE
#> 45      TRUE
#> 46      TRUE
#> 47      TRUE
#> 48      TRUE
#> 49      TRUE
#> 50      TRUE
#> 51      TRUE
#> 52      TRUE
#> 53      TRUE
#> 54      TRUE
#> 55      TRUE
#> 56      TRUE
#> 57      TRUE
#> 58      TRUE
#> 59      TRUE
#> 60      TRUE
#> 61      TRUE
#> 62      TRUE
#> 63      TRUE
#> 64      TRUE
#> 65      TRUE
#> 66      TRUE
#> 67      TRUE
#> 68      TRUE
#> 69      TRUE
#> 70      TRUE
#> 71      TRUE
#> 72      TRUE
#> 73      TRUE
#> 74      TRUE
#> 75      TRUE
#> 76      TRUE
#> 77      TRUE
#> 78      TRUE
#> 79      TRUE
#> 80      TRUE
#> 81      TRUE
#> 82      TRUE
#> 83      TRUE
#> 84      TRUE
#> 85      TRUE
#> 86      TRUE
#> 87      TRUE
#> 88      TRUE
#> 89      TRUE
#> 90      TRUE
#> 91      TRUE
#> 92      TRUE
#> 93      TRUE
#> 94      TRUE
#> 95      TRUE
#> 96      TRUE
#> 97      TRUE
#> 98      TRUE
#> 99      TRUE
#> 100     TRUE
#> 101     TRUE
#> 102     TRUE
#> 103     TRUE
#> 104     TRUE
#> 105     TRUE
#> 106     TRUE
#> 107     TRUE
#> 108     TRUE
#> 109     TRUE
#> 110     TRUE
#> 111     TRUE
#> 112     TRUE
#> 113     TRUE
#> 114     TRUE
#> 115     TRUE
#> 116     TRUE
#> 117     TRUE
#> 118     TRUE
#> 119     TRUE
#> 120     TRUE
#> 121     TRUE
#> 122     TRUE
#> 123     TRUE
#> 124     TRUE
#> 125     TRUE
#> 126     TRUE
#> 127     TRUE
#> 128     TRUE
#> 129     TRUE
#> 130     TRUE
#> 131     TRUE
#> 132     TRUE
#> 133     TRUE

logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
getDriver(x)

#> An object of class "GDALDriver"
#> Slot "handle":
#> <pointer: 0x6cd1840>
#> 

getDriverLongName(getDriver(x))

#> [1] "JPEG JFIF"

GDAL.close(x)