GDALMajorObject-class {rgdal}	R Documentation

Class "GDALMajorObject"

Description

"GDALMajorObject" is a virtual base class for all GDAL objects.

Usage

getDescription(object)

Arguments

object
an object inheriting from "GDALMajorObject"

Details

getDescription:

returns a descrption string associated with the object. No setter method is defined because GDAL dataset objects use the description to hold the filename attached to the dataset. It would not be good to change that mid-stream.

Objects from the Class

Objects can be created by calls of the form new("GDALMajorObject", ...), but are only created for classes that extend this class.

Slots

handle:

Object of class "externalptr", used internally; not for public consumption

Methods

No methods defined with class "GDALMajorObject" in the signature.

Author(s)

Timothy H. Keitt, modified by Roger Bivand

References

http://www.gdal.org/

See Also

GDALDriver-class, GDALReadOnlyDataset-class, GDALDataset-class and GDALTransientDataset-class

Examples

driver <- new('GDALDriver', as.character(getGDALDriverNames()[1,1]))
driver
rm(driver)
logo <- system.file("pictures/logo.jpg", package="rgdal")[1]
x <- new("GDALReadOnlyDataset", logo)
x
getDescription(x)
dim(x)
GDAL.close(x)

[Package rgdal version 1.4-8 Index]