RGB2PCT {rgdal}	R Documentation

Convert RGB three band to single band colour table

Description

This function converts a three-band GDALReadOnlyDataset into a single band of colour indices as a GDALTransientDataset.

Usage

RGB2PCT(x, band, driver.name = 'MEM', ncolors = 256, set.ctab = TRUE)

Arguments

x
a three-band GDALReadOnlyDataset object

band
a vector of numbers, recycled up to 3 in length

driver.name
default MEM

ncolors
a number of colours between 2 and 256

set.ctab
default TRUE, when the dithered dataset handle is returned, otherwise a list of the dataset and the PCT colour table

Value

The value returned is a either GDALTransientDataset or a list of a GDALTransientDataset and a colour table.

Author(s)

Tim Keitt

References

http://www.gdal.org/

Examples

logo <- system.file("pictures/Rlogo.jpg", package="rgdal")[1]
x <- GDAL.open(logo)
dim(x)
dx <- RGB2PCT(x, band=1:3)
displayDataset(dx)
dim(dx)
GDAL.close(x)
GDAL.close(dx)

[Package rgdal version 1.1-10 Index]