List PROJ 6 coordinate operations

List PROJ 6 coordinate operations for a pair of source/target coordinate reference systems

list_coordOps(src_crs, tgt_crs, area_of_interest = as.numeric(NA),
 strict_containment = FALSE, visualization_order = TRUE)

Arguments

src_crs
Source coordinate reference system string

tgt_crs
Target coordinate reference system string

area_of_interest
Numeric vector; either NA, or the xmin, ymin, xmax, ymax of the bounding box of the area of interest. This may be used to restrict the search for coordinate operations

strict_containment
default FALSE, permit partial matching of the area of interest; if TRUE strictly contain the area of interest. The area of interest is either as given, or as implied by the source/target coordinate reference systems (FIXME)

visualization_order
default TRUE, always choose x or longitude for the first axis; if FALSE, follow the axis orders given by the coordinate reference systems when constructing the coordinate operation

Details

(FIXME)

Value

A data frame with rows showing the coordinate operations found, and columns:

description

String describing the operation

definition

PROJ pipeline for executing the operation

accuracy

Accuracy in meters, if negative, unknown

instantiable

Can this operation be carried out with available resources

ballpark

Does this operation only have ballpark accuracy

number_grids

The number of grids required for the operation

The object has a "grids" attribute containing a nested list of grids for each coordinate operations found; if number_grids == 0, NULL, otherwise a list of grids. For each grid required, the short and long names of the grid are given, the package name if available in a PROJ grid package, and the download URL for that package. Three logical variables report whether the grid may be downloaded directly, whether it has an open license, and whether it is available.
References

https://proj.org/

Note

Fragile: work in progress

Examples

exts <- rgdal_extSoftVersion()
run <- substring(exts["GDAL"], 1, 1) >= "3" && substring(exts["PROJ"], 1, 1) >= "6"
if (run) {
  o <- list_coordOps("EPSG:27700", "EPSG:4326")
  print(o[1,])
}

#>                                                                           description
#> 1 Inverse of British National Grid + OSGB 1936 to WGS 84 (6) + axis order change (2D)
#>                                                                                                                                                                                                                                                                                                                                                                                          definition
#> 1 +proj=pipeline +step +inv +proj=tmerc +lat_0=49 +lon_0=-2 +k=0.9996012717 +x_0=400000 +y_0=-100000 +ellps=airy +step +proj=push +v_3 +step +proj=cart +ellps=airy +step +proj=helmert +x=446.448 +y=-125.157 +z=542.06 +rx=0.15 +ry=0.247 +rz=0.842 +s=-20.489 +convention=position_vector +step +inv +proj=cart +ellps=WGS84 +step +proj=pop +v_3 +step +proj=unitconvert +xy_in=rad +xy_out=deg
#>   accuracy instantiable ballpark number_grids
#> 1        2         TRUE    FALSE            0

if (run) {
  attr(o, "grids")[[1]][[1]]
}

#> NULL