Make a data frame of EPSG projection codes

Make a data frame of the now-defunct European Petroleum Survey Group (EPSG) geodetic parameter dataset as distributed with PROJ.4 software and included in this package. Because finding the correct projection specification is not easy, lists still known as EPSG lists are maintained, and more generally retrieved from databases. The data collated here are as distributed with PROJ.4.

make_EPSG(file)

Arguments

file
file name of the file matching EPSG codes and PROJ.4 arguments, should usually be autodetected

Value

returns a data frame with columns:

code

integer column of EPSG code numbers

note

character column of notes as included in the file

prj4

character column of PROJ.4 arguments for the equivalent projection definitions

...
References

http://www.epsg.org/

Note

See also Clifford J. Mugnier's Grids \& Datums columns in Photogrammetric Engineering & Remote Sensing, http://www.asprs.org/a/resources/grids/, see also GridsDatums.

Examples

EPSG <- make_EPSG()
attr(EPSG, "metadata") # PROJ.4 5.0.0 and later include EPSG version

#> [1] "<metadata> +version=9.2.0 +origin=EPSG +lastupdate=2017-12-17"

EPSG[grep("Oslo", EPSG$note), 1:2]

#>       code                              note
#> 428   4817                 # NGO 1948 (Oslo)
#> 4674 27391    # NGO 1948 (Oslo) / NGO zone I
#> 4675 27392   # NGO 1948 (Oslo) / NGO zone II
#> 4676 27393  # NGO 1948 (Oslo) / NGO zone III
#> 4677 27394   # NGO 1948 (Oslo) / NGO zone IV
#> 4678 27395    # NGO 1948 (Oslo) / NGO zone V
#> 4679 27396   # NGO 1948 (Oslo) / NGO zone VI
#> 4680 27397  # NGO 1948 (Oslo) / NGO zone VII
#> 4681 27398 # NGO 1948 (Oslo) / NGO zone VIII

EPSG[1925:1927, 3]

#> [1] "+proj=lcc +lat_1=38.56666666666667 +lat_2=37.26666666666667 +lat_0=36.66666666666666 +lon_0=-98.5 +x_0=399999.99998984 +y_0=399999.99998984 +datum=NAD83 +units=us-ft +no_defs"
#> [2] "+proj=tmerc +lat_0=34.75 +lon_0=-115.5833333333333 +k=0.9999 +x_0=200000.00001016 +y_0=8000000.000010163 +datum=NAD83 +units=us-ft +no_defs"                                   
#> [3] "+proj=tmerc +lat_0=34.75 +lon_0=-116.6666666666667 +k=0.9999 +x_0=500000.00001016 +y_0=6000000 +datum=NAD83 +units=us-ft +no_defs"                                             

EPSG[grep("Poland", EPSG$note), 1:2]

#>      code                                            note
#> 685  2171 # Pulkovo 1942(58) / Poland zone I (deprecated)
#> 686  2172             # Pulkovo 1942(58) / Poland zone II
#> 687  2173            # Pulkovo 1942(58) / Poland zone III
#> 688  2174             # Pulkovo 1942(58) / Poland zone IV
#> 689  2175              # Pulkovo 1942(58) / Poland zone V
#> 690  2176                 # ETRS89 / Poland CS2000 zone 5
#> 691  2177                 # ETRS89 / Poland CS2000 zone 6
#> 692  2178                 # ETRS89 / Poland CS2000 zone 7
#> 693  2179                 # ETRS89 / Poland CS2000 zone 8
#> 694  2180                          # ETRS89 / Poland CS92
#> 1625 3120              # Pulkovo 1942(58) / Poland zone I

EPSG[grep("Amersfoort", EPSG$note), 1:2]

#>       code                               note
#> 218   4289                       # Amersfoort
#> 4781 28991              # Amersfoort / RD Old
#> 4782 28992              # Amersfoort / RD New
#> 5815  7415 # Amersfoort / RD New + NAP height

EPSG[grep("North Carolina", EPSG$note), 1:2]

#>       code                                               note
#> 771   2264                    # NAD83 / North Carolina (ftUS)
#> 1863  3358                     # NAD83(HARN) / North Carolina
#> 1864  3359 # NAD83(HARN) / North Carolina (ftUS) (deprecated)
#> 1909  3404              # NAD83(HARN) / North Carolina (ftUS)
#> 2136  3631                 # NAD83(NSRS2007) / North Carolina
#> 2137  3632          # NAD83(NSRS2007) / North Carolina (ftUS)
#> 3228  6542                     # NAD83(2011) / North Carolina
#> 3229  6543              # NAD83(2011) / North Carolina (ftUS)
#> 4985 32019                           # NAD27 / North Carolina
#> 5056 32119                           # NAD83 / North Carolina

EPSG[2202, 3]

#> [1] "+proj=lcc +lat_1=46.76666666666667 +lat_2=45.56666666666667 +lat_0=45.16666666666666 +lon_0=-90 +x_0=600000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"