Make a data frame of EPSG projection codes

Make a data frame of the European Petroleum Survey Group (EPSG) geodetic parameter dataset as distributed with PROJ.4 software (prior to PROJ 6.0.0, March 2019, only the CSV file, from March 2019 with PROJ >= 6 from the SQLite database). Because finding the correct projection specification is not easy, lists still known as EPSG lists are maintained, and more generally retrieved from databases. The data collated here are as distributed with PROJ.4.

make_EPSG(file)

Arguments

file
file name of the file matching EPSG codes and PROJ.4 arguments, should usually be autodetected; not used for PROJ >= 6

Value

returns a data frame with columns:

code

integer column of EPSG code numbers

note

character column of notes as included in the file

prj4

character column of PROJ.4 arguments for the equivalent projection definitions

prj_method

extra character column from PROJ 6 showing the projection method

...
References

http://www.epsg.org/

Note

See also Clifford J. Mugnier's Grids \& Datums columns in Photogrammetric Engineering & Remote Sensing, http://www.asprs.org/a/resources/grids/, see also GridsDatums.

Examples

EPSG <- try(make_EPSG())
# from PROJ 6.0.0, EPSG data is no longer stored in a flat file
if (class(EPSG) != "try-error") attr(EPSG, "metadata") # PROJ.4 5.0.0 and later include EPSG version

#> NULL

if (class(EPSG) != "try-error") EPSG[grep("Oslo", EPSG$note), 1:2]

#>       code                            note
#> 515   4817                 NGO 1948 (Oslo)
#> 2386 27391    NGO 1948 (Oslo) / NGO zone I
#> 2387 27392   NGO 1948 (Oslo) / NGO zone II
#> 2388 27393  NGO 1948 (Oslo) / NGO zone III
#> 2389 27394   NGO 1948 (Oslo) / NGO zone IV
#> 2390 27395    NGO 1948 (Oslo) / NGO zone V
#> 2391 27396   NGO 1948 (Oslo) / NGO zone VI
#> 2392 27397  NGO 1948 (Oslo) / NGO zone VII
#> 2393 27398 NGO 1948 (Oslo) / NGO zone VIII

if (class(EPSG) != "try-error") EPSG[1925:1927, 3]

#> [1] "+proj=utm +zone=31 +ellps=GRS80 +units=m +no_defs +type=crs"
#> [2] "+proj=utm +zone=32 +ellps=GRS80 +units=m +no_defs +type=crs"
#> [3] "+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs +type=crs"

if (class(EPSG) != "try-error") EPSG[grep("Poland", EPSG$note), 1:2]

#>      code                               note
#> 1322 2171   Pulkovo 1942(58) / Poland zone I
#> 1323 2172  Pulkovo 1942(58) / Poland zone II
#> 1324 2173 Pulkovo 1942(58) / Poland zone III
#> 1325 2174  Pulkovo 1942(58) / Poland zone IV
#> 1326 2175   Pulkovo 1942(58) / Poland zone V
#> 1327 2176      ETRS89 / Poland CS2000 zone 5
#> 1328 2177      ETRS89 / Poland CS2000 zone 6
#> 1329 2178      ETRS89 / Poland CS2000 zone 7
#> 1333 2179      ETRS89 / Poland CS2000 zone 8
#> 1334 2180               ETRS89 / Poland CS92
#> 2969 3120   Pulkovo 1942(58) / Poland zone I

if (class(EPSG) != "try-error") EPSG[grep("Amersfoort", EPSG$note), 1:2]

#>       code                             note
#> 230   4289                       Amersfoort
#> 2653 28991              Amersfoort / RD Old
#> 2654 28992              Amersfoort / RD New
#> 6405  7415 Amersfoort / RD New + NAP height

if (class(EPSG) != "try-error") EPSG[grep("North Carolina", EPSG$note), 1:2]

#>       code                                                 note
#> 1468  2264                        NAD83 / North Carolina (ftUS)
#> 3158 32019                               NAD27 / North Carolina
#> 3239 32119                               NAD83 / North Carolina
#> 3819  3358                         NAD83(HARN) / North Carolina
#> 3820  3359                  NAD83(HARN) / North Carolina (ftUS)
#> 3865  3404                  NAD83(HARN) / North Carolina (ftUS)
#> 4092  3631                     NAD83(NSRS2007) / North Carolina
#> 4093  3632              NAD83(NSRS2007) / North Carolina (ftUS)
#> 5187  6542                         NAD83(2011) / North Carolina
#> 5188  6543                  NAD83(2011) / North Carolina (ftUS)
#> 6489  8768 NAD83 / North Carolina (ftUS) + NAVD88 height (ftUS)

if (class(EPSG) != "try-error") EPSG[2202, 3]

#> [1] "+proj=tmerc +lat_0=0 +lon_0=-96 +k=0.9999 +x_0=304800 +y_0=0 +ellps=GRS80 +units=m +no_defs +type=crs"