Make a data frame of EPSG projection codes

Make a data frame of the European Petroleum Survey Group (EPSG) geodetic parameter dataset as distributed with PROJ.4 software (prior to PROJ 6.0.0, March 2019, only the CSV file, from March 2019 with PROJ >= 6 from the SQLite database). Because finding the correct projection specification is not easy, lists still known as EPSG lists are maintained, and more generally retrieved from databases. The data collated here are as distributed with PROJ.4.

make_EPSG(file)

Arguments

file
file name of the file matching EPSG codes and PROJ.4 arguments, should usually be autodetected; not used for PROJ >= 6

Value

returns a data frame with columns:

code

integer column of EPSG code numbers

note

character column of notes as included in the file

prj4

character column of PROJ.4 arguments for the equivalent projection definitions

prj_method

extra character column from PROJ 6 showing the projection method

...
References

http://www.epsg.org/

Note

See also Clifford J. Mugnier's Grids \& Datums columns in Photogrammetric Engineering & Remote Sensing, http://www.asprs.org/a/resources/grids/, see also GridsDatums.

Examples

EPSG <- try(make_EPSG())
# from PROJ 6.0.0, EPSG data is no longer stored in a flat file
if (!inherits(EPSG, "try-error")) attr(EPSG, "metadata")

#> NULL

# PROJ.4 5.0.0 and later include EPSG version
if (!inherits(EPSG, "try-error")) EPSG[grep("Oslo", EPSG$note), 1:2]

#>       code                            note
#> 515   4817                 NGO 1948 (Oslo)
#> 2419 27391    NGO 1948 (Oslo) / NGO zone I
#> 2420 27392   NGO 1948 (Oslo) / NGO zone II
#> 2421 27393  NGO 1948 (Oslo) / NGO zone III
#> 2422 27394   NGO 1948 (Oslo) / NGO zone IV
#> 2423 27395    NGO 1948 (Oslo) / NGO zone V
#> 2424 27396   NGO 1948 (Oslo) / NGO zone VI
#> 2425 27397  NGO 1948 (Oslo) / NGO zone VII
#> 2426 27398 NGO 1948 (Oslo) / NGO zone VIII

if (!inherits(EPSG, "try-error")) EPSG[1925:1927, 3]

#> [1] "+proj=tmerc +lat_0=0 +lon_0=114 +k=1 +x_0=38500000 +y_0=0 +ellps=krass +units=m +no_defs +type=crs"
#> [2] "+proj=tmerc +lat_0=0 +lon_0=117 +k=1 +x_0=39500000 +y_0=0 +ellps=krass +units=m +no_defs +type=crs"
#> [3] "+proj=tmerc +lat_0=0 +lon_0=120 +k=1 +x_0=40500000 +y_0=0 +ellps=krass +units=m +no_defs +type=crs"

if (!inherits(EPSG, "try-error")) EPSG[grep("Poland", EPSG$note), 1:2]

#>      code                               note
#> 1355 2171   Pulkovo 1942(58) / Poland zone I
#> 1356 2172  Pulkovo 1942(58) / Poland zone II
#> 1357 2173 Pulkovo 1942(58) / Poland zone III
#> 1358 2174  Pulkovo 1942(58) / Poland zone IV
#> 1359 2175   Pulkovo 1942(58) / Poland zone V
#> 1360 2176      ETRS89 / Poland CS2000 zone 5
#> 1361 2177      ETRS89 / Poland CS2000 zone 6
#> 1362 2178      ETRS89 / Poland CS2000 zone 7
#> 1366 2179      ETRS89 / Poland CS2000 zone 8
#> 1367 2180               ETRS89 / Poland CS92
#> 3002 3120   Pulkovo 1942(58) / Poland zone I

if (!inherits(EPSG, "try-error")) EPSG[grep("Amersfoort", EPSG$note), 1:2]

#>       code                             note
#> 230   4289                       Amersfoort
#> 2686 28991              Amersfoort / RD Old
#> 2687 28992              Amersfoort / RD New
#> 6484  7415 Amersfoort / RD New + NAP height

if (!inherits(EPSG, "try-error")) EPSG[grep("North Carolina", EPSG$note), 1:2]

#>       code                                                 note
#> 1501  2264                        NAD83 / North Carolina (ftUS)
#> 3191 32019                               NAD27 / North Carolina
#> 3272 32119                               NAD83 / North Carolina
#> 3852  3358                         NAD83(HARN) / North Carolina
#> 3853  3359                  NAD83(HARN) / North Carolina (ftUS)
#> 3898  3404                  NAD83(HARN) / North Carolina (ftUS)
#> 4125  3631                     NAD83(NSRS2007) / North Carolina
#> 4126  3632              NAD83(NSRS2007) / North Carolina (ftUS)
#> 5220  6542                         NAD83(2011) / North Carolina
#> 5221  6543                  NAD83(2011) / North Carolina (ftUS)
#> 6568  8768 NAD83 / North Carolina (ftUS) + NAVD88 height (ftUS)

if (!inherits(EPSG, "try-error")) EPSG[2202, 3]

#> [1] "+proj=lcc +lat_0=46.5 +lon_0=-93.1 +lat_1=48.6333333333333 +lat_2=47.0333333333333 +x_0=800000.00001016 +y_0=99999.99998984 +datum=NAD83 +units=us-ft +no_defs +type=crs"