List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()
GDAL_OSR_PROJ()
GDALis3ormore()
PROJis6ormore()
new_proj_and_gdal()

Arguments

type

One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu. From PROJ 6, "datum" is not available. From PROJ 7.1.0, "units" returns the conversion factor as numeric, not character.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

https://proj.org/

Author

Roger Bivand Roger.Bivand@nhh.no

Note

Loading the rgdal package may change the PROJ_LIB environmental variable to the PROJ.4 support files if bundled with binary packages.

Examples

getPROJ4VersionInfo()
#> [1] "Rel. 9.1.1, December 1st, 2022, [PJ_VERSION: 911]"
#> attr(,"short")
#> [1] 911
projInfo()
#>             name
#> 1     adams_hemi
#> 2      adams_ws1
#> 3      adams_ws2
#> 4            aea
#> 5           aeqd
#> 6         affine
#> 7           airy
#> 8         aitoff
#> 9           alsk
#> 10         apian
#> 11        august
#> 12      axisswap
#> 13         bacon
#> 14    bertin1953
#> 15          bipc
#> 16         boggs
#> 17         bonne
#> 18       calcofi
#> 19          cart
#> 20          cass
#> 21            cc
#> 22          ccon
#> 23           cea
#> 24         chamb
#> 25         collg
#> 26     col_urban
#> 27        comill
#> 28         crast
#> 29      defmodel
#> 30   deformation
#> 31         denoy
#> 32          eck1
#> 33          eck2
#> 34          eck3
#> 35          eck4
#> 36          eck5
#> 37          eck6
#> 38       eqearth
#> 39           eqc
#> 40          eqdc
#> 41         euler
#> 42        etmerc
#> 43         fahey
#> 44          fouc
#> 45        fouc_s
#> 46          gall
#> 47          geoc
#> 48    geogoffset
#> 49          geos
#> 50         gins8
#> 51       gn_sinu
#> 52          gnom
#> 53         goode
#> 54          gs48
#> 55          gs50
#> 56         guyou
#> 57        hammer
#> 58        hatano
#> 59       healpix
#> 60      rhealpix
#> 61       helmert
#> 62    hgridshift
#> 63        horner
#> 64           igh
#> 65         igh_o
#> 66         imoll
#> 67       imoll_o
#> 68         imw_p
#> 69          isea
#> 70          kav5
#> 71          kav7
#> 72        krovak
#> 73         labrd
#> 74          laea
#> 75        lagrng
#> 76          larr
#> 77          lask
#> 78        lonlat
#> 79        latlon
#> 80           lcc
#> 81          lcca
#> 82          leac
#> 83        lee_os
#> 84         loxim
#> 85          lsat
#> 86         mbt_s
#> 87       mbt_fps
#> 88        mbtfpp
#> 89        mbtfpq
#> 90        mbtfps
#> 91          merc
#> 92        mil_os
#> 93          mill
#> 94       misrsom
#> 95          moll
#> 96   molobadekas
#> 97    molodensky
#> 98         murd1
#> 99         murd2
#> 100        murd3
#> 101     natearth
#> 102    natearth2
#> 103         nell
#> 104       nell_h
#> 105        nicol
#> 106        nsper
#> 107         nzmg
#> 108         noop
#> 109      ob_tran
#> 110         ocea
#> 111          oea
#> 112        omerc
#> 113        ortel
#> 114        ortho
#> 115       pconic
#> 116    patterson
#> 117     peirce_q
#> 118     pipeline
#> 119         poly
#> 120          pop
#> 121         push
#> 122        putp1
#> 123        putp2
#> 124        putp3
#> 125       putp3p
#> 126       putp4p
#> 127        putp5
#> 128       putp5p
#> 129        putp6
#> 130       putp6p
#> 131      qua_aut
#> 132          qsc
#> 133        robin
#> 134        rouss
#> 135        rpoly
#> 136           s2
#> 137          sch
#> 138          set
#> 139         sinu
#> 140       somerc
#> 141        stere
#> 142       sterea
#> 143      gstmerc
#> 144          tcc
#> 145         tcea
#> 146        times
#> 147     tinshift
#> 148       tissot
#> 149        tmerc
#> 150      tobmerc
#> 151  topocentric
#> 152        tpeqd
#> 153        tpers
#> 154  unitconvert
#> 155          ups
#> 156         urm5
#> 157       urmfps
#> 158          utm
#> 159        vandg
#> 160       vandg2
#> 161       vandg3
#> 162       vandg4
#> 163   vertoffset
#> 164        vitk1
#> 165   vgridshift
#> 166         wag1
#> 167         wag2
#> 168         wag3
#> 169         wag4
#> 170         wag5
#> 171         wag6
#> 172         wag7
#> 173      webmerc
#> 174        weren
#> 175        wink1
#> 176        wink2
#> 177       wintri
#> 178 xyzgridshift
#>                                                         description
#> 1                                      Adams Hemisphere in a Square
#> 2                                         Adams World in a Square I
#> 3                                        Adams World in a Square II
#> 4                                                 Albers Equal Area
#> 5                                             Azimuthal Equidistant
#> 6                                             Affine transformation
#> 7                                                              Airy
#> 8                                                            Aitoff
#> 9                                  Modified Stereographic of Alaska
#> 10                                                 Apian Globular I
#> 11                                              August Epicycloidal
#> 12                                                    Axis ordering
#> 13                                                   Bacon Globular
#> 14                                                      Bertin 1953
#> 15                              Bipolar conic of western hemisphere
#> 16                                                  Boggs Eumorphic
#> 17                                          Bonne (Werner lat_1=90)
#> 18                        Cal Coop Ocean Fish Invest Lines/Stations
#> 19                                   Geodetic/cartesian conversions
#> 20                                                          Cassini
#> 21                                              Central Cylindrical
#> 22                                                    Central Conic
#> 23                                           Equal Area Cylindrical
#> 24                                             Chamberlin Trimetric
#> 25                                                        Collignon
#> 26                                                   Colombia Urban
#> 27                                                   Compact Miller
#> 28                                   Craster Parabolic (Putnins P4)
#> 29                                                Deformation model
#> 30                                             Kinematic grid shift
#> 31                                          Denoyer Semi-Elliptical
#> 32                                                         Eckert I
#> 33                                                        Eckert II
#> 34                                                       Eckert III
#> 35                                                        Eckert IV
#> 36                                                         Eckert V
#> 37                                                        Eckert VI
#> 38                                                      Equal Earth
#> 39                           Equidistant Cylindrical (Plate Carree)
#> 40                                                Equidistant Conic
#> 41                                                            Euler
#> 42                                     Extended Transverse Mercator
#> 43                                                            Fahey
#> 44                                                          Foucaut
#> 45                                               Foucaut Sinusoidal
#> 46                                        Gall (Gall Stereographic)
#> 47                                              Geocentric Latitude
#> 48                                                Geographic Offset
#> 49                                     Geostationary Satellite View
#> 50                                        Ginsburg VIII (TsNIIGAiK)
#> 51                                        General Sinusoidal Series
#> 52                                                         Gnomonic
#> 53                                                 Goode Homolosine
#> 54                                Modified Stereographic of 48 U.S.
#> 55                                Modified Stereographic of 50 U.S.
#> 56                                                            Guyou
#> 57                                     Hammer & Eckert-Greifendorff
#> 58                                   Hatano Asymmetrical Equal Area
#> 59                                                          HEALPix
#> 60                                                         rHEALPix
#> 61                   3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 62                                            Horizontal grid shift
#> 63                                     Horner polynomial evaluation
#> 64                                     Interrupted Goode Homolosine
#> 65                        Interrupted Goode Homolosine Oceanic View
#> 66                                            Interrupted Mollweide
#> 67                               Interrupted Mollweide Oceanic View
#> 68                         International Map of the World Polyconic
#> 69                                    Icosahedral Snyder Equal Area
#> 70                                                     Kavrayskiy V
#> 71                                                   Kavrayskiy VII
#> 72                                                           Krovak
#> 73                                                          Laborde
#> 74                                     Lambert Azimuthal Equal Area
#> 75                                                         Lagrange
#> 76                                                         Larrivee
#> 77                                                        Laskowski
#> 78                                              Lat/long (Geodetic)
#> 79                                        Lat/long (Geodetic alias)
#> 80                                          Lambert Conformal Conic
#> 81                              Lambert Conformal Conic Alternative
#> 82                                         Lambert Equal Area Conic
#> 83                                        Lee Oblated Stereographic
#> 84                                                       Loximuthal
#> 85                                        Space oblique for LANDSAT
#> 86                           McBryde-Thomas Flat-Polar Sine (No. 1)
#> 87                            McBryde-Thomas Flat-Pole Sine (No. 2)
#> 88                              McBride-Thomas Flat-Polar Parabolic
#> 89                                McBryde-Thomas Flat-Polar Quartic
#> 90                             McBryde-Thomas Flat-Polar Sinusoidal
#> 91                                                         Mercator
#> 92                                     Miller Oblated Stereographic
#> 93                                               Miller Cylindrical
#> 94                                           Space oblique for MISR
#> 95                                                        Mollweide
#> 96                                Molodensky-Badekas transformation
#> 97                                             Molodensky transform
#> 98                                                        Murdoch I
#> 99                                                       Murdoch II
#> 100                                                     Murdoch III
#> 101                                                   Natural Earth
#> 102                                                 Natural Earth 2
#> 103                                                            Nell
#> 104                                                     Nell-Hammer
#> 105                                               Nicolosi Globular
#> 106                                          Near-sided perspective
#> 107                                            New Zealand Map Grid
#> 108                                                    No operation
#> 109                                  General Oblique Transformation
#> 110                                  Oblique Cylindrical Equal Area
#> 111                                              Oblated Equal Area
#> 112                                                Oblique Mercator
#> 113                                                   Ortelius Oval
#> 114                                                    Orthographic
#> 115                                               Perspective Conic
#> 116                                           Patterson Cylindrical
#> 117                                              Peirce Quincuncial
#> 118                                 Transformation pipeline manager
#> 119                                            Polyconic (American)
#> 120                   Retrieve coordinate value from pipeline stack
#> 121                         Save coordinate value on pipeline stack
#> 122                                                      Putnins P1
#> 123                                                      Putnins P2
#> 124                                                      Putnins P3
#> 125                                                     Putnins P3'
#> 126                                                     Putnins P4'
#> 127                                                      Putnins P5
#> 128                                                     Putnins P5'
#> 129                                                      Putnins P6
#> 130                                                     Putnins P6'
#> 131                                                Quartic Authalic
#> 132                                Quadrilateralized Spherical Cube
#> 133                                                        Robinson
#> 134                                         Roussilhe Stereographic
#> 135                                           Rectangular Polyconic
#> 136                                                              S2
#> 137                                    Spherical Cross-track Height
#> 138                                            Set coordinate value
#> 139                                   Sinusoidal (Sanson-Flamsteed)
#> 140                                            Swiss. Obl. Mercator
#> 141                                                   Stereographic
#> 142                               Oblique Stereographic Alternative
#> 143 Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 144                                  Transverse Central Cylindrical
#> 145                               Transverse Cylindrical Equal Area
#> 146                                                           Times
#> 147                              Triangulation based transformation
#> 148                                                          Tissot
#> 149                                             Transverse Mercator
#> 150                                                 Tobler-Mercator
#> 151                               Geocentric/Topocentric conversion
#> 152                                           Two Point Equidistant
#> 153                                              Tilted perspective
#> 154                                                 Unit conversion
#> 155                                   Universal Polar Stereographic
#> 156                                                        Urmaev V
#> 157                                    Urmaev Flat-Polar Sinusoidal
#> 158                             Universal Transverse Mercator (UTM)
#> 159                                             van der Grinten (I)
#> 160                                              van der Grinten II
#> 161                                             van der Grinten III
#> 162                                              van der Grinten IV
#> 163                                       Vertical Offset and Slope
#> 164                                                     Vitkovsky I
#> 165                                             Vertical grid shift
#> 166                                        Wagner I (Kavrayskiy VI)
#> 167                                                       Wagner II
#> 168                                                      Wagner III
#> 169                                                       Wagner IV
#> 170                                                        Wagner V
#> 171                                                       Wagner VI
#> 172                                                      Wagner VII
#> 173                                  Web Mercator / Pseudo Mercator
#> 174                                                   Werenskiold I
#> 175                                                        Winkel I
#> 176                                                       Winkel II
#> 177                                                   Winkel Tripel
#> 178                                           Geocentric grid shift
projInfo("ellps")
#>         name          major                  ell
#> 1      MERIT    a=6378137.0           rf=298.257
#> 2      SGS85    a=6378136.0           rf=298.257
#> 3      GRS80    a=6378137.0     rf=298.257222101
#> 4      IAU76    a=6378140.0           rf=298.257
#> 5       airy  a=6377563.396       rf=299.3249646
#> 6     APL4.9    a=6378137.0            rf=298.25
#> 7      NWL9D    a=6378145.0            rf=298.25
#> 8   mod_airy  a=6377340.189        b=6356034.446
#> 9     andrae   a=6377104.43             rf=300.0
#> 10    danish a=6377019.2563             rf=300.0
#> 11   aust_SA    a=6378160.0            rf=298.25
#> 12     GRS67    a=6378160.0    rf=298.2471674270
#> 13   GSK2011    a=6378136.5       rf=298.2564151
#> 14    bessel  a=6377397.155       rf=299.1528128
#> 15  bess_nam  a=6377483.865       rf=299.1528128
#> 16    clrk66    a=6378206.4          b=6356583.8
#> 17    clrk80  a=6378249.145          rf=293.4663
#> 18 clrk80ign    a=6378249.2 rf=293.4660212936269
#> 19       CPM    a=6375738.7            rf=334.29
#> 20    delmbr     a=6376428.             rf=311.5
#> 21   engelis   a=6378136.05          rf=298.2566
#> 22   evrst30  a=6377276.345          rf=300.8017
#> 23   evrst48  a=6377304.063          rf=300.8017
#> 24   evrst56  a=6377301.243          rf=300.8017
#> 25   evrst69  a=6377295.664          rf=300.8017
#> 26   evrstSS  a=6377298.556          rf=300.8017
#> 27   fschr60     a=6378166.             rf=298.3
#> 28  fschr60m     a=6378155.             rf=298.3
#> 29   fschr68     a=6378150.             rf=298.3
#> 30   helmert     a=6378200.             rf=298.3
#> 31     hough    a=6378270.0              rf=297.
#> 32      intl    a=6378388.0              rf=297.
#> 33     krass    a=6378245.0             rf=298.3
#> 34     kaula     a=6378163.            rf=298.24
#> 35     lerch     a=6378139.           rf=298.257
#> 36     mprts     a=6397300.              rf=191.
#> 37  new_intl    a=6378157.5          b=6356772.2
#> 38   plessis     a=6376523.           b=6355863.
#> 39      PZ90    a=6378136.0         rf=298.25784
#> 40    SEasia    a=6378155.0       b=6356773.3205
#> 41   walbeck    a=6376896.0       b=6355834.8467
#> 42     WGS60    a=6378165.0             rf=298.3
#> 43     WGS66    a=6378145.0            rf=298.25
#> 44     WGS72    a=6378135.0            rf=298.26
#> 45     WGS84    a=6378137.0     rf=298.257223563
#> 46    sphere    a=6370997.0          b=6370997.0
#>                                description
#> 1                               MERIT 1983
#> 2                Soviet Geodetic System 85
#> 3                     GRS 1980(IUGG, 1980)
#> 4                                 IAU 1976
#> 5                                Airy 1830
#> 6                      Appl. Physics. 1965
#> 7                 Naval Weapons Lab., 1965
#> 8                            Modified Airy
#> 9               Andrae 1876 (Den., Iclnd.)
#> 10          Andrae 1876 (Denmark, Iceland)
#> 11         Australian Natl & S. Amer. 1969
#> 12                       GRS 67(IUGG 1967)
#> 13                                GSK-2011
#> 14                             Bessel 1841
#> 15                   Bessel 1841 (Namibia)
#> 16                             Clarke 1866
#> 17                        Clarke 1880 mod.
#> 18                      Clarke 1880 (IGN).
#> 19         Comm. des Poids et Mesures 1799
#> 20                 Delambre 1810 (Belgium)
#> 21                            Engelis 1985
#> 22                            Everest 1830
#> 23                            Everest 1948
#> 24                            Everest 1956
#> 25                            Everest 1969
#> 26               Everest (Sabah & Sarawak)
#> 27            Fischer (Mercury Datum) 1960
#> 28                   Modified Fischer 1960
#> 29                            Fischer 1968
#> 30                            Helmert 1906
#> 31                                   Hough
#> 32 International 1924 (Hayford 1909, 1910)
#> 33                        Krassovsky, 1942
#> 34                              Kaula 1961
#> 35                              Lerch 1979
#> 36                         Maupertius 1738
#> 37                  New International 1967
#> 38                   Plessis 1817 (France)
#> 39                                   PZ-90
#> 40                          Southeast Asia
#> 41                                 Walbeck
#> 42                                  WGS 60
#> 43                                  WGS 66
#> 44                                  WGS 72
#> 45                                  WGS 84
#> 46               Normal Sphere (r=6370997)
projInfo("units")
#>        id     to_meter               name
#> 1      mm 1.000000e-03         millimetre
#> 2      cm 1.000000e-02         centimetre
#> 3       m 1.000000e+00              metre
#> 4      ft 3.048000e-01               foot
#> 5   us-ft 3.048006e-01     US survey foot
#> 6    fath 1.828800e+00             fathom
#> 7     kmi 1.852000e+03      nautical mile
#> 8   us-ch 2.011684e+01    US survey chain
#> 9   us-mi 1.609347e+03     US survey mile
#> 10     km 1.000000e+03          kilometre
#> 11 ind-ft 3.047984e-01 Indian foot (1937)
#> 12 ind-yd 9.143952e-01 Indian yard (1937)
#> 13     mi 1.609344e+03       Statute mile
#> 14     yd 9.144000e-01               yard
#> 15     ch 2.011680e+01              chain
#> 16   link 2.011680e-01               link
#> 17     dm 1.000000e-02          decimeter
#> 18     in 2.540000e-02               inch
#> 19 ind-ch 2.011670e+01       Indian chain
#> 20  us-in 2.540005e-02     US survey inch
#> 21  us-yd 9.144018e-01     US survey yard