List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

http://proj.maptools.org/

Note

Loading the rgdal package changes the PROJ\_LIB environmental variable to the PROJ.4 support files bundles with the package.

Examples

getPROJ4VersionInfo()

#> [1] "Rel. 5.2.0, September 15th, 2018, [PJ_VERSION: 520]"

projInfo()

#>            name                                                     description
#> 1           aea                                               Albers Equal Area
#> 2          aeqd                                           Azimuthal Equidistant
#> 3          airy                                                            Airy
#> 4        aitoff                                                          Aitoff
#> 5          alsk                                    Mod. Stereographic of Alaska
#> 6         apian                                                Apian Globular I
#> 7        august                                             August Epicycloidal
#> 8      axisswap                                                   Axis ordering
#> 9         bacon                                                  Bacon Globular
#> 10         bipc                             Bipolar conic of western hemisphere
#> 11        boggs                                                 Boggs Eumorphic
#> 12        bonne                                         Bonne (Werner lat_1=90)
#> 13      calcofi                       Cal Coop Ocean Fish Invest Lines/Stations
#> 14         cart                                  Geodetic/cartesian conversions
#> 15         cass                                                         Cassini
#> 16           cc                                             Central Cylindrical
#> 17         ccon                                                   Central Conic
#> 18          cea                                          Equal Area Cylindrical
#> 19        chamb                                            Chamberlin Trimetric
#> 20        collg                                                       Collignon
#> 21       comill                                                  Compact Miller
#> 22        crast                                  Craster Parabolic (Putnins P4)
#> 23  deformation                                            Kinematic grid shift
#> 24        denoy                                         Denoyer Semi-Elliptical
#> 25         eck1                                                        Eckert I
#> 26         eck2                                                       Eckert II
#> 27         eck3                                                      Eckert III
#> 28         eck4                                                       Eckert IV
#> 29         eck5                                                        Eckert V
#> 30         eck6                                                       Eckert VI
#> 31      eqearth                                                     Equal Earth
#> 32          eqc                           Equidistant Cylindrical (Plate Caree)
#> 33         eqdc                                               Equidistant Conic
#> 34        euler                                                           Euler
#> 35       etmerc                                    Extended Transverse Mercator
#> 36        fahey                                                           Fahey
#> 37         fouc                                                         Foucaut
#> 38       fouc_s                                              Foucaut Sinusoidal
#> 39         gall                                       Gall (Gall Stereographic)
#> 40         geoc                                             Geocentric Latitude
#> 41      geocent                                                      Geocentric
#> 42         geos                                    Geostationary Satellite View
#> 43        gins8                                       Ginsburg VIII (TsNIIGAiK)
#> 44      gn_sinu                                       General Sinusoidal Series
#> 45         gnom                                                        Gnomonic
#> 46        goode                                                Goode Homolosine
#> 47         gs48                                   Mod. Stereographic of 48 U.S.
#> 48         gs50                                   Mod. Stereographic of 50 U.S.
#> 49       hammer                                    Hammer & Eckert-Greifendorff
#> 50       hatano                                  Hatano Asymmetrical Equal Area
#> 51      healpix                                                         HEALPix
#> 52     rhealpix                                                        rHEALPix
#> 53      helmert                  3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 54   hgridshift                                           Horizontal grid shift
#> 55       horner                                    Horner polynomial evaluation
#> 56          igh                                    Interrupted Goode Homolosine
#> 57        imw_p                        International Map of the World Polyconic
#> 58         isea                                   Icosahedral Snyder Equal Area
#> 59         kav5                                                     Kavraisky V
#> 60         kav7                                                   Kavraisky VII
#> 61       krovak                                                          Krovak
#> 62        labrd                                                         Laborde
#> 63         laea                                    Lambert Azimuthal Equal Area
#> 64       lagrng                                                        Lagrange
#> 65         larr                                                        Larrivee
#> 66         lask                                                       Laskowski
#> 67       lonlat                                             Lat/long (Geodetic)
#> 68       latlon                                       Lat/long (Geodetic alias)
#> 69      latlong                                       Lat/long (Geodetic alias)
#> 70      longlat                                       Lat/long (Geodetic alias)
#> 71          lcc                                         Lambert Conformal Conic
#> 72         lcca                             Lambert Conformal Conic Alternative
#> 73         leac                                        Lambert Equal Area Conic
#> 74       lee_os                                       Lee Oblated Stereographic
#> 75        loxim                                                      Loximuthal
#> 76         lsat                                       Space oblique for LANDSAT
#> 77        mbt_s                          McBryde-Thomas Flat-Polar Sine (No. 1)
#> 78      mbt_fps                           McBryde-Thomas Flat-Pole Sine (No. 2)
#> 79       mbtfpp                             McBride-Thomas Flat-Polar Parabolic
#> 80       mbtfpq                               McBryde-Thomas Flat-Polar Quartic
#> 81       mbtfps                            McBryde-Thomas Flat-Polar Sinusoidal
#> 82         merc                                                        Mercator
#> 83       mil_os                                    Miller Oblated Stereographic
#> 84         mill                                              Miller Cylindrical
#> 85      misrsom                                          Space oblique for MISR
#> 86         moll                                                       Mollweide
#> 87   molodensky                                            Molodensky transform
#> 88        murd1                                                       Murdoch I
#> 89        murd2                                                      Murdoch II
#> 90        murd3                                                     Murdoch III
#> 91     natearth                                                   Natural Earth
#> 92    natearth2                                                 Natural Earth 2
#> 93         nell                                                            Nell
#> 94       nell_h                                                     Nell-Hammer
#> 95        nicol                                               Nicolosi Globular
#> 96        nsper                                          Near-sided perspective
#> 97         nzmg                                            New Zealand Map Grid
#> 98      ob_tran                                  General Oblique Transformation
#> 99         ocea                                  Oblique Cylindrical Equal Area
#> 100         oea                                              Oblated Equal Area
#> 101       omerc                                                Oblique Mercator
#> 102       ortel                                                   Ortelius Oval
#> 103       ortho                                                    Orthographic
#> 104      pconic                                               Perspective Conic
#> 105   patterson                                           Patterson Cylindrical
#> 106    pipeline                                 Transformation pipeline manager
#> 107        poly                                            Polyconic (American)
#> 108       putp1                                                      Putnins P1
#> 109       putp2                                                      Putnins P2
#> 110       putp3                                                      Putnins P3
#> 111      putp3p                                                     Putnins P3'
#> 112      putp4p                                                     Putnins P4'
#> 113       putp5                                                      Putnins P5
#> 114      putp5p                                                     Putnins P5'
#> 115       putp6                                                      Putnins P6
#> 116      putp6p                                                     Putnins P6'
#> 117     qua_aut                                                Quartic Authalic
#> 118         qsc                                Quadrilateralized Spherical Cube
#> 119       robin                                                        Robinson
#> 120       rouss                                         Roussilhe Stereographic
#> 121       rpoly                                           Rectangular Polyconic
#> 122         sch                                    Spherical Cross-track Height
#> 123        sinu                                   Sinusoidal (Sanson-Flamsteed)
#> 124      somerc                                            Swiss. Obl. Mercator
#> 125       stere                                                   Stereographic
#> 126      sterea                               Oblique Stereographic Alternative
#> 127     gstmerc Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 128         tcc                                  Transverse Central Cylindrical
#> 129        tcea                               Transverse Cylindrical Equal Area
#> 130       times                                                           Times
#> 131      tissot                                                          Tissot
#> 132       tmerc                                             Transverse Mercator
#> 133       tpeqd                                           Two Point Equidistant
#> 134       tpers                                              Tilted perspective
#> 135 unitconvert                                                 Unit conversion
#> 136         ups                                   Universal Polar Stereographic
#> 137        urm5                                                        Urmaev V
#> 138      urmfps                                    Urmaev Flat-Polar Sinusoidal
#> 139         utm                             Universal Transverse Mercator (UTM)
#> 140       vandg                                             van der Grinten (I)
#> 141      vandg2                                              van der Grinten II
#> 142      vandg3                                             van der Grinten III
#> 143      vandg4                                              van der Grinten IV
#> 144       vitk1                                                     Vitkovsky I
#> 145  vgridshift                                             Vertical grid shift
#> 146        wag1                                         Wagner I (Kavraisky VI)
#> 147        wag2                                                       Wagner II
#> 148        wag3                                                      Wagner III
#> 149        wag4                                                       Wagner IV
#> 150        wag5                                                        Wagner V
#> 151        wag6                                                       Wagner VI
#> 152        wag7                                                      Wagner VII
#> 153     webmerc                                  Web Mercator / Pseudo Mercator
#> 154       weren                                                   Werenskiold I
#> 155       wink1                                                        Winkel I
#> 156       wink2                                                       Winkel II
#> 157      wintri                                                   Winkel Tripel