List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()
GDAL_OSR_PROJ()
GDALis3ormore()
PROJis6ormore()
new_proj_and_gdal()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

https://proj.org/

Note

Loading the rgdal package changes the PROJ\_LIB environmental variable to the PROJ.4 support files bundles with the package.

Examples

getPROJ4VersionInfo()

#> [1] "Rel. 6.2.1, November 1st, 2019, [PJ_VERSION: 621]"

projInfo()

#>            name                                                     description
#> 1           aea                                               Albers Equal Area
#> 2          aeqd                                           Azimuthal Equidistant
#> 3        affine                                           Affine transformation
#> 4          airy                                                            Airy
#> 5        aitoff                                                          Aitoff
#> 6          alsk                                    Mod. Stereographic of Alaska
#> 7         apian                                                Apian Globular I
#> 8        august                                             August Epicycloidal
#> 9      axisswap                                                   Axis ordering
#> 10        bacon                                                  Bacon Globular
#> 11   bertin1953                                                     Bertin 1953
#> 12         bipc                             Bipolar conic of western hemisphere
#> 13        boggs                                                 Boggs Eumorphic
#> 14        bonne                                         Bonne (Werner lat_1=90)
#> 15      calcofi                       Cal Coop Ocean Fish Invest Lines/Stations
#> 16         cart                                  Geodetic/cartesian conversions
#> 17         cass                                                         Cassini
#> 18           cc                                             Central Cylindrical
#> 19         ccon                                                   Central Conic
#> 20          cea                                          Equal Area Cylindrical
#> 21        chamb                                            Chamberlin Trimetric
#> 22        collg                                                       Collignon
#> 23       comill                                                  Compact Miller
#> 24        crast                                  Craster Parabolic (Putnins P4)
#> 25  deformation                                            Kinematic grid shift
#> 26        denoy                                         Denoyer Semi-Elliptical
#> 27         eck1                                                        Eckert I
#> 28         eck2                                                       Eckert II
#> 29         eck3                                                      Eckert III
#> 30         eck4                                                       Eckert IV
#> 31         eck5                                                        Eckert V
#> 32         eck6                                                       Eckert VI
#> 33      eqearth                                                     Equal Earth
#> 34          eqc                          Equidistant Cylindrical (Plate Carree)
#> 35         eqdc                                               Equidistant Conic
#> 36        euler                                                           Euler
#> 37       etmerc                                    Extended Transverse Mercator
#> 38        fahey                                                           Fahey
#> 39         fouc                                                         Foucaut
#> 40       fouc_s                                              Foucaut Sinusoidal
#> 41         gall                                       Gall (Gall Stereographic)
#> 42         geoc                                             Geocentric Latitude
#> 43   geogoffset                                               Geographic Offset
#> 44         geos                                    Geostationary Satellite View
#> 45        gins8                                       Ginsburg VIII (TsNIIGAiK)
#> 46      gn_sinu                                       General Sinusoidal Series
#> 47         gnom                                                        Gnomonic
#> 48        goode                                                Goode Homolosine
#> 49         gs48                                   Mod. Stereographic of 48 U.S.
#> 50         gs50                                   Mod. Stereographic of 50 U.S.
#> 51       hammer                                    Hammer & Eckert-Greifendorff
#> 52       hatano                                  Hatano Asymmetrical Equal Area
#> 53      healpix                                                      HEAPJ_LPix
#> 54     rhealpix                                                     rHEAPJ_LPix
#> 55      helmert                  3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 56   hgridshift                                           Horizontal grid shift
#> 57       horner                                    Horner polynomial evaluation
#> 58          igh                                    Interrupted Goode Homolosine
#> 59        imw_p                        International Map of the World Polyconic
#> 60         isea                                   Icosahedral Snyder Equal Area
#> 61         kav5                                                     Kavraisky V
#> 62         kav7                                                   Kavraisky VII
#> 63       krovak                                                          Krovak
#> 64        labrd                                                         Laborde
#> 65         laea                                    Lambert Azimuthal Equal Area
#> 66       lagrng                                                        Lagrange
#> 67         larr                                                        Larrivee
#> 68         lask                                                       Laskowski
#> 69       lonlat                                             Lat/long (Geodetic)
#> 70       latlon                                       Lat/long (Geodetic alias)
#> 71          lcc                                         Lambert Conformal Conic
#> 72         lcca                             Lambert Conformal Conic Alternative
#> 73         leac                                        Lambert Equal Area Conic
#> 74       lee_os                                       Lee Oblated Stereographic
#> 75        loxim                                                      Loximuthal
#> 76         lsat                                       Space oblique for LANDSAT
#> 77        mbt_s                          McBryde-Thomas Flat-Polar Sine (No. 1)
#> 78      mbt_fps                           McBryde-Thomas Flat-Pole Sine (No. 2)
#> 79       mbtfpp                             McBride-Thomas Flat-Polar Parabolic
#> 80       mbtfpq                               McBryde-Thomas Flat-Polar Quartic
#> 81       mbtfps                            McBryde-Thomas Flat-Polar Sinusoidal
#> 82         merc                                                        Mercator
#> 83       mil_os                                    Miller Oblated Stereographic
#> 84         mill                                              Miller Cylindrical
#> 85      misrsom                                          Space oblique for MISR
#> 86         moll                                                       Mollweide
#> 87  molobadekas                               Molodensky-Badekas transformation
#> 88   molodensky                                            Molodensky transform
#> 89        murd1                                                       Murdoch I
#> 90        murd2                                                      Murdoch II
#> 91        murd3                                                     Murdoch III
#> 92     natearth                                                   Natural Earth
#> 93    natearth2                                                 Natural Earth 2
#> 94         nell                                                            Nell
#> 95       nell_h                                                     Nell-Hammer
#> 96        nicol                                               Nicolosi Globular
#> 97        nsper                                          Near-sided perspective
#> 98         nzmg                                            New Zealand Map Grid
#> 99         noop                                                    No operation
#> 100     ob_tran                                  General Oblique Transformation
#> 101        ocea                                  Oblique Cylindrical Equal Area
#> 102         oea                                              Oblated Equal Area
#> 103       omerc                                                Oblique Mercator
#> 104       ortel                                                   Ortelius Oval
#> 105       ortho                                                    Orthographic
#> 106      pconic                                               Perspective Conic
#> 107   patterson                                           Patterson Cylindrical
#> 108    pipeline                                 Transformation pipeline manager
#> 109        poly                                            Polyconic (American)
#> 110         pop                   Retrieve coordinate value from pipeline stack
#> 111        push                         Save coordinate value on pipeline stack
#> 112       putp1                                                      Putnins P1
#> 113       putp2                                                      Putnins P2
#> 114       putp3                                                      Putnins P3
#> 115      putp3p                                                     Putnins P3'
#> 116      putp4p                                                     Putnins P4'
#> 117       putp5                                                      Putnins P5
#> 118      putp5p                                                     Putnins P5'
#> 119       putp6                                                      Putnins P6
#> 120      putp6p                                                     Putnins P6'
#> 121     qua_aut                                                Quartic Authalic
#> 122         qsc                                Quadrilateralized Spherical Cube
#> 123       robin                                                        Robinson
#> 124       rouss                                         Roussilhe Stereographic
#> 125       rpoly                                           Rectangular Polyconic
#> 126         sch                                    Spherical Cross-track Height
#> 127        sinu                                   Sinusoidal (Sanson-Flamsteed)
#> 128      somerc                                            Swiss. Obl. Mercator
#> 129       stere                                                   Stereographic
#> 130      sterea                               Oblique Stereographic Alternative
#> 131     gstmerc Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 132         tcc                                  Transverse Central Cylindrical
#> 133        tcea                               Transverse Cylindrical Equal Area
#> 134       times                                                           Times
#> 135      tissot                                                          Tissot
#> 136       tmerc                                             Transverse Mercator
#> 137     tobmerc                                                 Tobler-Mercator
#> 138       tpeqd                                           Two Point Equidistant
#> 139       tpers                                              Tilted perspective
#> 140 unitconvert                                                 Unit conversion
#> 141         ups                                   Universal Polar Stereographic
#> 142        urm5                                                        Urmaev V
#> 143      urmfps                                    Urmaev Flat-Polar Sinusoidal
#> 144         utm                             Universal Transverse Mercator (UTM)
#> 145       vandg                                             van der Grinten (I)
#> 146      vandg2                                              van der Grinten II
#> 147      vandg3                                             van der Grinten III
#> 148      vandg4                                              van der Grinten IV
#> 149       vitk1                                                     Vitkovsky I
#> 150  vgridshift                                             Vertical grid shift
#> 151        wag1                                         Wagner I (Kavraisky VI)
#> 152        wag2                                                       Wagner II
#> 153        wag3                                                      Wagner III
#> 154        wag4                                                       Wagner IV
#> 155        wag5                                                        Wagner V
#> 156        wag6                                                       Wagner VI
#> 157        wag7                                                      Wagner VII
#> 158     webmerc                                  Web Mercator / Pseudo Mercator
#> 159       weren                                                   Werenskiold I
#> 160       wink1                                                        Winkel I
#> 161       wink2                                                       Winkel II
#> 162      wintri                                                   Winkel Tripel