List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()
GDAL_OSR_PROJ()
GDALis3ormore()
PROJis6ormore()
new_proj_and_gdal()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu. From PROJ 6, "datum" is not available. From PROJ 7.1.0, "units" returns the conversion factor as numeric, not character.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

https://proj.org/

Note

Loading the rgdal package may change the PROJ\_LIB environmental variable to the PROJ.4 support files if bundled with binary packages.

Examples

getPROJ4VersionInfo()

#> [1] "Rel. 7.1.0, August 1st, 2020, [PJ_VERSION: 710]"
#> attr(,"short")
#> [1] 710

projInfo()

#>             name
#> 1     adams_hemi
#> 2      adams_ws1
#> 3      adams_ws2
#> 4            aea
#> 5           aeqd
#> 6         affine
#> 7           airy
#> 8         aitoff
#> 9           alsk
#> 10         apian
#> 11        august
#> 12      axisswap
#> 13         bacon
#> 14    bertin1953
#> 15          bipc
#> 16         boggs
#> 17         bonne
#> 18       calcofi
#> 19          cart
#> 20          cass
#> 21            cc
#> 22          ccon
#> 23           cea
#> 24         chamb
#> 25         collg
#> 26        comill
#> 27         crast
#> 28      defmodel
#> 29   deformation
#> 30         denoy
#> 31          eck1
#> 32          eck2
#> 33          eck3
#> 34          eck4
#> 35          eck5
#> 36          eck6
#> 37       eqearth
#> 38           eqc
#> 39          eqdc
#> 40         euler
#> 41        etmerc
#> 42         fahey
#> 43          fouc
#> 44        fouc_s
#> 45          gall
#> 46          geoc
#> 47    geogoffset
#> 48          geos
#> 49         gins8
#> 50       gn_sinu
#> 51          gnom
#> 52         goode
#> 53          gs48
#> 54          gs50
#> 55         guyou
#> 56        hammer
#> 57        hatano
#> 58       healpix
#> 59      rhealpix
#> 60       helmert
#> 61    hgridshift
#> 62        horner
#> 63           igh
#> 64         igh_o
#> 65         imw_p
#> 66          isea
#> 67          kav5
#> 68          kav7
#> 69        krovak
#> 70         labrd
#> 71          laea
#> 72        lagrng
#> 73          larr
#> 74          lask
#> 75        lonlat
#> 76        latlon
#> 77           lcc
#> 78          lcca
#> 79          leac
#> 80        lee_os
#> 81         loxim
#> 82          lsat
#> 83         mbt_s
#> 84       mbt_fps
#> 85        mbtfpp
#> 86        mbtfpq
#> 87        mbtfps
#> 88          merc
#> 89        mil_os
#> 90          mill
#> 91       misrsom
#> 92          moll
#> 93   molobadekas
#> 94    molodensky
#> 95         murd1
#> 96         murd2
#> 97         murd3
#> 98      natearth
#> 99     natearth2
#> 100         nell
#> 101       nell_h
#> 102        nicol
#> 103        nsper
#> 104         nzmg
#> 105         noop
#> 106      ob_tran
#> 107         ocea
#> 108          oea
#> 109        omerc
#> 110        ortel
#> 111        ortho
#> 112       pconic
#> 113    patterson
#> 114     peirce_q
#> 115     pipeline
#> 116         poly
#> 117          pop
#> 118         push
#> 119        putp1
#> 120        putp2
#> 121        putp3
#> 122       putp3p
#> 123       putp4p
#> 124        putp5
#> 125       putp5p
#> 126        putp6
#> 127       putp6p
#> 128      qua_aut
#> 129          qsc
#> 130        robin
#> 131        rouss
#> 132        rpoly
#> 133          sch
#> 134          set
#> 135         sinu
#> 136       somerc
#> 137        stere
#> 138       sterea
#> 139      gstmerc
#> 140          tcc
#> 141         tcea
#> 142        times
#> 143       tissot
#> 144        tmerc
#> 145      tobmerc
#> 146        tpeqd
#> 147        tpers
#> 148  unitconvert
#> 149          ups
#> 150         urm5
#> 151       urmfps
#> 152          utm
#> 153        vandg
#> 154       vandg2
#> 155       vandg3
#> 156       vandg4
#> 157        vitk1
#> 158   vgridshift
#> 159         wag1
#> 160         wag2
#> 161         wag3
#> 162         wag4
#> 163         wag5
#> 164         wag6
#> 165         wag7
#> 166      webmerc
#> 167        weren
#> 168        wink1
#> 169        wink2
#> 170       wintri
#> 171 xyzgridshift
#>                                                         description
#> 1                                      Adams Hemisphere in a Square
#> 2                                         Adams World in a Square I
#> 3                                        Adams World in a Square II
#> 4                                                 Albers Equal Area
#> 5                                             Azimuthal Equidistant
#> 6                                             Affine transformation
#> 7                                                              Airy
#> 8                                                            Aitoff
#> 9                                      Mod. Stereographic of Alaska
#> 10                                                 Apian Globular I
#> 11                                              August Epicycloidal
#> 12                                                    Axis ordering
#> 13                                                   Bacon Globular
#> 14                                                      Bertin 1953
#> 15                              Bipolar conic of western hemisphere
#> 16                                                  Boggs Eumorphic
#> 17                                          Bonne (Werner lat_1=90)
#> 18                        Cal Coop Ocean Fish Invest Lines/Stations
#> 19                                   Geodetic/cartesian conversions
#> 20                                                          Cassini
#> 21                                              Central Cylindrical
#> 22                                                    Central Conic
#> 23                                           Equal Area Cylindrical
#> 24                                             Chamberlin Trimetric
#> 25                                                        Collignon
#> 26                                                   Compact Miller
#> 27                                   Craster Parabolic (Putnins P4)
#> 28                                                Deformation model
#> 29                                             Kinematic grid shift
#> 30                                          Denoyer Semi-Elliptical
#> 31                                                         Eckert I
#> 32                                                        Eckert II
#> 33                                                       Eckert III
#> 34                                                        Eckert IV
#> 35                                                         Eckert V
#> 36                                                        Eckert VI
#> 37                                                      Equal Earth
#> 38                           Equidistant Cylindrical (Plate Carree)
#> 39                                                Equidistant Conic
#> 40                                                            Euler
#> 41                                     Extended Transverse Mercator
#> 42                                                            Fahey
#> 43                                                          Foucaut
#> 44                                               Foucaut Sinusoidal
#> 45                                        Gall (Gall Stereographic)
#> 46                                              Geocentric Latitude
#> 47                                                Geographic Offset
#> 48                                     Geostationary Satellite View
#> 49                                        Ginsburg VIII (TsNIIGAiK)
#> 50                                        General Sinusoidal Series
#> 51                                                         Gnomonic
#> 52                                                 Goode Homolosine
#> 53                                    Mod. Stereographic of 48 U.S.
#> 54                                    Mod. Stereographic of 50 U.S.
#> 55                                                            Guyou
#> 56                                     Hammer & Eckert-Greifendorff
#> 57                                   Hatano Asymmetrical Equal Area
#> 58                                                          HEALPix
#> 59                                                         rHEALPix
#> 60                   3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 61                                            Horizontal grid shift
#> 62                                     Horner polynomial evaluation
#> 63                                     Interrupted Goode Homolosine
#> 64                        Interrupted Goode Homolosine Oceanic View
#> 65                         International Map of the World Polyconic
#> 66                                    Icosahedral Snyder Equal Area
#> 67                                                      Kavraisky V
#> 68                                                    Kavraisky VII
#> 69                                                           Krovak
#> 70                                                          Laborde
#> 71                                     Lambert Azimuthal Equal Area
#> 72                                                         Lagrange
#> 73                                                         Larrivee
#> 74                                                        Laskowski
#> 75                                              Lat/long (Geodetic)
#> 76                                        Lat/long (Geodetic alias)
#> 77                                          Lambert Conformal Conic
#> 78                              Lambert Conformal Conic Alternative
#> 79                                         Lambert Equal Area Conic
#> 80                                        Lee Oblated Stereographic
#> 81                                                       Loximuthal
#> 82                                        Space oblique for LANDSAT
#> 83                           McBryde-Thomas Flat-Polar Sine (No. 1)
#> 84                            McBryde-Thomas Flat-Pole Sine (No. 2)
#> 85                              McBride-Thomas Flat-Polar Parabolic
#> 86                                McBryde-Thomas Flat-Polar Quartic
#> 87                             McBryde-Thomas Flat-Polar Sinusoidal
#> 88                                                         Mercator
#> 89                                     Miller Oblated Stereographic
#> 90                                               Miller Cylindrical
#> 91                                           Space oblique for MISR
#> 92                                                        Mollweide
#> 93                                Molodensky-Badekas transformation
#> 94                                             Molodensky transform
#> 95                                                        Murdoch I
#> 96                                                       Murdoch II
#> 97                                                      Murdoch III
#> 98                                                    Natural Earth
#> 99                                                  Natural Earth 2
#> 100                                                            Nell
#> 101                                                     Nell-Hammer
#> 102                                               Nicolosi Globular
#> 103                                          Near-sided perspective
#> 104                                            New Zealand Map Grid
#> 105                                                    No operation
#> 106                                  General Oblique Transformation
#> 107                                  Oblique Cylindrical Equal Area
#> 108                                              Oblated Equal Area
#> 109                                                Oblique Mercator
#> 110                                                   Ortelius Oval
#> 111                                                    Orthographic
#> 112                                               Perspective Conic
#> 113                                           Patterson Cylindrical
#> 114                                              Peirce Quincuncial
#> 115                                 Transformation pipeline manager
#> 116                                            Polyconic (American)
#> 117                   Retrieve coordinate value from pipeline stack
#> 118                         Save coordinate value on pipeline stack
#> 119                                                      Putnins P1
#> 120                                                      Putnins P2
#> 121                                                      Putnins P3
#> 122                                                     Putnins P3'
#> 123                                                     Putnins P4'
#> 124                                                      Putnins P5
#> 125                                                     Putnins P5'
#> 126                                                      Putnins P6
#> 127                                                     Putnins P6'
#> 128                                                Quartic Authalic
#> 129                                Quadrilateralized Spherical Cube
#> 130                                                        Robinson
#> 131                                         Roussilhe Stereographic
#> 132                                           Rectangular Polyconic
#> 133                                    Spherical Cross-track Height
#> 134                                            Set coordinate value
#> 135                                   Sinusoidal (Sanson-Flamsteed)
#> 136                                            Swiss. Obl. Mercator
#> 137                                                   Stereographic
#> 138                               Oblique Stereographic Alternative
#> 139 Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 140                                  Transverse Central Cylindrical
#> 141                               Transverse Cylindrical Equal Area
#> 142                                                           Times
#> 143                                                          Tissot
#> 144                                             Transverse Mercator
#> 145                                                 Tobler-Mercator
#> 146                                           Two Point Equidistant
#> 147                                              Tilted perspective
#> 148                                                 Unit conversion
#> 149                                   Universal Polar Stereographic
#> 150                                                        Urmaev V
#> 151                                    Urmaev Flat-Polar Sinusoidal
#> 152                             Universal Transverse Mercator (UTM)
#> 153                                             van der Grinten (I)
#> 154                                              van der Grinten II
#> 155                                             van der Grinten III
#> 156                                              van der Grinten IV
#> 157                                                     Vitkovsky I
#> 158                                             Vertical grid shift
#> 159                                         Wagner I (Kavraisky VI)
#> 160                                                       Wagner II
#> 161                                                      Wagner III
#> 162                                                       Wagner IV
#> 163                                                        Wagner V
#> 164                                                       Wagner VI
#> 165                                                      Wagner VII
#> 166                                  Web Mercator / Pseudo Mercator
#> 167                                                   Werenskiold I
#> 168                                                        Winkel I
#> 169                                                       Winkel II
#> 170                                                   Winkel Tripel
#> 171                                           Geocentric grid shift

projInfo("ellps")

#>         name          major                  ell
#> 1      MERIT    a=6378137.0           rf=298.257
#> 2      SGS85    a=6378136.0           rf=298.257
#> 3      GRS80    a=6378137.0     rf=298.257222101
#> 4      IAU76    a=6378140.0           rf=298.257
#> 5       airy  a=6377563.396       rf=299.3249646
#> 6     APL4.9    a=6378137.0            rf=298.25
#> 7      NWL9D    a=6378145.0            rf=298.25
#> 8   mod_airy  a=6377340.189        b=6356034.446
#> 9     andrae   a=6377104.43             rf=300.0
#> 10    danish a=6377019.2563             rf=300.0
#> 11   aust_SA    a=6378160.0            rf=298.25
#> 12     GRS67    a=6378160.0    rf=298.2471674270
#> 13   GSK2011    a=6378136.5       rf=298.2564151
#> 14    bessel  a=6377397.155       rf=299.1528128
#> 15  bess_nam  a=6377483.865       rf=299.1528128
#> 16    clrk66    a=6378206.4          b=6356583.8
#> 17    clrk80  a=6378249.145          rf=293.4663
#> 18 clrk80ign    a=6378249.2 rf=293.4660212936269
#> 19       CPM    a=6375738.7            rf=334.29
#> 20    delmbr     a=6376428.             rf=311.5
#> 21   engelis   a=6378136.05          rf=298.2566
#> 22   evrst30  a=6377276.345          rf=300.8017
#> 23   evrst48  a=6377304.063          rf=300.8017
#> 24   evrst56  a=6377301.243          rf=300.8017
#> 25   evrst69  a=6377295.664          rf=300.8017
#> 26   evrstSS  a=6377298.556          rf=300.8017
#> 27   fschr60     a=6378166.             rf=298.3
#> 28  fschr60m     a=6378155.             rf=298.3
#> 29   fschr68     a=6378150.             rf=298.3
#> 30   helmert     a=6378200.             rf=298.3
#> 31     hough    a=6378270.0              rf=297.
#> 32      intl    a=6378388.0              rf=297.
#> 33     krass    a=6378245.0             rf=298.3
#> 34     kaula     a=6378163.            rf=298.24
#> 35     lerch     a=6378139.           rf=298.257
#> 36     mprts     a=6397300.              rf=191.
#> 37  new_intl    a=6378157.5          b=6356772.2
#> 38   plessis     a=6376523.           b=6355863.
#> 39      PZ90    a=6378136.0         rf=298.25784
#> 40    SEasia    a=6378155.0       b=6356773.3205
#> 41   walbeck    a=6376896.0       b=6355834.8467
#> 42     WGS60    a=6378165.0             rf=298.3
#> 43     WGS66    a=6378145.0            rf=298.25
#> 44     WGS72    a=6378135.0            rf=298.26
#> 45     WGS84    a=6378137.0     rf=298.257223563
#> 46    sphere    a=6370997.0          b=6370997.0
#>                        description
#> 1                       MERIT 1983
#> 2        Soviet Geodetic System 85
#> 3             GRS 1980(IUGG, 1980)
#> 4                         IAU 1976
#> 5                        Airy 1830
#> 6              Appl. Physics. 1965
#> 7         Naval Weapons Lab., 1965
#> 8                    Modified Airy
#> 9       Andrae 1876 (Den., Iclnd.)
#> 10  Andrae 1876 (Denmark, Iceland)
#> 11 Australian Natl & S. Amer. 1969
#> 12               GRS 67(IUGG 1967)
#> 13                        GSK-2011
#> 14                     Bessel 1841
#> 15           Bessel 1841 (Namibia)
#> 16                     Clarke 1866
#> 17                Clarke 1880 mod.
#> 18              Clarke 1880 (IGN).
#> 19 Comm. des Poids et Mesures 1799
#> 20         Delambre 1810 (Belgium)
#> 21                    Engelis 1985
#> 22                    Everest 1830
#> 23                    Everest 1948
#> 24                    Everest 1956
#> 25                    Everest 1969
#> 26       Everest (Sabah & Sarawak)
#> 27    Fischer (Mercury Datum) 1960
#> 28           Modified Fischer 1960
#> 29                    Fischer 1968
#> 30                    Helmert 1906
#> 31                           Hough
#> 32    International 1909 (Hayford)
#> 33                Krassovsky, 1942
#> 34                      Kaula 1961
#> 35                      Lerch 1979
#> 36                 Maupertius 1738
#> 37          New International 1967
#> 38           Plessis 1817 (France)
#> 39                           PZ-90
#> 40                  Southeast Asia
#> 41                         Walbeck
#> 42                          WGS 60
#> 43                          WGS 66
#> 44                          WGS 72
#> 45                          WGS 84
#> 46       Normal Sphere (r=6370997)

projInfo("units")

#>        id     to_meter               name
#> 1      mm 1.000000e-03         millimetre
#> 2      cm 1.000000e-02         centimetre
#> 3       m 1.000000e+00              metre
#> 4      ft 3.048000e-01               foot
#> 5   us-ft 3.048006e-01     US survey foot
#> 6    fath 1.828800e+00             fathom
#> 7     kmi 1.852000e+03      nautical mile
#> 8   us-ch 2.011684e+01    US survey chain
#> 9   us-mi 1.609347e+03     US survey mile
#> 10     km 1.000000e+03          kilometre
#> 11 ind-ft 3.047984e-01 Indian foot (1937)
#> 12 ind-yd 9.143952e-01 Indian yard (1937)
#> 13     mi 1.609344e+03       Statute mile
#> 14     yd 9.144000e-01               yard
#> 15     ch 2.011680e+01              chain
#> 16   link 2.011680e-01               link
#> 17     dm 1.000000e-02          decimeter
#> 18     in 2.540000e-02               inch
#> 19 ind-ch 2.011670e+01       Indian chain
#> 20  us-in 2.540005e-02     US survey inch
#> 21  us-yd 9.144018e-01     US survey yard