List PROJ.4 tag information

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()
GDAL_OSR_PROJ()
GDALis3ormore()
PROJis6ormore()
new_proj_and_gdal()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

References

https://proj.org/

Note

Loading the rgdal package changes the PROJ\_LIB environmental variable to the PROJ.4 support files bundles with the package.

Examples

getPROJ4VersionInfo()

#> [1] "Rel. 7.0.0, March 1st, 2020, [PJ_VERSION: 700]"

projInfo()

#>             name
#> 1            aea
#> 2           aeqd
#> 3         affine
#> 4           airy
#> 5         aitoff
#> 6           alsk
#> 7          apian
#> 8         august
#> 9       axisswap
#> 10         bacon
#> 11    bertin1953
#> 12          bipc
#> 13         boggs
#> 14         bonne
#> 15       calcofi
#> 16          cart
#> 17          cass
#> 18            cc
#> 19          ccon
#> 20           cea
#> 21         chamb
#> 22         collg
#> 23        comill
#> 24         crast
#> 25   deformation
#> 26         denoy
#> 27          eck1
#> 28          eck2
#> 29          eck3
#> 30          eck4
#> 31          eck5
#> 32          eck6
#> 33       eqearth
#> 34           eqc
#> 35          eqdc
#> 36         euler
#> 37        etmerc
#> 38         fahey
#> 39          fouc
#> 40        fouc_s
#> 41          gall
#> 42          geoc
#> 43    geogoffset
#> 44          geos
#> 45         gins8
#> 46       gn_sinu
#> 47          gnom
#> 48         goode
#> 49          gs48
#> 50          gs50
#> 51        hammer
#> 52        hatano
#> 53       healpix
#> 54      rhealpix
#> 55       helmert
#> 56    hgridshift
#> 57        horner
#> 58           igh
#> 59         imw_p
#> 60          isea
#> 61          kav5
#> 62          kav7
#> 63        krovak
#> 64         labrd
#> 65          laea
#> 66        lagrng
#> 67          larr
#> 68          lask
#> 69        lonlat
#> 70        latlon
#> 71           lcc
#> 72          lcca
#> 73          leac
#> 74        lee_os
#> 75         loxim
#> 76          lsat
#> 77         mbt_s
#> 78       mbt_fps
#> 79        mbtfpp
#> 80        mbtfpq
#> 81        mbtfps
#> 82          merc
#> 83        mil_os
#> 84          mill
#> 85       misrsom
#> 86          moll
#> 87   molobadekas
#> 88    molodensky
#> 89         murd1
#> 90         murd2
#> 91         murd3
#> 92      natearth
#> 93     natearth2
#> 94          nell
#> 95        nell_h
#> 96         nicol
#> 97         nsper
#> 98          nzmg
#> 99          noop
#> 100      ob_tran
#> 101         ocea
#> 102          oea
#> 103        omerc
#> 104        ortel
#> 105        ortho
#> 106       pconic
#> 107    patterson
#> 108     pipeline
#> 109         poly
#> 110          pop
#> 111         push
#> 112        putp1
#> 113        putp2
#> 114        putp3
#> 115       putp3p
#> 116       putp4p
#> 117        putp5
#> 118       putp5p
#> 119        putp6
#> 120       putp6p
#> 121      qua_aut
#> 122          qsc
#> 123        robin
#> 124        rouss
#> 125        rpoly
#> 126          sch
#> 127         sinu
#> 128       somerc
#> 129        stere
#> 130       sterea
#> 131      gstmerc
#> 132          tcc
#> 133         tcea
#> 134        times
#> 135       tissot
#> 136        tmerc
#> 137      tobmerc
#> 138        tpeqd
#> 139        tpers
#> 140  unitconvert
#> 141          ups
#> 142         urm5
#> 143       urmfps
#> 144          utm
#> 145        vandg
#> 146       vandg2
#> 147       vandg3
#> 148       vandg4
#> 149        vitk1
#> 150   vgridshift
#> 151         wag1
#> 152         wag2
#> 153         wag3
#> 154         wag4
#> 155         wag5
#> 156         wag6
#> 157         wag7
#> 158      webmerc
#> 159        weren
#> 160        wink1
#> 161        wink2
#> 162       wintri
#> 163 xyzgridshift
#>                                                         description
#> 1                                                 Albers Equal Area
#> 2                                             Azimuthal Equidistant
#> 3                                             Affine transformation
#> 4                                                              Airy
#> 5                                                            Aitoff
#> 6                                      Mod. Stereographic of Alaska
#> 7                                                  Apian Globular I
#> 8                                               August Epicycloidal
#> 9                                                     Axis ordering
#> 10                                                   Bacon Globular
#> 11                                                      Bertin 1953
#> 12                              Bipolar conic of western hemisphere
#> 13                                                  Boggs Eumorphic
#> 14                                          Bonne (Werner lat_1=90)
#> 15                        Cal Coop Ocean Fish Invest Lines/Stations
#> 16                                   Geodetic/cartesian conversions
#> 17                                                          Cassini
#> 18                                              Central Cylindrical
#> 19                                                    Central Conic
#> 20                                           Equal Area Cylindrical
#> 21                                             Chamberlin Trimetric
#> 22                                                        Collignon
#> 23                                                   Compact Miller
#> 24                                   Craster Parabolic (Putnins P4)
#> 25                                             Kinematic grid shift
#> 26                                          Denoyer Semi-Elliptical
#> 27                                                         Eckert I
#> 28                                                        Eckert II
#> 29                                                       Eckert III
#> 30                                                        Eckert IV
#> 31                                                         Eckert V
#> 32                                                        Eckert VI
#> 33                                                      Equal Earth
#> 34                           Equidistant Cylindrical (Plate Carree)
#> 35                                                Equidistant Conic
#> 36                                                            Euler
#> 37                                     Extended Transverse Mercator
#> 38                                                            Fahey
#> 39                                                          Foucaut
#> 40                                               Foucaut Sinusoidal
#> 41                                        Gall (Gall Stereographic)
#> 42                                              Geocentric Latitude
#> 43                                                Geographic Offset
#> 44                                     Geostationary Satellite View
#> 45                                        Ginsburg VIII (TsNIIGAiK)
#> 46                                        General Sinusoidal Series
#> 47                                                         Gnomonic
#> 48                                                 Goode Homolosine
#> 49                                    Mod. Stereographic of 48 U.S.
#> 50                                    Mod. Stereographic of 50 U.S.
#> 51                                     Hammer & Eckert-Greifendorff
#> 52                                   Hatano Asymmetrical Equal Area
#> 53                                                       HEAPJ_LPix
#> 54                                                      rHEAPJ_LPix
#> 55                   3(6)-, 4(8)- and 7(14)-parameter Helmert shift
#> 56                                            Horizontal grid shift
#> 57                                     Horner polynomial evaluation
#> 58                                     Interrupted Goode Homolosine
#> 59                         International Map of the World Polyconic
#> 60                                    Icosahedral Snyder Equal Area
#> 61                                                      Kavraisky V
#> 62                                                    Kavraisky VII
#> 63                                                           Krovak
#> 64                                                          Laborde
#> 65                                     Lambert Azimuthal Equal Area
#> 66                                                         Lagrange
#> 67                                                         Larrivee
#> 68                                                        Laskowski
#> 69                                              Lat/long (Geodetic)
#> 70                                        Lat/long (Geodetic alias)
#> 71                                          Lambert Conformal Conic
#> 72                              Lambert Conformal Conic Alternative
#> 73                                         Lambert Equal Area Conic
#> 74                                        Lee Oblated Stereographic
#> 75                                                       Loximuthal
#> 76                                        Space oblique for LANDSAT
#> 77                           McBryde-Thomas Flat-Polar Sine (No. 1)
#> 78                            McBryde-Thomas Flat-Pole Sine (No. 2)
#> 79                              McBride-Thomas Flat-Polar Parabolic
#> 80                                McBryde-Thomas Flat-Polar Quartic
#> 81                             McBryde-Thomas Flat-Polar Sinusoidal
#> 82                                                         Mercator
#> 83                                     Miller Oblated Stereographic
#> 84                                               Miller Cylindrical
#> 85                                           Space oblique for MISR
#> 86                                                        Mollweide
#> 87                                Molodensky-Badekas transformation
#> 88                                             Molodensky transform
#> 89                                                        Murdoch I
#> 90                                                       Murdoch II
#> 91                                                      Murdoch III
#> 92                                                    Natural Earth
#> 93                                                  Natural Earth 2
#> 94                                                             Nell
#> 95                                                      Nell-Hammer
#> 96                                                Nicolosi Globular
#> 97                                           Near-sided perspective
#> 98                                             New Zealand Map Grid
#> 99                                                     No operation
#> 100                                  General Oblique Transformation
#> 101                                  Oblique Cylindrical Equal Area
#> 102                                              Oblated Equal Area
#> 103                                                Oblique Mercator
#> 104                                                   Ortelius Oval
#> 105                                                    Orthographic
#> 106                                               Perspective Conic
#> 107                                           Patterson Cylindrical
#> 108                                 Transformation pipeline manager
#> 109                                            Polyconic (American)
#> 110                   Retrieve coordinate value from pipeline stack
#> 111                         Save coordinate value on pipeline stack
#> 112                                                      Putnins P1
#> 113                                                      Putnins P2
#> 114                                                      Putnins P3
#> 115                                                     Putnins P3'
#> 116                                                     Putnins P4'
#> 117                                                      Putnins P5
#> 118                                                     Putnins P5'
#> 119                                                      Putnins P6
#> 120                                                     Putnins P6'
#> 121                                                Quartic Authalic
#> 122                                Quadrilateralized Spherical Cube
#> 123                                                        Robinson
#> 124                                         Roussilhe Stereographic
#> 125                                           Rectangular Polyconic
#> 126                                    Spherical Cross-track Height
#> 127                                   Sinusoidal (Sanson-Flamsteed)
#> 128                                            Swiss. Obl. Mercator
#> 129                                                   Stereographic
#> 130                               Oblique Stereographic Alternative
#> 131 Gauss-Schreiber Transverse Mercator (aka Gauss-Laborde Reunion)
#> 132                                  Transverse Central Cylindrical
#> 133                               Transverse Cylindrical Equal Area
#> 134                                                           Times
#> 135                                                          Tissot
#> 136                                             Transverse Mercator
#> 137                                                 Tobler-Mercator
#> 138                                           Two Point Equidistant
#> 139                                              Tilted perspective
#> 140                                                 Unit conversion
#> 141                                   Universal Polar Stereographic
#> 142                                                        Urmaev V
#> 143                                    Urmaev Flat-Polar Sinusoidal
#> 144                             Universal Transverse Mercator (UTM)
#> 145                                             van der Grinten (I)
#> 146                                              van der Grinten II
#> 147                                             van der Grinten III
#> 148                                              van der Grinten IV
#> 149                                                     Vitkovsky I
#> 150                                             Vertical grid shift
#> 151                                         Wagner I (Kavraisky VI)
#> 152                                                       Wagner II
#> 153                                                      Wagner III
#> 154                                                       Wagner IV
#> 155                                                        Wagner V
#> 156                                                       Wagner VI
#> 157                                                      Wagner VII
#> 158                                  Web Mercator / Pseudo Mercator
#> 159                                                   Werenskiold I
#> 160                                                        Winkel I
#> 161                                                       Winkel II
#> 162                                                   Winkel Tripel
#> 163                                           Geocentric grid shift