projInfo {rgdal}	R Documentation

List PROJ.4 tag information

Description

The projInfo function lists known values and descriptions for PROJ.4 tags for tag in c("proj", "ellps", "datum", "units"); getPROJ4VersionInfo returns the version of the underlying PROJ.4 release, getPROJ4libPath returns the value of the PROJ\_LIB environment variable, projNAD detects the presence of NAD datum conversion tables (looking for conus).

Usage

projInfo(type = "proj")
getPROJ4VersionInfo()
getPROJ4libPath()
projNAD()

Arguments

type
One of these tags: c("proj", "ellps", "datum", "units")

Details

The output data frame lists the information given by the proj application with flags -lp, -le, -ld or -lu.

Value

A data frame with a name and description column, and two extra columns for the "ellps" and "datum" tags.

Note

Loading the rgdal package changes the PROJ\_LIB environmental variable to the PROJ.4 support files bundles with the package.

Author(s)

Roger Bivand Roger.Bivand@nhh.no

References

http://proj.maptools.org/

Examples

getPROJ4VersionInfo()
projInfo()

[Package rgdal version 1.4-8 Index]