Show Well-Known Text spatial reference system metadata

Use GDAL/OGR spatial reference objects to convert a PROJ.4 representation to a Well-Known Text representation, and report an EPSG code if it can be determined by OGR SRS services.

showWKT(p4s, file = NULL, morphToESRI = TRUE)
showP4(wkt, morphFromESRI=TRUE)
showEPSG(p4s)

Arguments

p4s
A valid PROJ.4 string representing a spatial reference system

file
if not NULL, a file name to which the output Well-Known Text representation should be written

morphToESRI
default TRUE, morph the WKT string to the representation used by ESRI

wkt
A valid WKT character string representing a spatial reference system

morphFromESRI
default TRUE, morph the WKT string from the representation used by ESRI

Value

A character string containing the WKT representation of the PROJ.4 string.

References

http://www.gdal.org/osr_tutorial.html

See also

Examples

cities <- readOGR(system.file("vectors", package = "rgdal")[1], "cities")

#> OGR data source with driver: ESRI Shapefile 
#> Source: "/home/rsb/topics/packages/R-Forge/rgdal/pkg/inst/vectors", layer: "cities"
#> with 606 features
#> It has 4 fields
#> Integer64 fields read as strings:  POPULATION 

readLines(system.file("vectors/cities.prj", package = "rgdal")[1])

#> [1] "GEOGCS[\"GCS_WGS_1984\",DATUM[\"D_WGS_1984\",SPHEROID[\"WGS_1984\",6378137,298.257223563]],PRIMEM[\"Greenwich\",0],UNIT[\"Degree\",0.0174532925199433]]"

showWKT(proj4string(cities))

#> [1] "GEOGCS[\"GCS_WGS_1984\",DATUM[\"D_WGS_1984\",SPHEROID[\"WGS_1984\",6378137,298.257223563]],PRIMEM[\"Greenwich\",0],UNIT[\"Degree\",0.017453292519943295]]"

showWKT("+init=epsg:28992")

#> [1] "PROJCS[\"Amersfoort_RD_New\",GEOGCS[\"GCS_Amersfoort\",DATUM[\"D_Amersfoort\",SPHEROID[\"Bessel_1841\",6377397.155,299.1528128]],PRIMEM[\"Greenwich\",0],UNIT[\"Degree\",0.017453292519943295]],PROJECTION[\"Double_Stereographic\"],PARAMETER[\"latitude_of_origin\",52.15616055555555],PARAMETER[\"central_meridian\",5.38763888888889],PARAMETER[\"scale_factor\",0.9999079],PARAMETER[\"false_easting\",155000],PARAMETER[\"false_northing\",463000],UNIT[\"Meter\",1]]"

showP4(showWKT("+init=epsg:28992"))

#> [1] "+proj=sterea +lat_0=52.15616055555555 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +ellps=bessel +units=m +no_defs "

showEPSG("+proj=utm +zone=30")

#> [1] "OGRERR_UNSUPPORTED_SRS"

showEPSG("+proj=longlat +ellps=WGS84")

#> [1] "4326"